==> 1_ecmo_ich_part_2.Rmd <==
---
title: "CCCC: ECMO and ICH patients (part 2)"
author: "Adrian Barnett"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  word_document: 
      reference_docx: rmarkdown-styles-SAP-landscape.docx
---
  
```{r, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, warning=FALSE, message=FALSE, error=FALSE, comment='', dpi=400)
options(width=1000) # Wide pages
options(scipen=999) # avoid scientific presentation
source('99_functions.R')
source('99_functions_joint.R') # specifically for joint models
library(dplyr)
library(stringr)
library(tidyr)
library(broom)
library(R2WinBUGS)
bugs_location = "c:/Program Files/WinBUGS14" # location of the file WinBUGS14.exe
library(mstate)
library(cmprsk) # for competing risks
library(survminer) # for nice plots of competing risks; also test of influential values
library(flextable) # for nice tables
library(captioner)
table_nums <- captioner(prefix = "Table", levels=2, type = c("n", "c")) # number and character
figure_nums <- captioner(prefix = "Figure", levels=2, type = c("n", "c"))
source("Martin/ext_mstate.R")
library(INLA) # for joint models

# graphics things:
library(diagram)
library(ggplot2)
theme_set(theme_bw())
library(ggupset)  # to plot combination of symptoms
g.theme = theme_bw() + theme(panel.grid.minor = element_blank())
cbPalette <- c("#E69F00", "#56B4E9", "#009E73", "#F0E442", "#0072B2", "#D55E00", "#999999", "#CC79A7")

## get the data
source = 'real_data'
#source = 'dummy_data'
if(source == 'dummy_data'){
  load(file='data/COVID_Data_stroke_dummy.RData') # from 0_read_data.R, dummy data
  # fixes with stroke
  patients = mutate(patients,
          stroke_group = ifelse(is.na(complication_stroke_date)==TRUE, 'None', stroke_group), # blank stroke group for those with no date
          stroke_group = ifelse(is.na(complication_stroke_date)==FALSE, sample(c('Hemorrhage','Ischemic','Other/Unknown','None'), size=n(), replace=TRUE), stroke_group) # randomly add stroke group for those with a date
                    )
}
if(source == 'real_data'){
  load(file='data/COVID_Data_stroke.RData') # from 0_read_data.R
}

# make last possible date based on database freeze
last_possible = as.Date(paste(data_date$day, data_date$month, '2020', sep='-'), format='%d-%b-%Y')
censor_day = 90 # maximum day for plots and estimates

## data edits
# a) patients
# slim down variables:
patients = select(patients, 'pin', 'site_name','country','age', 'sex', 'bmi','ethnicity', 'pregnant',
                  'date_icu','date_ecmo','date_ecmo_discontinued', 
                  'date_death','date_discharge','last_date','date_first_symptom',
                  'date_mech_vent_discontinued', 'date_mechanical_ventilation',
                  "height","weight",'date_admission','date_hospital_discharge',
                  starts_with('comorbidity_'),
                  'sofa','apache_ii',
                  'ecmo_type','ecmo_drainage_cannula_insertion_site',
                  'ecmo_return_cannula_insertion_site',
                  "ecmo_cardiac_arrest_before", "ecmo_worst_pa_o2_6hr_before",
                  "ecmo_worst_pa_co2_6hr_before", 'ecmo_continuous_rpt_before',
                  starts_with("complication_"), 'stroke_during_treatment','stroke_type', 'stroke_group', 'outcome','date_outcome','date_transfer') %>%
  mutate(pregnant = ifelse(pregnant==0, 'No', 'Yes'), # nicer label
         date_ecmo_discontinued = ifelse(is.na(date_ecmo), NA, date_ecmo_discontinued), # if no start date for ECMO, then remove any end date
         date_ecmo_discontinued = as.Date(date_ecmo_discontinued, origin='1970-01-01'))
## inclusion
patients = filter(patients,
    !country %in% c('Australia','United Kingdom')) # remove countries that did not collect stroke data
 
## exclusions
# age, must be adults
previous_n = nrow(patients)
n_exclude_age_stroke = nrow(filter(patients, age < 18, !is.na(complication_stroke_date))) # count those excluded in stroke group
patients = filter(patients,
    !is.na(age),
    age >= 18) # 
n_exclude_age = previous_n - nrow(patients) 
# no ICU date
previous_n = nrow(patients)
n_exclude_icu_stroke = nrow(filter(patients, is.na(date_icu), !is.na(complication_stroke_date))) # count those excluded in stroke
patients = filter(patients,
    !is.na(date_icu))
n_exclude_icu = previous_n -nrow(patients) 
# missing sex
previous_n = nrow(patients)
patients = filter(patients,
          sex != 'Missing', # remove those with missing sex as they are missing almost all their data
         !is.na(sex))
n_exclude_missing = nrow(patients) - previous_n
  
# further data edits
patients  = mutate(patients,
         # impossible admission date set to missing:
         date_admission = ifelse(date_admission >= last_possible, NA, date_admission), 
         date_admission = as.Date(date_admission, origin='1970-01-01'),
         # update stroke to add EOT
         complication_stroke = ifelse(!is.na(stroke_during_treatment) & stroke_during_treatment==1, 1, complication_stroke),
         # labels for factors
         complication_stroke = factor(complication_stroke, levels=0:1, labels=c('No','Yes')),
         complication_seizure = factor(complication_seizure, levels=0:1, labels=c('No','Yes')),
         complication_heart_failure = factor(complication_heart_failure, levels=0:1, labels=c('No','Yes')),
    comorbidity_chronic_kidney_disease = factor(comorbidity_chronic_kidney_disease, levels=0:1, labels=c('No','Yes')),
    comorbidity_malignant_neoplasm = factor(comorbidity_malignant_neoplasm, levels=0:1, labels=c('No','Yes')),
    comorbidity_smoking = factor(comorbidity_smoking, levels=0:1, labels=c('No','Yes')),
         comorbidity_severe_liver_disease = factor(comorbidity_severe_liver_disease, levels=0:1, labels=c('No','Yes')),
         comorbidity_diabetes = factor(comorbidity_diabetes, levels=0:1, labels=c('No','Yes')),
         comorbidity_hypertension = factor(comorbidity_hypertension, levels=0:1, labels=c('No','Yes')),
         comorbidity_chronic_neurological_disorder = factor(comorbidity_chronic_neurological_disorder, levels=0:1, labels=c('No','Yes')),
         comorbidity_chronic_cardiac_disease = factor(comorbidity_chronic_cardiac_disease, levels=0:1, labels=c('No','Yes')),
         ecmo_cardiac_arrest_before = factor(ecmo_cardiac_arrest_before, levels=0:1, labels=c('No','Yes')),
         ecmo_continuous_rpt_before = factor(ecmo_continuous_rpt_before, levels=0:1, labels=c('No','Yes')),
         any_mv = as.numeric(!is.na(date_mechanical_ventilation)),
         any_mv = factor(any_mv, levels=0:1, labels=c('No','Yes')),
         any_ecmo = as.numeric(ecmo_type!=''),
         any_ecmo = factor(any_ecmo, levels=0:1, labels=c('No','Yes')))

# b) daily data
daily = select(daily, pin, date_daily, ecmo, platelet_count, "p_h", "aptt", "inr", alt_sgpt, "ast_sgot", 'bilirubin','blood_urea_nitrogen',
               "lymphocyte_count", 'eotd_anticoagulants', "eotd_anticoagulants_type",
               "pa_o2", "pa_co2",
               "eotd_haemorrhagic_complication", "eotd_haemorrhagic_complication_source",
               ) %>%
  mutate(date_daily = ifelse(date_daily >= last_possible, NA, date_daily), # impossible date set to missing
         date_daily = as.Date(date_daily, origin='1970-01-01')) 

# combined
combined = full_join(patients, daily, by='pin')

# change 'missing' to NA
for_missing = select(patients, -pin, -any_ecmo, -stroke_type) %>% # remove ID; any_ecmo is never missing; stroke_type is a list variable
  mutate(
    ethnicity = ifelse(ethnicity=='Missing', NA, ethnicity),
    outcome = ifelse(outcome %in% c('','unknown'), NA, outcome)) %>% # tidy up missing outcome
  mutate_if(~is.character(.x), ~ifelse(.x=='', NA, .x)) # change '' to missing for character variables
```

This report uses the data from `r data_date$month` `r data_date$day`. 

Data from Australia and the UK were not included as these countries did not collect the necessary stroke data.

Patients were excluded for:

* being under age 18 (total excluded = `r n_exclude_age`, stroke patients excluded = `r n_exclude_age_stroke`)
* not having an ICU admission date (n = `r n_exclude_icu`, stroke patients excluded = `r n_exclude_icu_stroke`)

### Check of stroke numbers

```{r, include=FALSE}
# a) no stroke date
f = filter(patients, complication_stroke=='Yes') %>%
  select(pin, date_icu, complication_stroke_date, last_date, date_death, date_discharge, date_hospital_discharge, stroke_group) 
n_start = nrow(f) # starting number
n_missing_stroke_date = sum(is.na(f$complication_stroke_date)) # number excluded for missing stroke
# b) stroke before ICU admission
f2 = filter(f, 
            !is.na(complication_stroke_date)) # now use just complete
n_update = nrow(f2)
f2 = filter(f2, complication_stroke_date >= date_icu)
n_stroke_prior_icu = n_update - nrow(f2) # number excluded
# c) stroke after hospital discharge
f3 = mutate(f2, 
            diff = as.numeric(complication_stroke_date - date_hospital_discharge)) %>%
  filter(is.na(diff) | diff<=0)
n_stroke_post_discharge = nrow(f2) - nrow(f3)
n_final = nrow(f3)
# find patients knocked out by `c`
s = f2$pin[f2$pin %in% f3$pin==FALSE]
filter(f, pin%in%s)

## for patients without a stroke date, set them as 'None' for stroke type
index = is.na(patients$complication_stroke_date)
patients$complication_stroke[index] = 'No'
patients$stroke_group[index] = 'None'
#for (i in which(index==TRUE)){ # not working, but we do not use stroke_type here
#  patients$stroke_type[[i]] = NULL
#}
## same for patients with stroke before admission
index = patients$complication_stroke_date < patients$date_icu & !is.na(patients$complication_stroke_date) & !is.na(patients$date_icu)
patients$complication_stroke[index] = 'No'
patients$stroke_group[index] = 'None'
```

We start with `r n_start` patients who had a stroke.
Of these patients `r n_missing_stroke_date` have no stroke date, `r n_stroke_prior_icu` patients had their stroke prior to their ICU admission.
This gives `r n_final` patients with a known stroke date during their hospital admission.

# Descriptive statistics

The tables below are split by the three types of stroke: Hemorrhage, Ischemic and Stroke, not specified. There is also a columns for patients with no stroke.

```{r}
ftype = group_by(patients, stroke_group) %>%
  tally() %>%
  arrange(n) %>%
  mutate(percent = round(prop.table(n)*1000)/10,
         stroke_group = ifelse(stroke_group=='Other/Unknown', 'Stroke, not specified', stroke_group)) # nicer label
ftab = flextable(ftype) %>%
  theme_box() %>%
  autofit()
table_nums(name="stroke_type","Stroke type, frequency table.", level=2) # title
ftab
```

```{r, continuous_table}
#
ftab = desc_table_cont(
  indata=patients, 
  vars = c('age','bmi','sofa','apache_ii'), 
  group_var = 'stroke_group', 
  var_levels = c('Hemorrhage','Ischemic','Other/Unknown','None'),
  heading = c('Hemorrhage','Ischemic','Stroke, not specified','None'),
  autofit = FALSE,
  include_total = FALSE) %>%
  fontsize(size=8) # smaller font to fit
table_nums(name="cont_table","Continuous variables - table.", level=2) # title
ftab
```

```{r, categorical_table}
# combine ethnicity
patients = mutate(patients,
                  Ethnicity = case_when(
                    ethnicity == 'aboriginal'  ~ 'other',
                    ethnicity == 'arab'  ~ 'other',
                    ethnicity == 'east_asian'  ~ 'Asian',
                    ethnicity == 'west_asian'  ~ 'Asian',
                    ethnicity == 'south_asian'  ~ 'Asian',
                    ethnicity == 'latin_american'  ~ 'Hispanic',  
                    TRUE ~ as.character(ethnicity) # all others remain the same
                  ) )
#
ftab = desc_table_cat(indata = patients, 
                      vars = c('any_ecmo','any_mv','sex','Ethnicity','country'), 
                      group_var = 'stroke_group', 
                      var_levels = c('Hemorrhage','Ischemic','Other/Unknown','None'),
                      include_total = FALSE,
                      heading = c('Hemorrhage','Ischemic','Stroke, not specified','None')) # see 99_functions.R
table_nums(name="cat_table","Categorical variables.", level=2) # title
ftab
```

```{r, comorbid_table}
co_vars = c( 'comorbidity_hypertension','comorbidity_diabetes','comorbidity_chronic_kidney_disease','comorbidity_malignant_neoplasm','comorbidity_smoking','comorbidity_chronic_neurological_disorder','comorbidity_chronic_cardiac_disease')
ftab = desc_table_cat(indata = patients, 
                      vars = co_vars, 
                      group_var = 'stroke_group', 
                      var_levels = c('Hemorrhage','Ischemic','Other/Unknown','None'),
                      include_total = FALSE,
                      heading = c('Hemorrhage','Ischemic','Stroke, not specified','None') )
table_nums(name="comorb_table","Comorbidities.", level=2) # title
ftab
```

```{r, pregnant_table}
to_tab = filter(patients, sex=='Female')
ptab = ftab = desc_table_cat(indata = to_tab, 
                      vars = c('pregnant'), 
                      group_var = 'stroke_group', 
                      var_levels = c('Hemorrhage','Ischemic','Other/Unknown','None'),
                      include_total = FALSE,
                      heading = c('Hemorrhage','Ischemic','Stroke, not specified','None')) 
table_nums(name="preg_table","Pregnant (women only).", level=2) # title
ptab
```



```{r, time_table}
# create time variables
patients = mutate(patients,
      date_end = ifelse(is.na(date_discharge)==FALSE, date_hospital_discharge, date_death), # end date of death or discharge
      date_end = as.Date(date_end, origin='1970-01-01'),
      days_symptoms_hosp = as.numeric(date_admission - date_first_symptom) +0.5, 
      days_hosp = as.numeric(date_end - date_admission) + 0.5,
      #
      days_hosp_stroke = as.numeric(complication_stroke_date - date_admission) + 0.5,
      #
      duration_ecmo = as.numeric(date_ecmo_discontinued - date_ecmo) + 0.5,
      duration_mv = date_mech_vent_discontinued - date_mechanical_ventilation +0.5)
#
time_vars = c('days_symptoms_hosp','days_hosp','days_hosp_stroke','duration_ecmo','duration_mv')
#
ftab = desc_table_cont(indata = patients, 
                      vars = time_vars, 
                      group_var = 'stroke_group', 
                      var_levels = c('Hemorrhage','Ischemic','Other/Unknown','None'),
                      include_total = FALSE,
                      heading = c('Hemorrhage','Ischemic','Stroke, not specified','None')) %>% 
   fontsize(part='all', size=6) %>% # smaller font size to fit
  autofit()
table_nums(name="time_table", "Time variables (in days)", level=2) # title
ftab
```

```{r, outcome_table}
ftab = desc_table_cat(indata = patients, 
                      vars = 'outcome', 
                      group_var = 'stroke_group', 
                      var_levels = c('Hemorrhage','Ischemic','Other/Unknown','None'),
                      include_total = FALSE,
                      heading = c('Hemorrhage','Ischemic','Stroke, not specified','None')) 
table_nums(name="outcome_table", "Final outcome", level=2) # title
ftab
```

```{r}
patients = mutate(patients,
        stroke_group_better = ifelse(stroke_group=='Other/Unknown', 'Stroke, not specified', stroke_group))
bplot = ggplot(data=patients, aes(x=factor(stroke_group_better), y=age))+
  geom_boxplot(fill=grey(0.9))+
  xlab('Stroke type')+
  ylab('Age, years')
figure_nums(name="boxplot_age", 'Box plot of age by stroke type') # title
bplot
```

# Model diagram

The figure below shows the four states in the multi-state survival model. Patients start at ICU admission. Patients were censored if they were transferred to another hospital using the date of their transfer, or if they were still in hospital `r censor_day` after their ICU admission.

```{r}
states = c('ICU\nadmission','Stroke\n(3 types)','Death','Hospital\ndischarge')
# Matrix of joins between states:
M = matrix(data=c(
  2,1,
  3,1,
  4,1,
  3,2,
  4,2
), byrow = TRUE, ncol=2)
figure_nums(name="model_diagram", "The four states in the model and the possible movements between states", level=1) # title
multistate_diagram(states=states, links = M, pos = c(1, 2, 1), arr.pos=0.65) # see 99_functions.R
```

### Table of transitions between states

This table shows the transitions (movements) between the states (`r figure_nums("model_diagram", display = "cite")`).

```{r, include=FALSE}
### prepare data for survival analysis ###
# create multi-state data
my.data = time_transitions_string(indata=patients,
                intermediate_prior_ongoing= FALSE,
                censor_day = censor_day,
                start_state = 'date_icu',
                int_state = 'complication_stroke_date',
                int_state_end = NULL,
                discharge_date = 'date_hospital_discharge',
                dates= c('date_icu', 'date_ecmo', "date_ecmo_discontinued", 'complication_stroke_date', 'date_transfer', 'date_death','date_hospital_discharge')) #
# rename from and to states
my.data = mutate(my.data, 
                 to = ifelse(to=='date_transfer', 'censored', to),
                 to = ifelse(to=='date_int_state_start', 'stroke', to),
                 to = ifelse(to=='date_ecmo_discontinued', 'ecmo_end', to),
                 from = ifelse(from=='date_ecmo', 'ecmo_start', from),
                 from = ifelse(from=='date_int_state_start', 'stroke', from),
                 from = ifelse(from=='date_icu', 'icu', from),
                 from = ifelse(from=='date_ecmo_discontinued', 'ecmo_end', from),
                 from = ifelse(from=='icu' & to=='ecmo_end', 'ecmo_start', from)) %>% # assume ECMO at ICU start
  mutate(age_c = (age - 60)/10, # standardising
         bmi_c = (bmi - 28)/5) 
         
# with(my.data, table(from, to))
# specifically for stroke analysis
my.data.stroke = filter(my.data, !from == 'stroke') %>% # not interested in events post-stroke
  mutate(# update to state to allow multiple stroke types
         from = ifelse(from=='stroke' & stroke_group=='Ischemic', 'stroke_i', from), # 2= other, 3=Ischemic, 4=Hemorrhage
         from = ifelse(from=='stroke' & stroke_group=='Hemorrhage', 'stroke_h', from),
         to = ifelse(to=='stroke' & stroke_group=='Ischemic', 'stroke_i', to), # 2= other, 3=Ischemic, 4=Hemorrhage
         to = ifelse(to=='stroke' & stroke_group=='Hemorrhage', 'stroke_h', to)
         )

# check of stroke patients
#index = my.data$id %in% f3$pin # patients with a stroke
#my.data.stroke = my.data[index,]
#
#not.there = f2$pin[f2$pin%in% my.data$id ==FALSE] # stroke patients not in the transition data
#check = filter(patients, pin%in% not.there) %>%
#    select(pin, date_icu, complication_stroke_date, last_date, date_death, date_hospital_discharge) 
```

```{r}
columns = c('from', 'censored','date_ecmo','ecmo_end','stroke','stroke_i','stroke_h','discharge','death') # to get right order
tab = as.data.frame(with(my.data.stroke, table(from, to))) %>%
  mutate(from = case_when(
    from=='icu' ~ 'ICU admission', 
    from=='ecmo_start' ~ 'During ECMO', 
    from=='ecmo_end' ~ 'Post-ECMO')
  ) %>%
  tidyr::spread(to, Freq) %>%
  select(columns)
# table header
table_header <- data.frame(
  col_keys = columns,
  h1 = c('', rep("State moved to", 8)),
  h2 = c(' ', '','ECMO','ECMO','Stroke','Stroke','Stroke','',''),
  h3 = c('State moved from', 'Censored','Start', 'End','Stroke, not specified','Ischemic','Hemorrhage','Discharge','Death'),
  stringsAsFactors = FALSE )
#
ftab = flextable(tab) %>%
 set_header_df(mapping = table_header, key = "col_keys" ) %>% # add header
 merge_h(i=1, part = "header") %>% # merge "To" column headers
 merge_h(i=2, part = "header") %>% # merge stroke
  theme_box() %>%
  autofit() %>%
  align(align='center', part='all') # centre text
table_nums(name="transitions", "Transition numbers between states") # title
ftab
```

## Survival models for time to stroke

We use a Cox survival model to examine the time to stroke, examining ischemic stroke and hemorrhagic stroke separately. This models the time to stroke whilst accounting for censoring due to death and discharge, or administrative censoring if the patient is still in hospital at the end of the study.

Weibull and exponential models did not converge, likely due to the relatively small number of strokes.

```{r}
median_sofa = median(patients$sofa, na.rm=TRUE) # threshold for high/low SOFA
for_stroke = 
  filter(my.data.stroke, 
         sex != 'Not specified') %>% # remove small number missing
  mutate(#age2 = age_c*age_c, # quadratic
         # make variables with missing to avoid losing patients with partially complete data
         #bmi_c = case_when( # categorise BMI
        #   bmi <= 28 ~ 'Low',
        #   bmi > 28 ~ 'High',
        #   is.na(bmi) ~ 'Missing'
        # ),
         sofa_c = case_when( # categorise SOFA
           sofa <= median_sofa ~ 'Low',
           sofa > median_sofa ~ 'High',
           is.na(sofa) ~ 'Missing'
         ),
         sofa_c = relevel(factor(sofa_c), ref='Low')) # reference level
## cox model
#
control = coxph.control()
control$iter.max = 100 # increase iterations
# loop through outcomes
outcomes_list = list()
outcomes_list[[1]] = c('stroke','stroke_h','stroke_i') # all strokes combined
outcomes_list[[2]] = 'stroke_h'
outcomes_list[[3]] = 'stroke_i'
hr_table = events = NULL
for (s_type in outcomes_list){ # loop through stroke types
  # + factor(sofa_c) # does not converge, ditto for + bmi_c (BMI has around 10% missing)
  for_stroke = mutate(for_stroke, event = as.numeric(to %in% s_type)) # make event depending on stroke type
cmodel = coxph(Surv(time=exit, event= event, type='right') ~ age_c + sex + sofa_c + cluster(site_name), data=for_stroke, control = control)
digits = 2
ests = tidy(cmodel, conf.int = TRUE, exponentiate = TRUE) %>%
    mutate(
      stroke_type = paste(s_type,collapse = ','),
      HR = roundz(estimate, digits),
      conf.low = roundz(conf.low, digits),
      conf.high = roundz(conf.high, digits),
      CI = paste(conf.low, ' to ', conf.high, sep='')) %>%
    select(-std.error, -statistic, -p.value, -estimate, -robust.se, -starts_with('conf'))
hr_table = bind_rows(hr_table, ests)
events = c(events, cmodel$nevent)
}
hr_table = mutate(hr_table,
                  stroke_type = case_when(
                    stroke_type == 'stroke,stroke_h,stroke_i' ~ 'All strokes combined',
                    stroke_type == 'stroke_i' ~ 'Ischemic',
                    stroke_type == 'stroke_h' ~ 'Hemorrhage'
                  ),
                  term = ifelse(term=='age_c', 'Age (+10 years)', term),
                  term = ifelse(term=='bmi_c', 'BMI (+5 kg/m2)', term),
                  term = ifelse(term=='sexMale', 'Male', term)) %>%
  select(stroke_type, term, HR, CI) %>%
  rename('Variable' = 'term',
         'Stroke type' = 'stroke_type')
ftab = flextable(hr_table) %>%
  merge_v(j=1) %>%
  theme_box() %>%
  autofit() 
table_nums(name="hazard_table", "Hazard ratios and 95% confidence intervals for time to stroke", level=1) # title
ftab
```

The table shows the hazard ratios and 95% confidence intervals for the three types of stroke. I used two predictors of age and sex which are almost 100% complete. 

SOFA score has been added as a simple three category variable of missing, low or high, with low or high determined using the median of `r median_sofa`. I used this approach because of the large amount of missing data for SOFA. The reference category is low.

The confidence intervals for the effects of age and sex are relatively wide for some estimates, which also is likely because of the small number of strokes.

### Cumulative risk plot for time to stroke split by stroke type

```{r, fig.width=9, fig.height=5.5}
# for nice labels
my.data.stroke = mutate(my.data.stroke, from.nice=case_when(
  from=='icu' ~ 'ICU', # cannot be any spaces
  from=='ecmo_start' ~ 'ECMO_start', # cannot be any spaces
  from=='ecmo_end' ~ 'ECMO_end' # cannot be any spaces
  ))
# cumulative incidence
cum_inc = with(my.data.stroke, cuminc(ftime=exit, fstatus=to, group=from.nice, cencode='censored'))
# remove lines not needed (e.g., stroke to stroke)
any_ests = NULL
for (z in 1:(length(cum_inc)-1)){
  empty = any(cum_inc[[z]]$est !=0) # find missing
  any_ests = c(any_ests, empty)
}
index = rev(which(any_ests == FALSE))
for (z in index){
  cum_inc[[z]] = NULL # blank lines
}
# plot
g = ggcompetingrisks(cum_inc,
                 xlab = 'Days since ICU admission',
                 ylab = 'Cumulative probability',
                 xlim = c(0, censor_day), # 
                 title = '',
#                 gsep = '_', # dummy to prevent groups being created from spaces
                 cumcensor = TRUE, # does not work
                 legend.title='',
                 risk.table = TRUE,
                 conf_int = TRUE)
# create plot with better facet labels
to_plot = mutate(g$data, 
                 facet = paste('State moved from: ', group, sep=''))
cum_plot = ggplot(data=to_plot, aes(x=time, y=est, col=factor(event)))+
  geom_line(size = 1.05) + # make lines thicker
  facet_wrap(~facet)+
  xlab('Days since ICU admission')+
  ylab('Cumulative probability')+
  theme_bw()+
  scale_x_continuous(breaks=c(0, 20, 40, 60, 80))+
  scale_y_continuous(breaks=seq(0, 0.9, 0.1))+
  scale_colour_manual(name="State moved to:", values=cbPalette) +  # change labels?
  theme(
    panel.grid.minor = element_blank(),
    panel.grid.major = element_line(colour=grey(0.8)),
    text = element_text(size=14), 
    legend.position = 'top',
    legend.title = element_text(size=14),
    legend.text = element_text(size=14))
figure_nums(name="cuminc_stroke", "Cumulative probabilities over time using each stroke type.", level=1)
cum_plot
jpeg('figures/stroke/competing_risks_stroke_separate.jpg', width=6.8, height=6, units='in', res=400, quality=100)
print(cum_plot)
invisible(dev.off())
```

The panels of the plot are for the states of "ICU admission" and the three types of "stroke". The lines show where patients move to from these states.

There is a clear higher risk of death and lower risk of discharge for patients with a hemorrhagic stroke.

##### Alternative version of the above plot

```{r, fig.width=8.5, fig.height=5.5}
# alternative plot
to_plot = mutate(g$data,
                 event = ifelse(event=='date_ecmo', 'ECMO start', event),
                 event = paste("State moved to:", event))
alt_plot = ggplot(data= to_plot, aes(x=time, y=est, col=factor(group)))+
  geom_line(size=1.05)+
  xlab('Days since ICU')+
  ylab('Cumulative probability')+
scale_color_manual('State moved from:', values=(c("#CD950C", "#66CD00", "#FF00FF", "#ADD8E6")))+
  facet_wrap(~event, scale='free_y')+
  theme(legend.position = c(0.85,0.15),
        text = element_text(size=14))
figure_nums(name="cumunc_stroke_alternative", 'Cumulative probabilities over time of moving to ECMO, stroke (three types), discharge and death states') # title
alt_plot
#jpeg(filename='figures/stroke/cumulative_ecmo_stroke_alternative.jpg', width=5, height=4, units='in', res=400, quality = 100)
#print(alt_plot)
#invisible(dev.off())
```

This is an alternative arrangement of the previous plot. In this plot, the panels are the state moved to and the lines are coloured by the state moved from. Note the scales on the y-axes vary greatly by panel.

The risks for death change greatly over time, with patients in ICU have an initially higher risk of death followed by a lower risk after around 30 days (compared with patients who have come off ECMO). There is a relatively higher risk of hemorrhagic stroke for those on ECMO compared with those in a standard bed.

### Cumulative risk plot for stroke split for all strokes combined

```{r, fig.width=9, fig.height=5.5}
# cumulative incidence
cum_inc = with(my.data, cuminc(ftime=exit, fstatus=to, group=from, cencode='censored'))
# remove lines not needed (e.g., stroke to stroke)
any_ests = NULL
for (z in 1:(length(cum_inc)-1)){
  empty = any(cum_inc[[z]]$est !=0) # find missing
  any_ests = c(any_ests, empty)
}
index = rev(which(any_ests == FALSE))
for (z in index){
  cum_inc[[z]] = NULL # blank lines
}
# plot
g = ggcompetingrisks(cum_inc,
                 xlab = 'Days since ICU admission',
                 ylab = 'Cumulative probability',
                 xlim = c(0, censor_day), # 
                 title = '',
#                 gsep = '_', # dummy to prevent groups being created from spaces
                 cumcensor = TRUE, # does not work
                 legend.title='',
                 risk.table = TRUE,
                 conf_int = TRUE)
# create plot with better facet labels
to_plot = mutate(g$data, 
                 facet = paste('State moved from: ', group, sep=''))
cum_plot = ggplot(data=to_plot, aes(x=time, y=est, col=factor(event)))+
  geom_line(size = 1.05) + # make lines thicker
  facet_wrap(~facet, scales='free_y')+
  xlab('Days since ICU admission')+
  ylab('Cumulative probability')+
  theme_bw()+
  scale_x_continuous(breaks=c(0, 20, 40, 60, 80))+
  scale_y_continuous(breaks=seq(0, 0.9, 0.1))+
  scale_colour_manual(name="State moved to:", values=cbPalette) +  # change labels?
  theme(
    panel.grid.minor = element_blank(),
    panel.grid.major = element_line(colour=grey(0.8)),
    text = element_text(size=20), 
    legend.position = 'top',
    legend.title = element_text(size=16),
    legend.text = element_text(size=16))
figure_nums(name="cuminc_stroke_combined", "Cumulative probabilities over time for all strokes combined.")
cum_plot
jpeg('figures/stroke/competing_risks_stroke_combined.jpg', width=6.8, height=6, units='in', res=400, quality=100)
print(cum_plot)
invisible(dev.off())
```

The scales on the y-axis vary by panel. 

## Survival models

Here we model the survival time and outcome (discharge/death) with the aim of examining stroke whilst adjusting for confounders.

BMI was included as a predictor and the around 10% of missing values were randomly imputed as part of the Bayesian model.

```{r, winbugs_model_death_discharge}
# bugs model
bfile = 'bugs_survival_stroke_death.txt'
bugs = file(bfile, 'w')
cat('model{
    for (i in 1:N){
      days[i] ~ dweib(r, mu[i])I(cens[i],) # right-censoring
      log(mu[i]) <- int + beta[1]*sex[i] + beta[2]*age[i] + beta[3]*bmi[i] + beta[4]*stroke_o[i] + beta[5]*stroke_i[i] + beta[6]*stroke_h[i] + centre[site[i]]
      bmi[i] ~ dnorm(mu.bmi, tau.bmi) # missing BMI
    }
    # priors
    mu.bmi ~ dnorm(0, 0.1) # vaguely informative (scaled)
    tau.bmi ~ dgamma(1, 1)
    int ~ dnorm(0, 0.0001)
    for(k in 1:6){beta[k] ~ dnorm(0, 0.0001)}
    for(k in 1:M){ # random effect for site
      centre[k] ~ dnorm(0, tau.centre)
    }
    tau.centre ~ dgamma(1, 1)
    r ~ dgamma(10, 10)
}', file=bugs)
close(bugs)
```

#### Time to death and discharge

```{r, include=FALSE}
## prepare the data for 'winbugs' or 'R', need to use clock reset method
prepare_survival = function(indata, 
                            events, # what are the events, e.g. 'death'
                            clock_reset= TRUE, version='winbugs'){
  bugs_data = mutate(indata,
    site_num = as.numeric(as.factor(site_name)), # make site number
    # clock reset
    exit = ifelse(entry > 0, exit - entry, exit), # subtract starting days if delayed start
    # add tiny time because time zero not allowed
    days = ifelse(to %in% events == FALSE, NA, exit+0.001), # A right censored observation has a missing observed time ...
    cens = ifelse(to %in% events == FALSE, exit+0.001, 0) # ... and a minimum time equal to the censored time
                   )
#
if(version=='R'){
  return(bugs_data)
}
if(version=='winbugs'){
N = nrow(bugs_data)
M = max(bugs_data$site_num)

# set up survival data for winbugs
bdata = list(N = N, 
             M = M,
             site = bugs_data$site_num,
             days = bugs_data$days,
             cens = bugs_data$cens,
             stroke_o = as.numeric(bugs_data$from=='stroke'),
             stroke_i = as.numeric(bugs_data$from=='stroke_i'),
             stroke_h = as.numeric(bugs_data$from=='stroke_h'),
             sex = as.numeric(bugs_data$sex=='Male'),
             age = (bugs_data$age-60)/10, # scaled
             bmi = (bugs_data$bmi-30)/5) # scaled
  return(bdata)
} # end of if
} # end of function
# add stroke types
my.data.with.stroke.type = mutate(my.data, # update to state to allow multiple stroke types
         from = ifelse(from=='stroke' & stroke_group=='Ischemic', 'stroke_i', from), # 
         from = ifelse(from=='stroke' & stroke_group=='Hemorrhage', 'stroke_h', from),
         to = ifelse(to=='stroke' & stroke_group=='Ischemic', 'stroke_i', to), # 
         to = ifelse(to=='stroke' & stroke_group=='Hemorrhage', 'stroke_h', to)
         )
# create data
for_death = prepare_survival(indata=my.data.with.stroke.type, events=c('death')) #  event is death 
for_discharge = prepare_survival(indata=my.data.with.stroke.type, events=c('discharge')) # 
```

```{r, include=FALSE}
## run the survival models in WinBUGS
# could remove individual effect, it is not helping much
n.chains= 2
MCMC = 5000 # increase to 10,000 for final run
thin = 3 # increase to 5
debug = FALSE

# initial values for single variable outcome (use mean for intercept)
inits_start = list(r=1, beta=rep(0,6), int=0, tau.centre=1, centre=rep(0, for_death$M), mu.bmi=0, tau.bmi=1) 
parameters = c('r','beta','int','centre','tau.centre') # 
  
##
# a) death
# do not run model if there are already estimates (save time)
any_model = length(dir(path='results', pattern='bugs_death_stroke')) > 0
# run model
if(any_model == FALSE){
  # add initial values for days that are beyond censored time
  initial_days = for_death$cens+1
  initial_days[for_death$cens ==0] = NA
  inits = inits_start
  inits$days = initial_days
  inits = rep(list(inits), n.chains) # repeat initial values per chain
  # run model
  bugs_results_death = bugs(data=for_death, inits=inits, parameters=parameters, model.file=bfile, DIC=FALSE, n.chains=n.chains, n.iter=MCMC*thin*2, n.thin=thin, bugs.seed=87947, debug=debug, bugs.directory=bugs_location)
  save(bugs_results_death, file='results/survival_bugs_death_stroke.RData')
}
if(any_model == TRUE){
  load('results/survival_bugs_death_stroke.RData')
}
# b) discharge
any_model = length(dir(path='results', pattern='bugs_discharge_stroke')) > 0
# run model
if(any_model == FALSE){
  # add initial values for days that are beyond censored time
  initial_days = for_discharge$cens+1
  initial_days[for_discharge$cens ==0] = NA
  inits = inits_start
  inits$days = initial_days
  inits = rep(list(inits), n.chains) # repeat initial values per chain
  # run model
  bugs_results_discharge = bugs(data=for_discharge, inits=inits, parameters=parameters, model.file=bfile, DIC=FALSE, n.chains=n.chains, n.iter=MCMC*thin*2, n.thin=thin, bugs.seed=87947, debug=debug, bugs.directory=bugs_location)
  save(bugs_results_discharge, file='results/survival_bugs_discharge_stroke.RData')
}
if(any_model == TRUE){
  load('results/survival_bugs_discharge_stroke.RData')
}
# 
```

The table below is the hazard ratios (HRs) and 95% credible intervals for the competing risks of death and discharge. 
A hazard ratio of 1 indicates no change in risk, whereas a hazard ratio above 1 indicates increased risk, and a hazard ratio below 1 indicated decreased risk.

```{r}
## tabulate the results
# get beta estimates
beta_death = bugs_extract(inbugs=bugs_results_death, var='beta', exp=TRUE, names=c('Male','Age (+10 years)','BMI (+5 kg/m2)','Stroke - not specified','Stroke - Ischemic','Stroke - Hemorrhage'), digits=2, make.ci=TRUE) # from 99_functions.R
beta_discharge = bugs_extract(inbugs=bugs_results_discharge, var='beta', exp=TRUE, names=NULL, digits=2, make.ci=TRUE) # from 99_functions.R
for_table = bind_cols(beta_death, beta_discharge) %>%
  select(-row, -Var) 
names(for_table) = c('var','mean1','ci1','mean2','ci2')
#
header = read.table(header=TRUE, stringsAsFactors = FALSE, sep="!", text='
col_keys!colB!colA
var! !Variable
mean1!Death!HR
ci1!Death!95% CI
mean2!Discharge!HR
ci2!Discharge!95% CI
')
#
ftab = flextable(for_table) %>%
  set_header_df(mapping = header, key = "col_keys" ) %>%
  merge_h(part = "header", i = 1) %>%
  theme_box() %>% # nice theme 
  align(align='center', part='all') %>% # centre text
  autofit()
table_nums(name="hr_weibull", "Hazard ratios and 95% credible intervals for death and discharge from the Weibull model", level=1)
ftab
# get number administratively censored
censored = group_by(my.data, pin) %>%
  arrange(pin, entry) %>%
  slice(n()) %>% # last per patient
  filter(to == '0') # admin censored
```

Older age increases the hazard of death and decreases the hazard of discharge. 

There is a very strong increase in the hazard of death for hemorrhagic stroke, and a strong increase for 'not specified' stroke. The uncertainty in the hazard ratios is relatively wide, which is due to the relatively small number of strokes.

## Cumulative survival models

The survival models in this section focus on the patients' final (or cumulative) outcome of death and discharge, whereas the previous models examined the instantaneous hazard of experiencing death or discharge. 

The table below shows the cumulative hazard ratios and 95% confidence intervals for death and discharge.

```{r}
my.data_crr = mutate(my.data, 
   # update to state to allow multiple stroke types
   from = ifelse(from=='stroke' & stroke_group=='Ischemic', 'stroke_i', from), 
   from = ifelse(from=='stroke' & stroke_group=='Hemorrhage', 'stroke_h', from),
   to = ifelse(to=='stroke' & stroke_group=='Ischemic', 'stroke_i', to), 
   to = ifelse(to=='stroke' & stroke_group=='Hemorrhage', 'stroke_h', to),
  # clock reset
  exit = ifelse(entry > 0, exit - entry, exit), # subtract starting days if delayed start
  # prepare variables
   stroke_other = as.numeric(from=='stroke'), # # 2= other, 3=Ischemic, 4=Hemorrhage
   stroke_ischemic = as.numeric(from=='stroke_i'),
   stroke_hemorrhage = as.numeric(from=='stroke_h'),
   sex = as.numeric(sex=='Male'))
# small amounts of missing
my.data_crr = mutate(my.data_crr,
          bmi_c = ifelse(is.na(bmi_c)==TRUE, 0, bmi_c)) # use mean
#
cov1 = as.matrix(select(my.data_crr, sex, age_c, bmi_c, stroke_other, stroke_ischemic, stroke_hemorrhage))
# death model (to = 6)
cum_rr_death = with(my.data_crr, crr(ftime=exit, fstatus=to=='death', cov1=cov1, cencode='censored', maxiter = 100))
# discharge model (to =5)
cum_rr_discharge = with(my.data_crr, crr(ftime=exit, fstatus=to=='discharge', cov1=cov1, cencode='censored', maxiter = 100))
s_death = summary(cum_rr_death) # cannot use broom
s_discharge = summary(cum_rr_discharge) # cannot use broom
table_death = as.data.frame(s_death$coef); names(table_death) = c('coef','exp_coef','SE','z','pval')
table_discharge = as.data.frame(s_discharge$coef); names(table_discharge) = c('coef','exp_coef','SE','z','pval')
tab = bind_rows(table_death, table_discharge, .id='type') # combine tables
table = mutate(tab,
         z = qnorm(0.975),
         lower = coef - (z*SE), # make CIs
         upper = coef + (z*SE),
         HR = roundz(exp(coef), 2),
         lower = roundz(exp(lower), 2),
         upper = roundz(exp(upper), 2),
         CI = paste(lower, ' to ', upper, sep='')) %>%
  select(type, HR, CI) 
ftable = bind_cols(filter(table, type==1) %>% select(-type),
                  filter(table, type==2) %>% select(-type))
names(ftable) = c('HR1','CI1','HR2','CI2')
#
table_header <- data.frame(
  col_keys = c('Variable', 'HR1','CI1','HR2','CI2'),
  h2 = c('', 'Death','Death','Discharge','Discharge'),
  h3 = c('Variable', 'HR','95% CI', 'HR','95% CI'),
  stringsAsFactors = FALSE )
# nice table
ftable = mutate(ftable, Variable=c('Male','Age (+10 years)','BMI (+5 kg/m2)',"Stroke - other","Stroke - Ischemic","Stroke - Hemorrhage")) %>%
  select('Variable', everything()) %>% # move Variable first
  flextable() %>%
  set_header_df(mapping = table_header, key = "col_keys") %>% # add header
  merge_h(i=1, part = "header") %>% # merge  column headers
  theme_box() %>% # nice theme
  align(align='center', part='all') %>% # centre text
  autofit() 
table_nums(name="crr_hazards", "Hazard ratios and 95% confidence intervals using a cumulative probability regression model")
ftable
```

Older age increases the cumulative risk of death and decreases the cumulative risk of discharge. 

Hemorrhagic stroke were associated with a greatly increased probability of death, and a reduced probability of discharge.

## Anticoagulation type at the time of stroke

```{r}
# function to run case-control analysis depending on stroke type
case_control = function(s_type, number_of_controls = 5){
set.seed(212121) # for random control selection
# split into cases and controls by stroke
cases = filter(patients, 
               !is.na(complication_stroke_date), # has a stroke ...
               stroke_group %in% s_type) %>% # ... of this type
  select(pin, age, complication_stroke_date, date_icu) %>%
  mutate(days = as.numeric(complication_stroke_date - date_icu)) %>%
  filter(!is.na(days),
         days > 0) %>% # stroke must be after ICU admission
  mutate(days_c = case_when( # combine days as they get sparse
    days > 20 & days<=30 ~ 25,
    days > 30 & days<=40 ~ 35,
    days > 40 & days<=50 ~ 45,
    days > 50 ~ 55,
    TRUE ~ days
  ))
controls = filter(patients, 
                  !pin %in% unique(cases$pin), # not in the stroke case data
                  !is.na(date_icu)) %>% # with ICU data
  select(pin, age, date_icu)
# merge cases with daily data
cases = left_join(cases, daily, by=c('pin'='pin', 'complication_stroke_date'='date_daily'))
# merge controls with daily data (take all days)
controls = left_join(controls, daily, by=c('pin'='pin')) %>%
  mutate(days = as.numeric(date_daily - date_icu)) %>%
  filter(!is.na(days),
         days > 0) %>%
  mutate(days_c = case_when( # combine days as they get sparse
    days > 20 & days<=30 ~ 25,
    days > 30 & days<=40 ~ 35,
    days > 40 & days<=50 ~ 45,
    days > 50 ~ 55,
    TRUE ~ days
  ))
# match controls to case:
matched = NULL
for (k in 1:nrow(cases)){
  this_case = cases[k,] %>%
    mutate(case=1)
  random_controls = filter(controls, 
                           days_c == this_case$days_c, # same (grouped) day in ICU
                           age >= (this_case$age - 2), # age within 2 years
                           age <= (this_case$age + 2)) %>%
    sample_n(number_of_controls) %>% # randomly sample
    mutate(case = 0)
  comb = bind_rows(this_case, random_controls) %>%
    mutate(pair = k)
  matched = bind_rows(matched, comb)
}
#
table = group_by(matched, case, eotd_anticoagulants) %>%
  tally() %>%
  group_by(case) %>%
  mutate(percent = round(prop.table(n)*100),
         cell = paste(n, ' (', percent, ')', sep='')) %>%
  select(-n, -percent) %>%
  ungroup() %>%
  spread(case, cell) %>%
  mutate(eotd_anticoagulants = case_when(
    eotd_anticoagulants == 0 ~ 'No',
    eotd_anticoagulants == 1 ~ 'Yes',
    is.na(eotd_anticoagulants) ~ 'Missing'
  ))
#
# table header
table_header <- data.frame(
  col_keys = c('eotd_anticoagulants','0','1'),
  h1 = c('', 'Stroke', 'Stroke'),
  h2 = c('Anticoagulants','No','Yes'),
  stringsAsFactors = FALSE )
#
ftab = flextable(table) %>% 
  set_header_df(mapping = table_header, key = "col_keys" ) %>% # add header
 merge_h(i=1, part = "header") %>% # merge column headers
  theme_box() %>%
  autofit()
# chi-squared test
  tab = with(matched, table(case, eotd_anticoagulants))
  chisq = chisq.test(tab, simulate.p.value = TRUE)
# return
  to_return = list()
  to_return$chisq = chisq
  to_return$table = ftab
  to_return$matched = matched
  return(to_return)
} # end of function
#
cc_Hemorrhage = case_control(s_type = 'Hemorrhage')
table_nums(name="anti_table_h","Anticoagulants on the day of hemorrhagic stroke compared with matched controls.", level=1)
cc_Hemorrhage$table
```

The table shows the anticoagulant use on the day of hemorrhagic stroke. The controls without hemorrhagic stroke were matched to the cases based on age (within 2 years) and days since ICU admission. We used five randomly selected controls per case. The cells are the number and percentage.

A chi-squared test on the above table gives X2=`r roundz(cc_Hemorrhage$chisq$statistic,1)` with p-value=`r format.pval(cc_Hemorrhage$chisq$p.value, digits=2, eps=0.001)`.

```{r}
#
cc_ischemic = case_control(s_type = 'Ischemic')
table_nums(name="anti_table_i","Anticoagulants on the day of ischemic stroke compared with matched controls.")
cc_ischemic$table
```

```{r}
#
matched = mutate(cc_Hemorrhage$matched,
                 eotd_anticoagulants_type = ifelse(is.na(eotd_anticoagulants_type), '', eotd_anticoagulants_type))
table = group_by(matched, case, eotd_anticoagulants_type) %>%
  tally() %>%
  group_by(case) %>%
  mutate(percent = round(prop.table(n)*100),
         cell = paste(n, ' (', percent, ')', sep='')) %>%
  select(-n, -percent) %>%
  ungroup() %>%
  spread(case, cell) %>%
  mutate(eotd_anticoagulants_type = ifelse(eotd_anticoagulants_type=='', 'Missing', eotd_anticoagulants_type))
#
# table header
table_header <- data.frame(
  col_keys = c('eotd_anticoagulants_type','0','1'),
  h1 = c('', 'Stroke', 'Stroke'),
  h2 = c('Anticoagulant type','No','Yes'),
  stringsAsFactors = FALSE )
#
ftab = flextable(table) %>% 
  set_header_df(mapping = table_header, key = "col_keys" ) %>% # add header
 merge_h(i=1, part = "header") %>% # merge column headers
  theme_box() %>%
  autofit()
table_nums(name="anti_table_type_h","Anticoagulant type on the day of hemorrhagic stroke compared with matched controls.") # title
ftab
```

This table shows the type of anticoagulant on the day of hemorrhagic stroke.

```{r}
#
matched = mutate(cc_ischemic$matched,
                 eotd_anticoagulants_type = ifelse(is.na(eotd_anticoagulants_type), '', eotd_anticoagulants_type))
table = group_by(matched, case, eotd_anticoagulants_type) %>%
  tally() %>%
  group_by(case) %>%
  mutate(percent = round(prop.table(n)*100),
         cell = paste(n, ' (', percent, ')', sep='')) %>%
  select(-n, -percent) %>%
  ungroup() %>%
  spread(case, cell) %>%
  mutate(eotd_anticoagulants_type = ifelse(eotd_anticoagulants_type=='', 'Missing', eotd_anticoagulants_type))
#
# table header
table_header <- data.frame(
  col_keys = c('eotd_anticoagulants_type','0','1'),
  h1 = c('', 'Stroke', 'Stroke'),
  h2 = c('Anticoagulant type','No','Yes'),
  stringsAsFactors = FALSE )
#
ftab = flextable(table) %>% 
  set_header_df(mapping = table_header, key = "col_keys" ) %>% # add header
 merge_h(i=1, part = "header") %>% # merge column headers
  theme_box() %>%
  autofit()
table_nums(name="anti_table_type_i","Anticoagulant type on the day of hemorrhagic stroke compared with matched controls.") # title
ftab
```

### Platelet count on the day of stroke

```{r}
# combine two stroke types
to_count = bind_rows(cc_Hemorrhage$matched, cc_ischemic$matched, .id='s_type') %>%
  select(s_type, case, pair, platelet_count) %>%
  mutate(s_type = ifelse(s_type==1, 'Hemorrhage', 'Ischemic'),
         case = ifelse(case==1, 'Case', 'Control'))
# make table
table = group_by(to_count, s_type, case) %>%
  summarise(N = n(), 
            Missing = sum(is.na(platelet_count)), 
            mean=mean(platelet_count, na.rm=TRUE),
            sd=sd(platelet_count, na.rm=TRUE)) %>%
  ungroup() %>%
  mutate(N = N - Missing, # subtract missing
         mean = round(mean*10)/10,
         sd = round(sd*10)/10)
ftab = flextable(table) %>%
  theme_box() %>%
  autofit()
ftab
```

The table shows the number non-missing, the number missing, and the mean and variance.

```{r}
bplot = ggplot(data=filter(to_count, !is.na(platelet_count)), aes(x=factor(case), y=platelet_count, fill=factor(case)))+
  geom_boxplot()+
  ylab('Platelet count')+
  facet_wrap(~s_type)+
  theme(legend.position = 'none')
bplot
```

# Does ECMO increase the risk of stroke?

As ECMO is a time-dependent exposure we need to use a multi-state model to examine this question.

```{r, fig.height=6, fig.width=6}
states = c('ICU\nnon-ECMO','ECMO','Post-ECMO','Stroke','Death','Hospital\ndischarge')
# Matrix of joins between states:
M = matrix(data=c(
  2,1,
  3,2,
  4,1,
  4,2,
  4,3,
  5,1,
  5,2,
  5,3,
  5,4,
  6,4,
  6,3,
  6,2,
  6,1
), byrow = TRUE, ncol=2)
figure_nums(name="model_diagram_ecmo", "The six states in the model and the possible movements between states", level=1) # title
multistate_diagram(states=states, links = M, pos = c(3,1,2), arr.pos=0.65) # see 99_functions.R
```


### Table of transitions between states

This table shows the transitions (movements) between the states (`r figure_nums("model_diagram_ecmo", display = "cite")`).

```{r}
states = c('from', 'date_ecmo','ecmo_end','stroke','discharge','death','censored') # columns in order
tab = as.data.frame(with(my.data, table(from, to))) %>%
  tidyr::spread(to, Freq) %>%
  select(all_of(states))
# table header
table_header <- data.frame(
  col_keys = states,
  h1 = c('', rep("State moved to", 6)),
  h2 = c('State moved from', 'ECMO start','ECMO end','Stroke','Discharge','Death','Censored'),
  stringsAsFactors = FALSE )
ftab = flextable(tab) %>%
  set_header_df(mapping = table_header, key = "col_keys" ) %>% # add header
  merge_h(i=1, part = "header") %>% # merge "To" column headers
  theme_box() %>%
  align(align='center', part='all') %>% # centre text
  autofit() 
table_nums(name="transitions_ecmo", "Transition numbers between states") # title
ftab
```

## Movements between states over time

```{r, fig.width=9, fig.height=6.5}
# cumulative incidence
cum_inc = with(my.data, cuminc(ftime=exit, fstatus=to, group=from, cencode='censored'))
# remove lines not needed (e.g., stroke to stroke)
any_ests = NULL
for (z in 1:(length(cum_inc)-1)){
  empty = any(cum_inc[[z]]$est !=0) # find missing
  any_ests = c(any_ests, empty)
}
index = rev(which(any_ests == FALSE))
for (z in index){
  cum_inc[[z]] = NULL # blank lines
}
# plot
g = ggcompetingrisks(cum_inc,
                     xlab = 'Days since ICU',
                     ylab = 'Cumulative probability',
                     xlim = c(0,censor_day), # 
                     title = '',
                     # gsep = '_', # dummy to prevent groups being created from spaces
                     cumcensor = TRUE, # does not work
                     legend.title='',
                     risk.table = TRUE,
                     conf_int = TRUE)
# create plot with better facet labels
to_plot = mutate(g$data, 
                 facet = paste('State moved from: ', group, sep=''))
cum_plot = ggplot(data=to_plot, aes(x=time, y=est, col=factor(event)))+
  geom_line(size = 1.05) + # make lines thicker
  facet_wrap(~facet, scales='free_y')+
  xlab('Days since ICU')+
  ylab('Cumulative probability')+
  theme_bw()+
  scale_x_continuous(breaks=c(0, 20, 40, 60, 80))+
  scale_y_continuous(breaks=seq(0, 0.9, 0.1))+
  scale_colour_manual(name="State moved to:", values=cbPalette) +  # change labels?
  theme(
    panel.grid.minor = element_blank(),
    panel.grid.major = element_line(colour=grey(0.8)),
    text = element_text(size=20), 
    legend.position = 'top',
    legend.title = element_text(size=16),
    legend.text = element_text(size=16))
figure_nums(name="cumunc_stroke_ecmo", 'Cumulative probabilities over time of moving to ECMO, stroke, discharge and death states') # title
cum_plot
jpeg(filename='figures/stroke/cumulative_ecmo_stroke.jpg', width=5, height=4, units='in', res=400, quality = 100)
print(cum_plot)
invisible(dev.off())
```

The scales on the y-axes vary by panel. 

```{r, fig.width=8.5, fig.height=6.5}
# 
to_plot = mutate(g$data, event = paste('State moved to:', event))
# alternative plot
alt_plot = ggplot(data= to_plot, aes(x=time, y=est, col=factor(group)))+
  geom_line(size=1.05)+
  xlab('Days since ICU')+
  ylab('Cumulative probability')+
#scale_color_manual('State moved from:', values=(c("#CD950C", "#66CD00", "#FF00FF", "#ADD8E6")))+
  facet_wrap(~event, scale='free_y')+
  theme(legend.position = c(0.85,0.15),
        text = element_text(size=12))
figure_nums(name="cumunc_stroke_ecmo_alternative", 'Cumulative probabilities over time of moving to ECMO, stroke, discharge and death states') # title
alt_plot
jpeg(filename='figures/stroke/cumulative_ecmo_stroke_alternative.jpg', width=5, height=4, units='in', res=400, quality = 100)
print(alt_plot)
invisible(dev.off())
```

This is an alternative arrangement of the previous plot. In this plot, the panels are the state moved to and the lines are coloured by the state moved from. Note the scales on the y-axes vary by panel.

The higher risk of death post stroke and post-ECMO is clear in this figure (compared with patients remainin in the ICU state). Also clear is the lower risk of discharge.

Most transitions to ECMO happen within the first 20 days.

### Cox survival models of time to stroke


```{r}
for_stroke = mutate(my.data,
                    from = relevel(factor(from), ref='icu')) %>%
  filter(from!='stroke') # exclude stroke for this analysis
#
c_stroke = coxph(Surv(time=entry, time2=exit, event= to=='death') ~ I(sex=='Male') + age_c + from + cluster(site_name), data=for_stroke)
digits = 2
ests_stroke = tidy(c_stroke, conf.int = TRUE, exponentiate = TRUE) %>%
  filter(!str_detect(string=term, pattern = 'fromstroke')) %>%
    mutate(
      term = case_when(
        str_detect(term, 'Male') ==TRUE ~ 'Male',
        str_detect(term, 'age')==TRUE ~ 'Age (+10 years)',
        str_detect(term, 'fromecmo_start')==TRUE ~ 'During ECMO',
        str_detect(term, 'fromecmo_end')==TRUE ~ 'Post-ECMO'
      ),
      estimate = roundz(estimate, digits),
      conf.low = roundz(conf.low, digits),
      conf.high = roundz(conf.high, digits),
      CI = paste(conf.low, ' to ', conf.high, sep='')) %>%
    rename('HR' = 'estimate',
           'CI' = 'CI') %>%
  select('term','HR','CI')
ftab = flextable(ests_stroke) %>%
  theme_box() %>%
  autofit()
table_nums(name="cox_stroke_ecmo", 'Cox survival model of stroke with ECMO and post-ECMO as a time-dependent exposures') # title
ftab
```

Being on ECMO dramatically reduces the hazard of stroke, but the risk of stroke is far higher for patients after they come off ECMO.
The reference group for during ECMO and post-ECMO is patients in a standard ICU bed.

### Cox survival models of times to death and discharge

```{r}
# reference category
my.data = mutate(my.data, from = relevel(factor(from), ref='icu'))
# death
c_death = coxph(Surv(time=entry, time2=exit, event= to=='death') ~ I(sex=='Male') + age_c + from + cluster(site_name), data=my.data)
# discharge
c_discharge = coxph(Surv(time=entry, time2=exit, event= to=='discharge') ~ I(sex=='Male') + age_c + from + cluster(site_name), data=my.data)
#
ests_death = tidy(c_death, conf.int = TRUE, exponentiate = TRUE)
ests_disch = tidy(c_discharge, conf.int = TRUE, exponentiate = TRUE) 
ests_d = bind_rows(ests_death, ests_disch, .id='outcome') %>%
    mutate(
      outcome = ifelse(outcome==1, 'Death','Discharge'),
      term = case_when(
        str_detect(term, 'Male') ==TRUE ~ 'Male',
        str_detect(term, 'age')==TRUE ~ 'Age (+10 years)',
        str_detect(term, 'fromecmo_start')==TRUE ~ 'During ECMO',
        str_detect(term, 'fromecmo_end')==TRUE ~ 'Post-ECMO',
        str_detect(term, 'fromstroke')==TRUE ~ 'Stroke',
        TRUE ~ as.character(term)
      ),
      estimate = roundz(estimate, digits),
      conf.low = roundz(conf.low, digits),
      conf.high = roundz(conf.high, digits),
      CI = paste(conf.low, ' to ', conf.high, sep='')) %>%
    rename('HR' = 'estimate',
           'CI' = 'CI') %>%
  select('outcome','term','HR','CI')
ftab = flextable(ests_d) %>%
  theme_box() %>%
  autofit()
table_nums(name="cox_stroke_ecmo_death", 'Cox survival models for death and discharge with stroke, ECMO and post-ECMO as time-dependent exposures') # title
ftab
```

This table is for all stroke types combined.

### Cumulative subdistribution hazard models of death and discharge

```{r}
# prepare data
for_cum = mutate(my.data,
                    exit_reset = exit - entry, # clock reset
                    sexn = as.numeric(sex=='Male'),
                    stroke = as.numeric(from=='stroke'),
                    ecmo_start = as.numeric(from=='ecmo_start'),
                    post_ecmo = as.numeric(from=='ecmo_end'))
#
cov1 = as.matrix(select(for_cum, sexn, age_c, stroke, ecmo_start, post_ecmo))
col_names = c('Sex = male','Age (+10 years)','Stroke','During ECMO','Post-ECMO') # used later
cum_rr_death = with(for_cum, crr(ftime=exit_reset, fstatus= to=='death', cov1=cov1, cencode='censored', maxiter = 100))
cum_rr_discharge = with(for_cum, crr(ftime=exit_reset, fstatus= to=='discharge', cov1=cov1, cencode='censored', maxiter = 100))
#
s_death = summary(cum_rr_death) # cannot use broom
s_discharge = summary(cum_rr_discharge) # cannot use broom
table_death = as.data.frame(s_death$coef); names(table_death) = c('coef','exp_coef','SE','z','pval')
table_discharge = as.data.frame(s_discharge$coef); names(table_discharge) = c('coef','exp_coef','SE','z','pval')
tab = bind_rows(table_death, table_discharge, .id='type') # combine tables
table = mutate(tab,
               z = qnorm(0.975),
               lower = coef - (z*SE), # make CIs
               upper = coef + (z*SE),
               HR = roundz(exp(coef), 2),
               lower = roundz(exp(lower), 2),
               upper = roundz(exp(upper), 2),
               CI = paste(lower, ' to ', upper, sep='')) %>%
  select(type, HR, CI) 
ftable = bind_cols(filter(table, type==1) %>% select(-type),
                   filter(table, type==2) %>% select(-type))
names(ftable) = c('HR1','CI1','HR2','CI2')
ftable = mutate(ftable, Variable=col_names) %>%
  select('Variable', everything()) %>% # move Variable first
  flextable() %>%
  add_header('Variable'='', 'HR1' = 'Death', 'CI1' = 'Death', 'HR2' = 'Discharge', 'CI2' = 'Discharge') %>%
  merge_h(i=1, part = "header") %>% # merge "To" column headers
  theme_box() %>% # nice theme
  align(align='center', part='all') %>% # centre text
  autofit() 
table_nums(name="crr_hazards_death", "Hazard ratios and 95% confidence intervals using a cumulative probability regression model for death and discharge")
ftable
```

Older age increased the risk of death and decreased the risk of discharge.

### Plot of patient probabilities over time

```{r, include=FALSE}
# Set transition matrix for mstate
tra <- transMat(x = list(c(2,4,5,6), c(3,4,5,6), c(4,5,6), c(3,5,6),c(),c()),
                names = c("icu","ecmo_start", "ecmo_end","stroke","discharge","death"))
# collapse ecmo end to icu
my.data.ext = mutate(my.data, 
                 to = ifelse(to=='date_ecmo', 'ecmo_start', to))
# transition must by numbers for ext_mstate
rownames(tra) = colnames(tra) = 1:6
my.data.ext = mutate(my.data.ext,
                 to = case_when(
  to=='censored' ~ 0,
  to=='icu' ~ 1,
  to=='ecmo_start' ~ 2,
  to=='ecmo_end' ~ 3,
  to=='stroke' ~ 4,
  to=='discharge' ~ 5,
  to=='death' ~ 6
),
from = case_when(
  from=='icu' ~ 1,
  from=='ecmo_start' ~ 2,
  from=='ecmo_end' ~ 3,
  from=='stroke' ~ 4,
  from=='discharge' ~ 5,
  from=='death' ~ 6
))

# Add transition vectors (not censored)
my.data.ext$trans <- NA
for (i in rownames(tra))
{
  for (j in rownames(tra))
  {
    my.data.ext$trans[which(my.data.ext$from == i & my.data.ext$to ==j)] <- tra[i, j]  
  }
}
with(my.data.ext, table(from, to))

# Add status vector, indicates observed transition
my.data.ext$status <- 1
# Status vector for censored observations set to '0'
my.data.ext$status[my.data.ext$to == '0'] <- 0
# Rename 'to' == 0  to 'cens' for use in 'ext_mstate' function
my.data.ext$to[my.data.ext$to == '0'] <- 'cens'
# Create data frame with all possible transitions for when a patient is at risk
my.data_ext <- ext_mstate(my.data.ext, tra, id.var='pin')
```

```{r, death_probs}
# get starting numbers
starts = filter(my.data.ext, entry==0) %>% 
  group_by(from) %>% 
  summarise(n=n()) %>% 
  ungroup() %>% 
  mutate(p = prop.table(n))
init_dis_2 <- c(starts$p, 0, 0) # to update (last two zeros are death and discharge)

## Cox model stratified by transition
c_2 <- coxph(Surv(entry, exit, status==1) ~ strata(trans), data= my.data_ext, method = "breslow")
# msfit calculates baseline hazards
msf_2 <- msfit(c_2, trans = tra)
# probtrans calculates transition probabilities, prediction from day 0
pt_2 <- probtrans(msf_2, predt = 0)

# Create weighted average of transition probabilities for progress of entire cohort
# Used to produce full cohort results in Table 3 
pt_fc_2 <- pt_2
fc_2 <- pt_fc_2[[1]] * starts$p[1] + pt_fc_2[[2]] * starts$p[2]
pt_fc_2[[1]] <- fc_2
```


```{r, fig.width=8, fig.height=6}
figure_nums(name="state_plot", "Plot of predicted probabilities for the six states") # title
plot_it = function(){
  par(mai=c(0.9,0.9,0.5,0.1), las=1) # b, l, t, r; leave room at top for legend
colours = c("khaki1","indianred1","cornflowerblue","purple",'light green','grey')
plot(pt_fc_2, from = 1, 
     ord = c(5,4,3,2,1,6), # 5 = discharge at bottom
     type= "filled",
     cols = colours,
     lwd= 2, xlab = "Days since ICU admission", ylab = "Predicted Probabilities", 
     cex.lab = 1.25,
     legend.pos='top',
     legend = rep("", 6), # no labels in plot
     main= NULL, 
     axes=FALSE)
# better x-axis ticks:
axis(side=2, at=seq(0,1,0.2), labels=TRUE) # y-axis
# 90 days
axis(side=1, at=seq(0,90,10), labels=FALSE)
axis(side=1, at=seq(0,90,5), labels=TRUE, tick=FALSE)
# legend at top
par(xpd=TRUE)
labels = c('ICU','ECMO','Post-ECMO','Stroke','Discharge','Death')
legend(0, 1.12, labels, col=colours, border=NA, fill=colours, bty='n', ncol=3)
}
plot_it()
jpeg(filename='figures/stroke/patients_states_time.jpg', width=5, height=4, units='in', res=400, quality = 100)
plot_it()
invisible(dev.off())
```

The state of "stroke" has very few patients and is hard to see.

## Joint models for time to stroke

Here we use a joint modelling approach to try to evaluate the risk of stroke by modelling repeated patient test results over time and modelling the time to stroke.
We use a Bayesian approach and hence the results are presented with 95% credible intervals instead of 95% confidence intervals.

The models were fitted using a log-logistic distribution. The estimates are time multipliers of the median time, for example, a multipler of 2 for men would mean that the median times for men are twice as long as that for women. So multipliers over 1 mean longer times and reduced risk of stroke, whereas multipliers under 1 mean short times and an increased risk of stroke.

The models ignore missing data with no imputation.

The tables have estimates for two models, one for the longitudinal model, and one for the survival model.

We use 'spaghetti' plots to highlight differences between patients with and without a stroke in terms of the test results measured repeatedly over time.
Each line is the results for one patient.
We show all patients with a stroke and a random selection of 50 patients without a stroke as plotting all patients without a stroke would make the plot too busy.
The blue line is the mean from a simple linear regression of time. 
We use a log scale on the y-axis because of a strong positive skew.
We split the plot according to the patients current state: ICU, During ECMO and Post-ECMO.

```{r, include=FALSE}
## separate datasets depending on stroke type 
# create multi-state data
# a) hemorrhagic
hemorrhagic = mutate(patients,
                     # blank all dates that are not hemorrhagic
     complication_stroke_date = ifelse(stroke_group!='Hemorrhage', NA, complication_stroke_date),
     complication_stroke_date = as.Date(complication_stroke_date, origin='1970-01-01'))
my.data.hemorrhagic = time_transitions_string(indata=hemorrhagic,
                intermediate_prior_ongoing= FALSE,
                censor_day = censor_day,
                start_state = 'date_icu',
                int_state = 'complication_stroke_date',
                int_state_end = NULL,
                discharge_date = 'date_hospital_discharge',
                dates= c('date_icu', 'date_ecmo', "date_ecmo_discontinued", 'complication_stroke_date', 'date_transfer', 'date_death','date_hospital_discharge')) %>% #
  mutate(from !='stroke') # only interested in transitions up to stroke
# rename from and to states
my.data.hemorrhagic = mutate(my.data.hemorrhagic, 
                 to = ifelse(to=='date_transfer', 'censored', to),
                 to = ifelse(to=='date_int_state_start', 'stroke', to),
                 to = ifelse(to=='date_ecmo_discontinued', 'ecmo_end', to),
                 from = ifelse(from=='date_ecmo', 'ecmo_start', from),
                 from = ifelse(from=='date_int_state_start', 'stroke', from),
                 from = ifelse(from=='date_icu', 'icu', from),
                 from = ifelse(from=='date_ecmo_discontinued', 'ecmo_end', from),
                 from = ifelse(from=='icu' & to=='ecmo_end', 'ecmo_start', from)) %>% # assume ECMO at ICU start
  mutate(age_c = (age - 60)/10, # standardising
         bmi_c = (bmi - 28)/5) 
# b) Ischemic          
ischemic = mutate(patients,
                     # blank all dates that are not hemorrhagic
     complication_stroke_date = ifelse(stroke_group!='Ischemic', NA, complication_stroke_date),
     complication_stroke_date = as.Date(complication_stroke_date, origin='1970-01-01'))
my.data.ischemic = time_transitions_string(indata=ischemic,
                intermediate_prior_ongoing= FALSE,
                censor_day = censor_day,
                start_state = 'date_icu',
                int_state = 'complication_stroke_date',
                int_state_end = NULL,
                discharge_date = 'date_hospital_discharge',
                dates= c('date_icu', 'date_ecmo', "date_ecmo_discontinued", 'complication_stroke_date', 'date_transfer', 'date_death','date_hospital_discharge')) %>% #
  mutate(from !='stroke') # only interested in transitions up to stroke
# rename from and to states
my.data.ischemic = mutate(my.data.ischemic, 
                 to = ifelse(to=='date_transfer', 'censored', to),
                 to = ifelse(to=='date_int_state_start', 'stroke', to),
                 to = ifelse(to=='date_ecmo_discontinued', 'ecmo_end', to),
                 from = ifelse(from=='date_ecmo', 'ecmo_start', from),
                 from = ifelse(from=='date_int_state_start', 'stroke', from),
                 from = ifelse(from=='date_icu', 'icu', from),
                 from = ifelse(from=='date_ecmo_discontinued', 'ecmo_end', from),
                 from = ifelse(from=='icu' & to=='ecmo_end', 'ecmo_start', from)) %>% # assume ECMO at ICU start
  mutate(age_c = (age - 60)/10, # standardising
         bmi_c = (bmi - 28)/5) 
```


### Joint model for time to hemorrhagic stroke using APTT

```{r, joint_model_aptt_h}
joint_ests_aptt_hemorrhagic = joint_survival(indata.surv = my.data.hemorrhagic, 
                                 indata.long = daily, 
                                 event_name='stroke',
                                 long_var = 'aptt', 
                                 log_long = TRUE, 
                                 censor_day = censor_day, 
                                 intercept.survival = FALSE)
# table of estimates
table_nums(name="joint_aptt_hemorrhagic","Estimates from the joint model of APTT over time and survival for hemorrhagic stroke.", level=1) # title, increase first number
joint_ests_aptt_hemorrhagic$table
# for text below
mult = filter(joint_ests_aptt_hemorrhagic$ests, model=='Time to stroke', term=='aptt') 

## competing models for death/discharge
joint_ests_aptt_hemorrhagic_death = joint_survival(indata.surv = my.data.hemorrhagic, 
                                 indata.long = daily, 
                                 event_name='death',
                                 long_var = 'aptt', 
                                 log_long = TRUE, 
                                 censor_day = censor_day, 
                                 intercept.survival = FALSE)
joint_ests_aptt_hemorrhagic_discharge = joint_survival(indata.surv = my.data.hemorrhagic, 
                                 indata.long = daily, 
                                 event_name='discharge',
                                 long_var = 'aptt', 
                                 log_long = TRUE, 
                                 censor_day = censor_day, 
                                 intercept.survival = FALSE)
# store estimates
aptt_ests_hemorrhagic = bind_rows(filter(joint_ests_aptt_hemorrhagic_death$ests, is.long==1),
                      filter(joint_ests_aptt_hemorrhagic_discharge$ests, is.long==1), .id ='model') %>%
  mutate(
    stroke_type = 'hemorrhagic',
    model = factor(model, levels=1:2, labels=c('Death','Discharge')))
```

The estimates use `r joint_ests_aptt_hemorrhagic$numbers$N_patients` patients with `r format(joint_ests_aptt_hemorrhagic$numbers$N_long, big.mark=',')` longitudinal observations. 
The median number of longitudinal results per patient was `r joint_ests_aptt_hemorrhagic$numbers$long_median` (inter-quartile range `r joint_ests_aptt_hemorrhagic$numbers$long_q1` to `r joint_ests_aptt_hemorrhagic$numbers$long_q3`). 
The time to stroke model used `r joint_ests_aptt_hemorrhagic$numbers$N_event` strokes.

A higher APTT reduces the risk of stroke. A doubling of APTT multiplies the median time to stroke by `r mult$mean` (95% credible interval `r mult$lower` to `r mult$upper`).

### Joint model for time to ischemic stroke using APTT

```{r, joint_model_aptt}
joint_ests_aptt_ischemic = joint_survival(indata.surv = my.data.ischemic, 
                                 indata.long = daily, 
                                 long_var = 'aptt', 
                                 event_name='stroke',
                                 log_long = TRUE, 
                                 censor_day = censor_day, 
                                 intercept.survival = FALSE)
# table of estimates
table_nums(name="joint_aptt_ischemic","Estimates from the joint model of APTT over time and survival for ischemic stroke.") 
joint_ests_aptt_ischemic$table
# for text below
mult = filter(joint_ests_aptt_ischemic$ests, model=='Time to stroke', term=='aptt') 

## competing models for death/discharge
joint_ests_aptt_ischemic_death = joint_survival(indata.surv = my.data.ischemic, 
                                 indata.long = daily, 
                                 event_name='death',
                                 long_var = 'aptt', 
                                 log_long = TRUE, 
                                 censor_day = censor_day, 
                                 intercept.survival = FALSE)
joint_ests_aptt_ischemic_discharge = joint_survival(indata.surv = my.data.ischemic, 
                                 indata.long = daily, 
                                 event_name='discharge',
                                 long_var = 'aptt', 
                                 log_long = TRUE, 
                                 censor_day = censor_day, 
                                 intercept.survival = FALSE)
# store estimates
aptt_ests_ischemic = bind_rows(filter(joint_ests_aptt_ischemic_death$ests, is.long==1),
                      filter(joint_ests_aptt_ischemic_discharge$ests, is.long==1), .id ='model') %>%
  mutate(
    stroke_type = 'ischemic',
    model = factor(model, levels=1:2, labels=c('Death','Discharge')))
```

The estimates use `r joint_ests_aptt_ischemic$numbers$N_patients` patients with `r format(joint_ests_aptt_ischemic$numbers$N_long, big.mark=',')` longitudinal observations. 
The median number of longitudinal results per patient was `r joint_ests_aptt_ischemic$numbers$long_median` (inter-quartile range `r joint_ests_aptt_ischemic$numbers$long_q1` to `r joint_ests_aptt_ischemic$numbers$long_q3`). The time to stroke model used `r joint_ests_aptt_ischemic$numbers$N_event` strokes.

A higher APTT reduces the risk of stroke. A doubling of APTT multiplies the median time to stroke by `r mult$mean` (95% credible interval `r mult$lower` to `r mult$upper`).


##### Longitudinal plot of APTT for stroke and non-stroke patients (ischemic)

```{r}
to_plot = joint_ests_aptt_ischemic$graph +
  ylab('APTT (log scale)')+
  xlab('Day since ICU admission')
figure_nums(name="joint_aptt","APTT over time for stroke and non-stroke patients.", level=1) # increase first number 
print(to_plot)
```

### Joint model for time to hemorrhagic stroke using ph

ph was the only variable that was not log-transformed as the original distribution did not have a strong skew.

```{r, joint_model_ph_h}
joint_ests_ph_hemorrhagic = joint_survival(indata.surv = my.data.hemorrhagic, 
                                           indata.long = daily, 
                                           long_var = 'p_h', 
                                           log_long = FALSE, 
                                           event_name = 'stroke',
                                           censor_day = censor_day, 
                                           scale = 0.1, 
                                           intercept.survival=FALSE)
# table of estimates
table_nums(name="joint_ph_hemorrhagic","Estimates from the joint model of ph over time and time to hemorrhagic stroke")
joint_ests_ph_hemorrhagic$table
# for text below
mult = filter(joint_ests_ph_hemorrhagic$ests, model=='Time to stroke', term=='p_h') 

## competing models for death/discharge
joint_ests_ph_hemorrhagic_death = joint_survival(indata.surv = my.data.hemorrhagic, 
                                 indata.long = daily, 
                                 event_name='death',
                                 long_var = 'p_h', 
                                 scale = 0.1,
                                 log_long = FALSE, 
                                 censor_day = censor_day, 
                                 intercept.survival = FALSE)
joint_ests_ph_hemorrhagic_discharge = joint_survival(indata.surv = my.data.hemorrhagic, 
                                 indata.long = daily, 
                                 event_name='discharge',
                                 long_var = 'p_h', 
                                 scale = 0.1,
                                 log_long = FALSE, 
                                 censor_day = censor_day, 
                                 intercept.survival = FALSE)
# store estimates
ph_ests_hemorrhagic = bind_rows(
  filter(joint_ests_ph_hemorrhagic_death$ests, is.long==1),
  filter(joint_ests_ph_hemorrhagic_discharge$ests, is.long==1), .id ='model') %>%
  mutate(
    stroke_type = 'hemorrhagic',
    model = factor(model, levels=1:2, labels=c('Death','Discharge')))
```

The estimates use `r joint_ests_ph_hemorrhagic$numbers$N_patients` patients with `r format(joint_ests_ph_hemorrhagic$numbers$N_long, big.mark=',')` longitudinal observations. The time to stroke model used `r joint_ests_ph_hemorrhagic$numbers$N_event` strokes.

### Joint model for time to ischemic stroke using ph

```{r, joint_model_ph}
joint_ests_ph_ischemic = joint_survival(indata.surv=my.data.ischemic, 
                                        indata.long=daily, 
                                        long_var='p_h', 
                                 event_name='stroke',
                                        log_long=FALSE, 
                                        censor_day=censor_day, 
                                        scale=0.1, 
                                        intercept.survival=TRUE)
# table of estimates
table_nums(name="joint_ph_ischemic","Estimates from the joint model of ph over time and time to ischemic stroke")
joint_ests_ph_ischemic$table
# for text below
mult = filter(joint_ests_ph_ischemic$ests, model=='Time to stroke', term=='p_h') 

## competing models for death/discharge
joint_ests_ph_ischemic_death = joint_survival(indata.surv = my.data.ischemic, 
                                 indata.long = daily, 
                                 event_name='death',
                                 long_var = 'p_h', 
                                 log_long = FALSE, 
                                 censor_day = censor_day, 
                                 intercept.survival = FALSE)
joint_ests_ph_ischemic_discharge = joint_survival(indata.surv = my.data.ischemic, 
                                 indata.long = daily, 
                                 event_name='discharge',
                                 long_var = 'p_h', 
                                 log_long = FALSE, 
                                 censor_day = censor_day, 
                                 intercept.survival = FALSE)
# store estimates
ph_ests_ischemic = bind_rows(
  filter(joint_ests_ph_ischemic_death$ests, is.long==1),
  filter(joint_ests_ph_ischemic_discharge$ests, is.long==1), .id ='model') %>%
  mutate(
    stroke_type = 'ischemic',
    model = factor(model, levels=1:2, labels=c('Death','Discharge')))
```

The estimates use `r joint_ests_ph_ischemic$numbers$N_patients` patients with `r format(joint_ests_ph_ischemic$numbers$N_long, big.mark=',')` longitudinal observations. The time to stroke model used `r joint_ests_ph_ischemic$numbers$N_event` strokes.

A higher ph reduces the risk of stroke. Each 0.1 unit increase of ph multiplies the median time to stroke by `r mult$mean` (95% credible interval `r mult$lower` to `r mult$upper`).

##### Longitudinal plot of ph for stroke and non-stroke patients (ischemic)

```{r}
to_plot = joint_ests_ph_ischemic$graph +
  ylab('ph')+
  xlab('Day since ICU admission')
figure_nums(name="joint_ph","ph over time for ischemic stroke and non-ischemic-stroke patients.")
print(to_plot)
```


### Joint model for time to hemorrhagic stroke using PaO2

```{r, joint_model_pa_o2_h}
# pa_o2 is skewed so log-transform 
joint_ests_pa_o2_hemorrhagic = joint_survival(indata.surv=my.data.hemorrhagic, 
                                              indata.long=daily, 
                                 event_name='stroke',
                                              long_var='pa_o2', 
                                              log_long=TRUE, 
                                              censor_day=censor_day, 
                                              intercept.survival=FALSE)
# table of estimates
table_nums(name="joint_pao2_hemorrhagic","Estimates from the joint model of PaO2 over time and time to hemorrhagic stroke") # title
joint_ests_pa_o2_hemorrhagic$table
# for text below
mult = filter(joint_ests_pa_o2_hemorrhagic$ests, model=='Time to stroke', term=='pa_o2') 

## competing models for death/discharge
joint_ests_pa_o2_hemorrhagic_death = joint_survival(indata.surv = my.data.hemorrhagic, 
                                 indata.long = daily, 
                                 event_name='death',
                                 long_var = 'pa_o2', 
                                 log_long = TRUE, 
                                 censor_day = censor_day, 
                                 intercept.survival = FALSE)
joint_ests_pa_o2_hemorrhagic_discharge = joint_survival(indata.surv = my.data.hemorrhagic, 
                                 indata.long = daily, 
                                 event_name='discharge',
                                 long_var = 'pa_o2', 
                                 log_long = TRUE, 
                                 censor_day = censor_day, 
                                 intercept.survival = FALSE)
# store estimates
pa_o2_ests_hemorrhagic = bind_rows(
  filter(joint_ests_pa_o2_hemorrhagic_death$ests, is.long==1),
  filter(joint_ests_pa_o2_hemorrhagic_discharge$ests, is.long==1), .id ='model') %>%
  mutate(
    stroke_type = 'hemorrhagic',
    model = factor(model, levels=1:2, labels=c('Death','Discharge')))
```

The estimates use `r joint_ests_pa_o2_hemorrhagic$numbers$N_patients` patients with `r format(joint_ests_pa_o2_hemorrhagic$numbers$N_long, big.mark=',')` longitudinal observations. The survival model used `r joint_ests_pa_o2_hemorrhagic$numbers$N_event` strokes.

A higher PaO2 reduces the risk of stroke. A doubling of PaO2 multiplies the median time to stroke by `r mult$mean` (95% credible interval `r mult$lower` to `r mult$upper`).

### Joint model for time to ischemic stroke using PaO2

```{r, joint_model_pa_o2}
# pa_o2 is skewed so log-transform 
joint_ests_pa_o2_ischemic = joint_survival(indata.surv=my.data.ischemic, 
                                           indata.long=daily, 
                                 event_name='stroke',
                                           long_var='pa_o2', 
                                           log_long=TRUE, 
                                           censor_day=censor_day, 
                                           intercept.survival=FALSE)
# table of estimates
table_nums(name="joint_pao2_ischemic","Estimates from the joint model of PaO2 over time and time to ischemic stroke") # title
joint_ests_pa_o2_ischemic$table
# for text below
mult = filter(joint_ests_pa_o2_ischemic$ests, model=='Time to stroke', term=='pa_o2') 

## competing models for death/discharge
joint_ests_pa_o2_ischemic_death = joint_survival(indata.surv = my.data.ischemic, 
                                 indata.long = daily, 
                                 event_name='death',
                                 long_var = 'pa_o2', 
                                 log_long = TRUE, 
                                 censor_day = censor_day, 
                                 intercept.survival = FALSE)
joint_ests_pa_o2_ischemic_discharge = joint_survival(indata.surv = my.data.ischemic, 
                                 indata.long = daily, 
                                 event_name='discharge',
                                 long_var = 'pa_o2', 
                                 log_long = TRUE, 
                                 censor_day = censor_day, 
                                 intercept.survival = FALSE)
# store estimates
pa_o2_ests_ischemic = bind_rows(
  filter(joint_ests_pa_o2_ischemic_death$ests, is.long==1),
  filter(joint_ests_pa_o2_ischemic_discharge$ests, is.long==1), .id ='model') %>%
  mutate(
    stroke_type = 'ischemic',
    model = factor(model, levels=1:2, labels=c('Death','Discharge')))
```

The estimates use `r joint_ests_pa_o2_ischemic$numbers$N_patients` patients with `r format(joint_ests_pa_o2_ischemic$numbers$N_long, big.mark=',')` longitudinal observations. The survival model used `r joint_ests_pa_o2_ischemic$numbers$N_event` strokes.

A higher PaO2 reduces the risk of stroke. A doubling of PaO2 multiplies the median time to stroke by `r mult$mean` (95% credible interval `r mult$lower` to `r mult$upper`).


##### Longitudinal plot of PaO2 for stroke and non-stroke patients (ischemic)

```{r}
to_plot = joint_ests_pa_o2_ischemic$graph +
  ylab('PaO2 (log scale)')+
  xlab('Day since ICU admission')
figure_nums(name="joint_pa_o2","PaO2 over time for ischemic stroke and non-ischemic stroke patients.") 
print(to_plot)
```

### Joint model for time to hemorrhagic stroke using PaCO2

```{r, joint_model_pa_co2_h}
# pa_co2 is skewed so log-transform
joint_ests_pa_co2_hemorrhagic = joint_survival(indata.surv=my.data.hemorrhagic, 
                                               indata.long=daily, 
                                 event_name='stroke',
                                               long_var='pa_co2', 
                                               log_long=TRUE, 
                                               censor_day=censor_day, 
                                               intercept.survival=FALSE)
# table of estimates
table_nums(name="joint_pa_co2_hemorrhagic","Estimates from the joint model of PaCO2 over time and time to hemorrhagic stroke") # title
joint_ests_pa_co2_hemorrhagic$table
# for text below
mult = filter(joint_ests_pa_co2_hemorrhagic$ests, model=='Time to stroke', term=='pa_co2') 

## competing models for death/discharge
joint_ests_pa_co2_hemorrhagic_death = joint_survival(indata.surv = my.data.hemorrhagic, 
                                 indata.long = daily, 
                                 event_name='death',
                                 long_var = 'pa_co2', 
                                 log_long = TRUE, 
                                 censor_day = censor_day, 
                                 intercept.survival = FALSE)
joint_ests_pa_co2_hemorrhagic_discharge = joint_survival(indata.surv = my.data.hemorrhagic, 
                                 indata.long = daily, 
                                 event_name='discharge',
                                 long_var = 'pa_co2', 
                                 log_long = TRUE, 
                                 censor_day = censor_day, 
                                 intercept.survival = FALSE)
# store estimates
pa_co2_ests_hemorrhagic = bind_rows(
  filter(joint_ests_pa_co2_hemorrhagic_death$ests, is.long==1),
  filter(joint_ests_pa_co2_hemorrhagic_discharge$ests, is.long==1), .id ='model') %>%
  mutate(
    stroke_type = 'hemorrhagic',
    model = factor(model, levels=1:2, labels=c('Death','Discharge')))
```

The estimates use `r joint_ests_pa_co2_hemorrhagic$numbers$N_patients` patients with `r format(joint_ests_pa_co2_hemorrhagic$numbers$N_long, big.mark=',')` longitudinal observations. The survival model used `r joint_ests_pa_co2_hemorrhagic$numbers$N_event` strokes.

A higher PaCO2 reduces the risk of stroke. A doubling of PaCO2 multiplies the median time to stroke by `r mult$mean` (95% credible interval `r mult$lower` to `r mult$upper`).

### Joint model for time to ischemic stroke using PaCO2

```{r, joint_model_pa_co2_i}
# pa_co2 is skewed so log-transform
joint_ests_pa_co2_ischemic = joint_survival(indata.surv=my.data.ischemic, 
                                            indata.long=daily, 
                                 event_name='stroke',
                                            long_var='pa_co2', 
                                            log_long=TRUE, 
                                            censor_day=censor_day, 
                                            intercept.survival=FALSE)
# table of estimates
table_nums(name="joint_pa_co2_ischemic","Estimates from the joint model of PaCO2 over time and time to ischemic stroke") # title
joint_ests_pa_co2_ischemic$table
# for text below
mult = filter(joint_ests_pa_co2_ischemic$ests, model=='Time to stroke', term=='pa_co2') 

## competing models for death/discharge
joint_ests_pa_co2_ischemic_death = joint_survival(indata.surv = my.data.ischemic, 
                                 indata.long = daily, 
                                 event_name='death',
                                 long_var = 'pa_co2', 
                                 log_long = TRUE, 
                                 censor_day = censor_day, 
                                 intercept.survival = FALSE)
joint_ests_pa_co2_ischemic_discharge = joint_survival(indata.surv = my.data.ischemic, 
                                 indata.long = daily, 
                                 event_name='discharge',
                                 long_var = 'pa_co2', 
                                 log_long = TRUE, 
                                 censor_day = censor_day, 
                                 intercept.survival = FALSE)
# store estimates
pa_co2_ests_ischemic = bind_rows(
  filter(joint_ests_pa_co2_ischemic_death$ests, is.long==1),
  filter(joint_ests_pa_co2_ischemic_discharge$ests, is.long==1), .id ='model') %>%
  mutate(
    stroke_type = 'ischemic',
    model = factor(model, levels=1:2, labels=c('Death','Discharge')))
```

##### Longitudinal plot of PaCO2 for stroke and non-stroke patients (hemorrhagic)

```{r}
to_plot = joint_ests_pa_co2_hemorrhagic$graph +
  ylab('PaCO2 (log scale)')+
  xlab('Day since ICU admission')
figure_nums(name="joint_pa_co2","PaCO2 counts over time for hemorrhagic stroke and patients with a hemorrhagic stroke") 
print(to_plot)
```

### Joint model for time to hemorrhagic stroke using platelet count

```{r, joint_model_platelet}
# platelet_count is skewed so log-transform 
joint_ests_platelet_hemorrhagic = joint_survival(indata.surv=my.data.hemorrhagic, 
                                                 indata.long=daily, 
                                 event_name='stroke',
                                                 long_var='platelet_count', 
                                                 log_long=TRUE, 
                                                 censor_day=censor_day, 
                                                 intercept.survival = FALSE,
                                                 outliers = 2) # cut a few very odd low platelet counts
# table of estimates
table_nums(name="joint_platelet_hemorrhagic","Estimates from the joint model of Platelet counts over time and time to hemorrhagic stroke") # title
joint_ests_platelet_hemorrhagic$table
# for text below
mult = filter(joint_ests_platelet_hemorrhagic$ests, model=='Time to stroke', term=='platelet_count') 

## competing models for death/discharge
joint_ests_platelet_count_hemorrhagic_death = joint_survival(indata.surv = my.data.hemorrhagic, 
                                 indata.long = daily, 
                                 event_name='death',
                                 long_var = 'platelet_count', 
                                 log_long = TRUE, 
                                 censor_day = censor_day, 
                                 intercept.survival = FALSE, 
                                 outliers = 2)
joint_ests_platelet_count_hemorrhagic_discharge = joint_survival(indata.surv = my.data.hemorrhagic, 
                                 indata.long = daily, 
                                 event_name='discharge',
                                 long_var = 'platelet_count', 
                                 log_long = TRUE, 
                                 censor_day = censor_day, 
                                 intercept.survival = FALSE,
                                 outliers = 2)
# store estimates
platelet_count_ests_hemorrhagic = bind_rows(
  filter(joint_ests_platelet_count_hemorrhagic_death$ests, is.long==1),
  filter(joint_ests_platelet_count_hemorrhagic_discharge$ests, is.long==1), .id ='model') %>%
  mutate(
    stroke_type = 'hemorrhagic',
    model = factor(model, levels=1:2, labels=c('Death','Discharge')))
```

The estimates use `r joint_ests_platelet_hemorrhagic$numbers$N_patients` patients with `r format(joint_ests_platelet_hemorrhagic$numbers$N_long, big.mark=',')` longitudinal observations. The survival model used `r joint_ests_platelet_hemorrhagic$numbers$N_event` strokes.

Platelet count had no clear effect on the risk of stroke. A doubling of platelet count multiplies the median time to stroke by `r mult$mean` (95% credible interval `r mult$lower` to `r mult$upper`). As the 95% CI includes 1 the effect of platelet count is not certain.

### Joint model for time to ischemic stroke using platelet count

```{r, joint_model_platelet_i}
# platelet_count is skewed so log-transform 
joint_ests_platelet_ischemic = joint_survival(indata.surv=my.data.ischemic, 
                                              indata.long=daily, 
                                 event_name='stroke',
                                              long_var='platelet_count', 
                                              log_long=TRUE, 
                                              censor_day=censor_day, 
                                              intercept.survival = FALSE,
                                              outliers = 2)
# table of estimates
table_nums(name="joint_platelet_ischemic","Estimates from the joint model of Platelet counts over time and time to ischemic stroke") # title
joint_ests_platelet_ischemic$table
# for text below
mult = filter(joint_ests_platelet_ischemic$ests, model=='Time to stroke', term=='platelet_count') 

## competing models for death/discharge
joint_ests_platelet_count_ischemic_death = joint_survival(indata.surv = my.data.ischemic, 
                                 indata.long = daily, 
                                 event_name='death',
                                 long_var = 'platelet_count', 
                                 log_long = TRUE, 
                                 censor_day = censor_day, 
                                 intercept.survival = FALSE,
                                 outliers = 2)
joint_ests_platelet_count_ischemic_discharge = joint_survival(indata.surv = my.data.ischemic, 
                                 indata.long = daily, 
                                 event_name='discharge',
                                 long_var = 'platelet_count', 
                                 log_long = TRUE, 
                                 censor_day = censor_day, 
                                 intercept.survival = FALSE,
                                 outliers = 2)
# store estimates
platelet_count_ests_ischemic = bind_rows(
  filter(joint_ests_platelet_count_ischemic_death$ests, is.long==1),
  filter(joint_ests_platelet_count_ischemic_discharge$ests, is.long==1), .id ='model') %>%
  mutate(
    stroke_type = 'ischemic',
    model = factor(model, levels=1:2, labels=c('Death','Discharge')))

```

##### Longitudinal plot of platelet count for hemorrhagic stroke and patients without a hemorrhagic stroke

```{r}
to_plot = joint_ests_platelet_hemorrhagic$graph +
  ylab('Platelet count (log scale)')+
  xlab('Day since ICU admission')
figure_nums(name="joint_platelet","Platelet counts over time for stroke and non-stroke patients.") 
print(to_plot)
```


### Joint models summary plot

The plot below shows the estimates for the joint models for the longitudinal association for the time to stroke.

```{r, summary_plot}
# combine estimates
comb = bind_rows(joint_ests_pa_o2_hemorrhagic$ests,
                 joint_ests_pa_o2_ischemic$ests,
                 joint_ests_pa_co2_hemorrhagic$ests,
                 joint_ests_pa_co2_ischemic$ests,
                 joint_ests_aptt_hemorrhagic$ests,
                 joint_ests_aptt_ischemic$ests,
                 joint_ests_platelet_hemorrhagic$ests,
                 joint_ests_platelet_ischemic$ests,
                 joint_ests_ph_hemorrhagic$ests,
                 joint_ests_ph_ischemic$ests, .id='hi') %>%
  filter(is.long==1) %>% # just the longitudinal
  mutate(
  group_nice = as.numeric(as.numeric(hi)%%2 == 0), # even numbers
  group_nice = factor(group_nice, levels=0:1, labels=c('Hemorrhagic','Ischemic')),
  xaxis = case_when(
    term =='aptt' ~ 1,
    term =='p_h' ~ 2,
    term =='platelet_count' ~ 3,
    term =='pa_o2' ~ 4,
    term =='pa_co2' ~ 5
  ),
  xaxis = xaxis + (as.numeric(hi)%%2/10)) #dodge
#
xaxis_labels = c('APTT (double)','pH (per 0.1)','Platelet count (double)','PaO2 (double)','PaCO2 (double)')
# annotation
labels = data.frame(xaxis=0.9, mean=1, lower=1, upper=1, label=c('Increased risk','Decreased risk'))
# 
gplot = ggplot(data=comb, aes(x=xaxis, y=mean, ymin=lower, ymax=upper, col=factor(group_nice)))+
  geom_hline(lty=2, yintercept=1)+ # reference line
  scale_x_continuous(breaks=1:5, labels=xaxis_labels)+
  geom_text(data=labels, size=3, aes(x=xaxis, y=mean, label=label), col='black', hjust=c(-0.1, 1.1))+
  geom_point(size=2)+
  geom_errorbar(width=0, size=1.05)+
  scale_color_manual(NULL, values=c('darkred','seagreen3'))+
  theme_bw()+
  xlab('')+
  ylab('Multiplier')+
  coord_flip()+
  theme(legend.position = 'top')
figure_nums(name="joint_summary","Summary plot of the joint model estimates on the time to stroke")
gplot
```


### Within-patient correlation for longitudinal variables

Some of the effect sizes about are remarkably similar, this could be because they are strongly correlated within patients. Here we examine the within patient correlation in variables. We log transformed (base 2) all the variables except pH.

```{r}
for_corr = select(daily, aptt, platelet_count, pa_co2, pa_o2, p_h) %>%
  mutate(aptt = log2(aptt),
         pa_o2 = log2(pa_o2),
         pa_co2 = log2(pa_co2),
         platelet_count = log2(platelet_count)) %>%
  as.matrix()
c = as.data.frame(cor(for_corr, use='pairwise.complete.obs')) %>%
  mutate_all(funs(roundz(., 2)))
tab = mutate(c, Variable = colnames(c)) %>%
  select(Variable, everything())
ftab = flextable(tab) %>%
  theme_box
ftab
```


### Joint models summary plot for competing outcomes of death and discharge

The plot below shows the estimates for the joint models for the effect of the longitudinal variable on death and discharge.

```{r, summary_plot_competing, fig.width=7, fig.height=6}
# combine estimates
comb = bind_rows(aptt_ests_ischemic,
                 aptt_ests_hemorrhagic,
                 pa_o2_ests_ischemic,
                 pa_o2_ests_hemorrhagic,
                 pa_co2_ests_ischemic,
                 pa_co2_ests_hemorrhagic,
                 platelet_count_ests_ischemic,
                 platelet_count_ests_hemorrhagic,
                 ph_ests_ischemic,
                 ph_ests_hemorrhagic) %>%
  mutate(
  xaxis = case_when(
    term =='aptt' ~ 1,
    term =='p_h' ~ 2,
    term =='platelet_count' ~ 3,
    term =='pa_o2' ~ 4,
    term =='pa_co2' ~ 5
  ),
  xaxis = xaxis + (as.numeric(stroke_type=='ischemic')/8)) #dodge
#
xaxis_labels = c('APTT (double)','pH (per 0.1)','Platelet count (double)','PaO2 (double)','PaCO2 (double)')
# annotation
labels = data.frame(xaxis=0.9, mean=1, lower=0, upper=0, stroke_type=NA, 
                    model = factor(c(1,1,2,2), levels=c(1,2), labels=c('Death','Discharge')),
                    label=c('Increased risk','Decreased risk','Increased risk','Decreased risk'))
# 
gplot = ggplot(data=comb, aes(x=xaxis, y=mean, ymin=lower, ymax=upper, col=factor(stroke_type)))+
  geom_hline(lty=2, yintercept=1)+ # reference line
  geom_text(data=labels, size=3, aes(x=xaxis, y=mean, label=label), col='black', hjust=c(-0.1, 1.1, -0.1, 1.1))+
  scale_x_continuous(breaks=1:5, labels=xaxis_labels)+
  scale_y_log10()+
  geom_point(size=2)+
  geom_errorbar(width=0, size=1.05)+
  scale_color_manual(NULL, values=c('darkred','seagreen3'))+
  theme_bw()+
  xlab('')+
  ylab('Multiplier')+
  coord_flip()+
  facet_wrap(~model)+
  theme(legend.position = 'top')
figure_nums(name="joint_summary_competing","Summary plot of the joint model estimates on the time to death and discharge")
gplot
```

The results for hemorrhagic and ischemic stroke are almost identical for each combination of longitudinal variable and outcome (death/discharge). This is not surprising given the relatively small number of strokes, which means that the risks of death and discharge are relatively unchanged for the two stroke types.


# Appendices

These appendices contain additional results and checks of the models. These could be appendices to the paper.

##### Site effects (death and discharge)

```{r, fig.width=5, fig.height=6}
# get centre estimates
centres_death = bugs_extract(inbugs=bugs_results_death, var='^centre', exp=TRUE) %>% # from 99_functions.R
  mutate(type = 1)
# annotation
centres_discharge = bugs_extract(inbugs=bugs_results_discharge, var='^centre', exp=TRUE) %>% # from 99_functions.R
  mutate(type = 2)
# annotation
labels = data.frame(row=0, mean=1, lower=1, upper=1, facet=1, label=c('Increased hazard','Decreased hazard'))
# combine estimates
centres = bind_rows(centres_death, centres_discharge) %>%
  group_by(row) %>%
  mutate(mean_hr = mean(mean)) %>% # calculate average HR
  ungroup() %>%
  mutate(order = as.numeric(as.factor(mean_hr)),
         order = ifelse(type==2, order+0.2, order), # jitter
         facet = ifelse(order >= 60, 2, 1),
         order = ifelse(facet==2, order-60, order))
# plot
cplot = ggplot(data=centres, aes(x=order, y=mean, ymin=lower, ymax=upper, col=factor(type)))+
  geom_hline(lty=2, yintercept=1)+
  geom_point()+
  scale_y_log10()+
  scale_x_continuous(breaks=NULL, labels = NULL, limits=c(0, 61+0.2), expand=c(0.01,0.01))+
  scale_color_manual(NULL, values=c('indianred','skyblue'), labels=c('Death','Discharge'))+
  geom_errorbar(width=0)+
  geom_text(data=labels, size=3, aes(x=row, y=mean, label=label), col='black', hjust=c(-0.1, 1.1))+
  xlab('')+
  ylab('Hazard ratio')+
  coord_flip()+
  theme_bw()+
  theme(legend.position='top',
        panel.grid.minor.y = element_blank(), 
        panel.grid.major.y = element_blank(),
        strip.background = element_blank(), # remove facet labels
        strip.text.x = element_blank())+
  facet_wrap(~facet)
figure_nums(name="site_death", "Site effects from the Weibull survival model for death and discharge", level=1)
cplot
```

The hazard ratio axis is on log scale (base 10). Each row is a site and there are two estimates per site: one for death and one for discharge. The sites are ordered from lowest to highest risk. The plot has been split into two panels because of the number of sites. There are a few sites where the hazards of death or discharge are quite different from other sites.

### Alternative Cox survival models for death and discharge

Here we use non-parametric Cox models to confirm the results from the parametric Weibull survival models. These models use the multi-state approach. These models account for correlated results from the same site. We use a simple mean imputation for patients missing BMI.

The table shows the mean hazard ratios (HRs) and 95% confidence intervals for death and discharge.

```{r}
# use function to run Cox models (also used below in leave-one-out analysis)
run_cox_models = function(in_discharge, in_death, digits=2){
  # convert lists to data frames
  if(class(in_discharge) == 'list'){in_discharge = as.data.frame(in_discharge)}
  if(class(in_death) == 'list'){in_death = as.data.frame(in_death)}
  # a) discharge
  in_discharge = mutate(in_discharge,
          bmi = ifelse(is.na(bmi)==TRUE, 0, bmi)) # simple imputation of mean
  # cox
  in_discharge = mutate(in_discharge,
    days = ifelse(is.na(days)==TRUE, cens, days)) # get days from censored or uncensored
  c_disch = coxph(Surv(days, event=cens==0, type='right') ~ sex + age + bmi + stroke_o + stroke_i + stroke_h + cluster(site), data=in_discharge)
  ests_disch = tidy(c_disch, conf.int = TRUE, exponentiate = TRUE) %>%
    mutate(
      estimate = roundz(estimate, digits),
      conf.low = roundz(conf.low, digits),
      conf.high = roundz(conf.high, digits),
      CI = paste(conf.low, ' to ', conf.high, sep='')) %>%
    select(-term, -std.error, -statistic, -p.value) %>%
    rename('HR2' = 'estimate',
           'CI2' = 'CI',
           'conf.low2' = 'conf.low',
           'conf.high2' = 'conf.high')
  # b) death
  in_death = mutate(in_death,
          bmi = ifelse(is.na(bmi)==TRUE, 0, bmi)) # simple imputation of mean
  # cox
  in_death = mutate(in_death,
            days = ifelse(is.na(days)==TRUE, cens, days)) # get days from censored or uncensored
  c_death = coxph(Surv(days, event=cens == 0, type='right') ~ sex + age + bmi + stroke_o + stroke_i + stroke_h + cluster(site), data=in_death)
  ests_death = tidy(c_death, conf.int = TRUE, exponentiate = TRUE) %>%
    mutate(
      estimate = roundz(estimate, digits),
      conf.low = roundz(conf.low, digits),
      conf.high = roundz(conf.high, digits),
      CI = paste(conf.low, ' to ', conf.high, sep = '')) %>%
    select(-std.error, -statistic, -p.value) %>%
    rename('HR1' = 'estimate',
           'CI1' = 'CI',
           'conf.low1' = 'conf.low',
           'conf.high1' = 'conf.high')
  for_tab = bind_cols(ests_death, ests_disch) %>%
    mutate(term = str_remove(string=term, pattern='in_death\\$'))
 # return
 to.return = list()
 to.return$table = for_tab
 to.return$discharge_model = c_disch
 to.return$death_model = c_death
 to.return$death_data = in_death # add data
 to.return$discharge_data = in_discharge # 
 return(to.return) 
}
cox_models = run_cox_models(in_discharge = for_discharge, in_death = for_death) # run above function
for_tab = cox_models$table %>% 
  select(term, HR1, CI1, HR2, CI2) %>%
  mutate(term = ifelse(term=='age', 'Age (+10 years)', term),
         term = ifelse(term=='bmi', 'BMI (+5 kg/m2)', term),
         term = ifelse(term=='sex', 'Male', term),
         term = ifelse(term=='sofa_cHigh', 'SOFA = high', term),
         term = ifelse(term=='sofa_cMissing', 'SOFA = missing', term),
         term = ifelse(term=='stroke_o', 'Stroke - not specified', term),
         term = ifelse(term=='stroke_i', 'Stroke - Ischemic', term),
         term = ifelse(term=='stroke_h', 'Stroke - Hemorrhage', term))
## nice table
  header = read.table(header=TRUE, stringsAsFactors = FALSE, sep="!", text='
col_keys!colB!colA
term! !Variable
HR1!Death!HR
CI1!Death!95% CI
HR2!Discharge!HR
CI2!Discharge!95% CI
') 
ftab = flextable(for_tab) %>%
  set_header_df(mapping = header, key = "col_keys") %>%
  merge_h(part = "header", i = 1) %>%
  theme_box() %>%# nice theme
  align(align='center', part='all') %>%# centre text
  autofit()
table_nums(name="hr_cox", "Hazard ratios and 95% credible intervals for death and discharge from the Cox model", level=1)
ftab 
```



#### Check of proportional hazards assumption for Cox models

```{r}
prop_disch = cox.zph(run_cox_models(in_discharge = for_discharge, in_death = for_death)$discharge_model)
prop_death = cox.zph(run_cox_models(in_discharge = for_death, in_death = for_death)$death_model)
for_table = bind_rows(as.data.frame(prop_disch$table), as.data.frame(prop_death$table), .id='outcome') %>%
  mutate(Variable = row.names(.),
         Variable = str_remove_all(string=Variable, pattern='in_discharge\\$|\\.|[0-9]'), # tidy names
         chisq = round(chisq*100, 0)/100,
         p = format.pval(p, eps=0.001, digits=3),
         outcome = factor(as.numeric(outcome), levels=1:2, labels=c("Discharge",'Death'))) %>%
  select(outcome, Variable, everything())
ftab = flextable(for_table) %>%
  merge_v(j=1) %>% # 
  theme_box() %>%
  autofit()
table_nums(name="ph_check", "Check of proportional hazards assumption for Cox models")
ftab
```

There are strong non-proportional hazards for age in the discharge model and hemorrhagic stroke in the death model.

The plot below shows the estimated effect of age over time on discharge. There is a stronger protective effect of age in the first few weeks (negative beta), with a weaker protective effect thereafter. We can account for this by splitting the effect of age into and early and late effect (making it time-dependent). 

```{r}
figure_nums(name="cox_zph_age", "Plot of the estimated effect of age on discharge over time in ICU. The green line is the estimated effect with dotted lines for a 95% confidence interval. The horizontal red line indicates no effect of age.")
par(mai=c(1,1,0.1,0.1))
plot(prop_disch[2], xlab='Days since ICU admission', se=TRUE, col=3, resid=FALSE)
abline(0, 0, col=2)
```

The plot below shows the estimated effect of age over time on death. There is a large change in the effect of age, which grows over time from an effect close to zero on admission to ICU. The effect of age stabilises after around 9 days. After this time age consistently increases the risk of death.

```{r}
figure_nums(name="cox_zph_age_death", "Plot of the estimated effect of age on death over time in ICU. The green line is the estimated effect with dotted lines for a 95% confidence interval. The horizontal red line indicates no effect of age.")
par(mai=c(1,1,0.1,0.1))
plot(prop_death[2], xlab='Days since ICU admission', se=TRUE, col=3, resid=FALSE)
abline(0, 0, col=2)
```

The plots for the non-proportional effects of hemorrhagic stroke (not shown) indicate influential observations which we address below.

#### Discharge model with time-dependent effect of age

Here we fit a Cox survival model for discharge with a separate effect of age before and after 20 days in ICU. 

```{r}
# see: vignette("timedep", package = "survival")
for_split = select(my.data, -stroke_type) %>%
  mutate(sex = as.numeric(sex=='Male'),
             age = (age-50)/10, # scaled
             bmi = (bmi-30)/5)
# add split to data at 20 days
for_split_split <- survSplit(Surv(time=entry, time2=exit, event=to=='discharge') ~ ., data= for_split, cut=c(20), episode= "tgroup", id="idx")
split_m = coxph(Surv(time=entry, time2=exit, event) ~ sex + age:strata(tgroup) + bmi + factor(from), data=for_split_split)
ests_disch_split = tidy(split_m, conf.int = TRUE, exponentiate = TRUE) %>%
    mutate(
      estimate = roundz(estimate, digits),
      conf.low = roundz(conf.low, digits),
      conf.high = roundz(conf.high, digits),
      CI = paste(conf.low, ' to ', conf.high, sep=''),
      term = ifelse(str_detect(term, 'age') &  str_detect(term, 'group=1'), 'Age (+10 years) before 20 days in ICU', term),
      term = ifelse(str_detect(term, 'age') &  str_detect(term, 'group=2'), 'Age (+10 years) after 20 days in ICU', term),
      term = ifelse(str_detect(term, 'from\\(stroke\\)'), 'Stroke, all 3 types', term),
      term = ifelse(str_detect(term, 'ecmo_start'), 'During ECMO', term),
      term = ifelse(str_detect(term, 'ecmo_end'), 'Post-ECMO', term),
      term = ifelse(str_detect(term, 'from') &  str_detect(term, '2'), 'Stroke, not specified', term),
      term = ifelse(str_detect(term, 'from') &  str_detect(term, '3'), 'Stroke, ischemic', term),
      term = ifelse(str_detect(term, 'from') &  str_detect(term, '4'), 'Stroke, hemorrhage', term),
      term = ifelse(term=='bmi', 'BMI (+5 kg/m2)', term),
      term = ifelse(term=='sex', 'Male', term)) %>%
    select(term, estimate, CI) %>%
    rename('HR' = 'estimate',
           'CI' = 'CI')
ftab = flextable(ests_disch_split) %>%
  theme_box()%>%
  autofit()
ftab
```

The stronger protective effect of age at early times is shown by the hazard ratio being smaller in the first 20 days compared with later days.

#### Death model with time-dependent effect of age

Here we fit a Cox survival model for death with a separate effect of age before and after 9 days in ICU. 

```{r}
# see: vignette("timedep", package = "survival")
for_split = select(my.data, -stroke_type) %>%
  mutate(sex = as.numeric(sex=='Male'),
             age = (age-50)/10, # scaled
             bmi = (bmi-30)/5)
# add split to data at 20 days
for_split_split <- survSplit(Surv(time=entry, time2=exit, event=to=='death') ~ ., data= for_split, cut=c(9), episode= "tgroup", id="idx")
split_m = coxph(Surv(time=entry, time2=exit, event) ~ sex + age:strata(tgroup) + bmi + factor(from), data=for_split_split)
ests_disch_split = tidy(split_m, conf.int = TRUE, exponentiate = TRUE) %>%
    mutate(
      estimate = roundz(estimate, digits),
      conf.low = roundz(conf.low, digits),
      conf.high = roundz(conf.high, digits),
      CI = paste(conf.low, ' to ', conf.high, sep=''),
      term = ifelse(str_detect(term, 'age') &  str_detect(term, 'group=1'), 'Age (+10 years) before 9 days in ICU', term),
      term = ifelse(str_detect(term, 'age') &  str_detect(term, 'group=2'), 'Age (+10 years) after 9 days in ICU', term),
      term = ifelse(str_detect(term, 'from') &  str_detect(term, '2'), 'Stroke, not specified', term),
      term = ifelse(str_detect(term, 'from') &  str_detect(term, '3'), 'Stroke, ischemic', term),
      term = ifelse(str_detect(term, 'from') &  str_detect(term, '4'), 'Stroke, hemorrhage', term),
      term = ifelse(str_detect(term, 'ecmo_start'), 'During ECMO', term),
      term = ifelse(str_detect(term, 'ecmo_end'), 'Post-ECMO', term),
      term = ifelse(term=='bmi', 'BMI (+5 kg/m2)', term),
      term = ifelse(term=='sex', 'Male', term)) %>%
    select(term, estimate, CI) %>%
    rename('HR' = 'estimate',
           'CI' = 'CI')
ftab = flextable(ests_disch_split) %>%
  theme_box()%>%
  autofit()
ftab
```

The stronger effect of age at later times is clear in the increasing hazard ratio for death.

Accounting for the non-proportional effect of age had little effect on the estimates for stroke, and hemorrhagic stroke was still strongly associated with an increased risk of death.

#### Checking for influential patients

In this section we check whether there are influential patients in the Cox survival analysis. An influential patient is one who has a large effect of the estimates and hence any inference.

The plot below shows the 'dfbetas' which are the estimated changes in the regression coefficients after deleting each observation patient in turn. The estimates are on the log-scale and zero means no change. We have separate plots for the discharge and death models.

```{r}
figure_nums(name="inf_discharge", "Influential observations for the Cox survival model of discharge", level=2) # title
# test of influential observations
influence_discharge = ggcoxdiagnostics(fit=cox_models$discharge_model, sline=FALSE, type = 'dfbeta', linear.predictions = FALSE, ggtheme = theme_bw())
influence_discharge
```

The most influential patients are shown below and are the few patients who had a stroke.

```{r}
# look at influential observation(s)
index = filter(influence_discharge$data, covariate=='stroke_h', abs(res) >= 0.1) %>% pull(xval)
to_table = cox_models$discharge_data[index,] %>%
  select(site, sex, age, bmi, stroke_h, days, cens) %>%
  mutate(
    cens = ifelse(cens==0, 'No', 'Yes'), # nicer labels for table
    sex = ifelse(sex==1, 'Male', 'Female'),
    days = round(days),
    age = (age*10)+50, # back-transform
    bmi = (bmi*5)+30,
    bmi = round(bmi*10)/10) %>%
  rename('censored' = 'cens')
ftab = flextable(to_table) %>%
  autofit() %>%
  theme_box()
ftab
```

```{r}
figure_nums(name="inf_death", "Influential observations for the Cox survival model of death", level=2) # title
# test of influential observations
influence_death = ggcoxdiagnostics(fit=cox_models$death_model, sline=FALSE, type = 'dfbeta', linear.predictions = FALSE, ggtheme = theme_bw())
influence_death
```

The most influential patients are shown below and are the few patients who had a stroke.

```{r}
# look at influential observation(s)
index = filter(influence_death$data, covariate=='stroke_h', abs(res) >= 0.1) %>% pull(xval)
to_table = cox_models$discharge_data[index,] %>%
  select(site, sex, age, bmi, stroke_h, days, cens) %>%
  mutate(
    cens = ifelse(cens==0, 'No', 'Yes'), # nicer labels for table
    sex = ifelse(sex==1, 'Male', 'Female'),
    days = round(days),
    age = (age*10)+50, # back-transform
    bmi = (bmi*5)+30,
    bmi = round(bmi*10)/10) %>%
  rename('censored' = 'cens')
ftab = flextable(to_table) %>%
  autofit() %>%
  theme_box()
ftab
```

## Convergence of Bayesian models

Here we visually check the convergence of the estimates for the Bayesian survival models for death and discharge. We plot the intercept (int), regression parameters (beta's), and Weibull shape parameter (r).

```{r, fig.width=7}
# plot r, beta and intercept
to_plot = c('^r$','^beta','^int')
index = str_detect(pattern=paste(to_plot, collapse='|'), string=dimnames(bugs_results_death$sims.array)[[3]])
chain1 = reshape2::melt(bugs_results_death$sims.array[,1,index])
chain2 = reshape2::melt(bugs_results_death$sims.array[,2,index])
chains = bind_rows(chain1, chain2, .id='chain')
chain_plot = ggplot(data=chains, aes(x=Var1, y=value, col=factor(chain)))+
  geom_line()+
  scale_color_manual('Chain', values=c(c("#FFB90F", "#00B2EE")))+
  xlab('')+
  ylab('')+
  theme_bw()+
  facet_wrap(~Var2, scales='free_y')+
  theme(legend.position = c(0.8,0.1))
figure_nums(name="chains_death", "Markov Chain Monte Carlo estimates for the Weibull survival model of death", level=2) # title
chain_plot
```

```{r, fig.width=7}
# plot r, beta and intercept
to_plot = c('^r$','^beta','^int')
index = str_detect(pattern=paste(to_plot, collapse='|'), string=dimnames(bugs_results_discharge$sims.array)[[3]])
chain1 = reshape2::melt(bugs_results_discharge$sims.array[,1,index])
chain2 = reshape2::melt(bugs_results_discharge$sims.array[,2,index])
chains = bind_rows(chain1, chain2, .id='chain')
chain_plot = ggplot(data=chains, aes(x=Var1, y=value, col=factor(chain)))+
  geom_line()+
  scale_color_manual('Chain', values=c(c("#FFB90F", "#00B2EE")))+
  xlab('')+
  ylab('')+
  theme_bw()+
  facet_wrap(~Var2, scales='free_y')+
  theme(legend.position = c(0.8,0.1))
figure_nums(name="chains_discharge", "Markov Chain Monte Carlo estimates for the Weibull survival model of discharge") # title
chain_plot
```

#### R version and packages used

This information can be useful for reproducibility purposes.

```{r}
sessionInfo()
```
