==> 1_ecmo_prone_multistate.Rmd <==
---
title: "Analysis of ECMO and prone positioning using a multi-state survival model"
author: "Adrian Barnett"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  word_document:
  reference_docx: rmarkdown-styles-SAP-landscape.docx
link-citations: yes
bibliography: bibs.bib
---
  
```{r, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, warning=FALSE, message=FALSE, error=FALSE, comment='', dpi=400)
options(width=1000) # Wide pages
options(scipen=999) # avoid scientific presentation
debug = FALSE # for winbugs

## libraries and functions
library(captioner)
table_nums <- captioner(prefix = "Table", levels=2, type = c("n", "c")) # number and character
figure_nums <- captioner(prefix = "Figure", levels=2, type = c("n", "c"))
library(diagram)
library(dplyr)
options(dplyr.summarise.inform = FALSE)
library(mstate)
library(stringr)
library(flextable)
library(lme4)
library(R2WinBUGS)
bugs_location = "c:/Program Files/WinBUGS14" # location of the file WinBUGS14.exe
library(broom) 
library(cmprsk)
library(survminer) # for nice plots of competing risks; also test of influential values
# function for preparing data set for mstate package
source('99_functions.R')
source("Martin/ext_mstate.R")
# function for bootstrap confidence intervals
source("Martin/LOS_boot.R")
source('99_functions.R')
# plots
library(ggplot2)
library(gridExtra)

# maximum day for plots and estimates
censor_day = 90

# Data preparation-------------------------------------------------------------

# State 0: Censored
# State 1: ECMO & supine
# State 2: Prone
# State 3: Discharge
# State 4: Death

## prepare CCC set (with id, from, to, entry, exit; from and to are state numbers)
#load('data/COVID_Data_stroke.RData') # 0_read_data.R
load('data/COVID_Data_stroke_Oct_31.RData') # database freez
#load('data/COVID_Data_stroke_dummy.RData') # 0_read_data.R - dummy data for testing code
## apply inclusions
patients = filter(patients,
                  #                  diagnosis_coronavirus ==1, # with diagnosis
                  #pin != 'OX_007100057', # without influential patient
                  age >= 18, # adults
                  !is.na(date_ecmo)) # must have started ECMO
## apply exclusions and count excluded
ex1 = filter(patients,
             ecmo_type == 'Venous-Venous') # must be V-V ECMO
n_exclude_ecmo_type = nrow(patients) - nrow(ex1)
# patient must be in daily data 
u = c(unique(daily$pin)) # 
patients = filter(ex1, pin %in% u)
n_exclude_no_daily = nrow(ex1) - nrow(patients)
# final edits
patients = mutate(patients, 
                  # any prone during ECMO:
                  any_prone = as.numeric(!is.na(prone_start) & (prone_start>=date_ecmo | prone_end > date_ecmo)), # start date not missing and after ECMO -> 1
                  any_prone = ifelse(prone_end == date_ecmo, 0, any_prone), # assume supine if ends on same day as ECMO start
                  any_prone = ifelse(is.na(prone_start), 0, any_prone), # no if prone start is missing
                  any_prone = factor(any_prone, levels=0:1, labels=c('No','Yes')),
                  # if no final outcome then unknown:
                  outcome = ifelse(outcome=='', 'Unknown', outcome),
                  # if no death date then switch to unknown:
                  outcome = ifelse(outcome=='Death' & is.na(date_death), 'Unknown', outcome),
                  # make pre-ECMO ventilation
                  pre_ecmo_vent = ifelse(is.na(date_mechanical_ventilation) | date_mechanical_ventilation >= date_ecmo, 0, 1),
                  pre_ecmo_vent = factor(pre_ecmo_vent, levels=0:1, labels=c('No','Yes')),
                  # labels
                  ecmo_prone_before = factor(ecmo_prone_before, levels=0:1, labels=c('No','Yes')),
                  # duration of pre-ECMO ventilation
                  dur_pre_ecmo_vent = as.numeric(date_ecmo - date_mechanical_ventilation) + 0.5,
                  dur_pre_ecmo_vent = ifelse(is.na(dur_pre_ecmo_vent) | dur_pre_ecmo_vent < 0, 0, dur_pre_ecmo_vent)
) 
# estimates for text
n_patients = nrow(patients)
n_sites = length(unique(patients$site_name))
ecmo_start = min(patients$date_ecmo)
ecmo_end = max(patients$date_ecmo)

## daily data on date of ECMO ##
for_join = select(patients, pin, date_ecmo, any_prone) %>%
  #  mutate(date_ecmo = date_ecmo + 1) %>% # use first full date on ECMO? ...
  rename('date_daily' = 'date_ecmo') # ... currently matched on date of ECMO
daily_ecmo = left_join(for_join, daily, by=c('pin','date_daily'))

## for checking
# check = select(patients, pin, date_icu, date_death, date_hospital_discharge, last_date, date_outcome, outcome) %>% filter(!is.na(date_icu), is.na(date_death), is.na(date_hospital_discharge), is.na(date_outcome))
```


```{r, include=FALSE}
### prepare data for survival analysis ###
## ---- transitions
my.data = time_transitions(indata=patients, 
                           start_state='date_ecmo',
                           int_state = 'prone_start',
                           int_state_end = 'prone_end', 
                           discharge_date = 'date_hospital_discharge',
                           censor_day = censor_day, check=TRUE) %>% # from 99_functions.R
  select(pin, from, to, start, end, date_ecmo, age, any_prone, sex, bmi, site_name, country, ecmo_prone_before) %>%
  rename('entry' = 'start',
         'exit' = 'end')

# alternative function to verify, dummy function, do not run
alternative = function(){
  library(tidyr)
my.data.alt = time_transitions_string(indata=patients,
                intermediate_prior_ongoing= TRUE,
                censor_day = censor_day,
                start_state = 'date_ecmo',
                int_state = 'prone_start',
                int_state_end = 'prone_end',
                discharge_date = 'date_hospital_discharge',
                dates= c('date_ecmo', 'prone_start', 'prone_end', 'date_hospital_discharge', 'date_transfer','date_death'))
## remove some transitions
# no outcomes after transfer
index = my.data.alt$from == 'date_transfer' 
my.data.alt = my.data.alt[!index,]
# 
index = my.data.alt$from == 'date_ecmo' & my.data.alt$to=='date_int_state_start'& my.data.alt$entry==0&my.data.alt$exit==0.5
my.data.alt = my.data.alt[!index,]
#
index = my.data.alt$from == 'date_int_state_start' & my.data.alt$to=='date_int_state_end'& my.data.alt$entry==0&my.data.alt$exit==0.5
my.data.alt = my.data.alt[!index,]
# rename states
my.data.alt = rename(my.data.alt,
                     'from_c' = 'from',
                     'to_c' = 'to') %>%
  mutate(to = case_when(
    to_c == 'censored' ~ 0, # censored
    to_c == 'date_transfer' ~ 0, # censored
    to_c == 'last_date' ~ 0, # censored
    to_c == 'date_int_state_end' ~ 1,
    to_c == 'date_int_state_start' ~ 2,
    to_c == 'discharge_date' ~ 3,
    to_c == 'discharge' ~ 3,
    to_c == 'date_death' ~ 4,
    to_c == 'death' ~ 4
  ),
  from = case_when(
    from_c == 'date_ecmo' & to_c=='date_int_state_end' ~ 2,
    from_c == 'date_ecmo' & to_c!='date_int_state_end' ~ 1,
    from_c == 'date_int_state_start' ~ 2,
    from_c == 'date_int_state_end' ~ 1,
  ))
# compare and check
with(my.data.alt, table(from_c, to_c))
with(my.data.alt, table(from, to))
#with(my.data, table(from, to))
#f1 = filter(my.data.alt, from==2, to==4) %>% select(pin, from_c, to_c, entry, exit)
#f2 = filter(my.data, from==2, to==4) %>% select(id, from, to, entry, exit)
#f1$pin[f1$pin %in% f2$id == FALSE]
#f2$id[f2$id%in%f1$pin  == FALSE]
#f3 = filter(patients, pin=='OX_00549-0048') %>% select(pin, 'date_ecmo', 'prone_start', 'prone_end', 'date_hospital_discharge', 'date_death')
#f4 = filter(my.data.alt, pin=='OX_00549-0048') %>% select(pin, from_c, to_c, from, to, entry, exit)
#f5 = filter(my.data, id=='OX_00549-0048') %>% select(id, from, to, entry, exit)
} # end of if

# Set transition matrix for mstate
tra <- transMat(x = list(c(2,3,4), c(1,3,4), c(),c()),
                names = c("ECMO", "Prone", "Discharge","Death"))
# version with no transitions from prone to discharge
#tra <- transMat(x = list(c(2,3,4), c(1,4), c(),c()),
#                names = c("ECMO", "Prone", "Discharge","Death"))

# Add transition vectors
my.data$trans <- NA
for (i in 1:nrow(tra))
{
  for (j in 1:ncol(tra))
  {
    my.data$trans[which(my.data$from == i & my.data$to ==j)] <- tra[i, j]  
  }
}

# Add status vector, indicates observed transition
my.data$status <- 1
# Status vector for censored observations set to '0'
my.data$status[my.data$to == 0] <- 0
# Rename 'to' == 0  to 'cens' for use in 'ext_mstate' function
my.data$to[my.data$to == 0] <- 'cens'
my.data$from = as.character(my.data$from) # needed for function below
#my.data$id = my.data$pin #needed for function below
# Create data frame with all possible transitions for when a patient is at risk
my.data_ext <- ext_mstate(my.data, tra, id.var='pin')
# output for check
save(my.data, file='data/ecmo_transitions.RData') # for 99_censored_ecmo.R
```

This report uses the data from `r data_date$month` `r data_date$day`. The planned database freeze is October 31.

We include patients with a date of ECMO initiation. We include patients aged 18 or over. We excluded patients not on V-V ECMO (n = `r n_exclude_ecmo_type`), and patients without daily observations (n = `r n_exclude_no_daily`). The total number of patients is `r n_patients`. The data are from `r n_sites` sites.

The patients started on ECMO between the dates `r format(ecmo_start, '%d-%B-%Y')` to `r format(ecmo_end, '%d-%B-%Y')`.

The key research question is: if a patient is sick enough to need ECMO, then does proning help? We predominantly use survival analysis to answer this question. 

### Descriptive statistics by prone status

Here we show descriptive statistics to summarise the included patients.
We show the statistics for the two groups of: "no prone" and "some prone", based on the patients receiving prone after starting ECMO.

```{r}
#
cont_vars = c('age','bmi','sofa','apache_ii')
ftab = desc_table_cont(
  indata=patients, 
  vars = cont_vars, 
  group_var = 'any_prone', 
  var_levels=c('No','Yes'),
  heading = c('No prone','Some prone'),
  include_total = TRUE) %>%
  fontsize(size=10, part='all') %>%
  autofit()
table_nums(name="cont_table","Continuous variables - table.", level=2) # title
ftab
```

The table shows the median and inter-quartile range (IQR) which covers the central 50%. The overall number of patients is `r n_patients`.

```{r}
for_plot = select(patients, pin, all_of(cont_vars)) %>%
  tidyr::gather(key='Variable', value='Value', -pin)
hplot = ggplot(data=for_plot, aes(x=Value))+
  geom_histogram(fill='cyan3')+
  xlab('')+
  ylab('Count')+
  facet_wrap(~Variable, scales='free')+
  theme_bw()
figure_nums(name="cont_histo", "Continuous variables - histograms.", level=2) # title
hplot
```


```{r}
cont_vars = c('compliance_respiratory_system','eotd_peep','pa_co2','pa_o2','eotd_fi_o2','pa_o2_fi_o2')
ftab = desc_table_cont(indata=daily_ecmo, vars=cont_vars, group_var='any_prone', var_levels=c('No','Yes'), heading=c('No prone','Some prone')) %>% # see 99_functions.R
  fontsize(size=8, part='all') %>%
  autofit()
table_nums(name="cont_table_daily","Continuous variables from daily data.", level=2) # title
ftab
```

These are the continuous variables that are collected daily. Here we show the results for the date the patient started ECMO.

```{r}
#
ftab = desc_table_cat(indata = patients, 
                      vars = c('sex','ethnicity','country','pre_ecmo_vent','ecmo_prone_before'), 
                      group_var = 'any_prone', 
                      var_levels = c('No','Yes'),
                      include_total = TRUE,
                      heading = c('No prone','Some prone')) # see 99_functions.R
table_nums(name="cat_table","Categorical variables.", level=2) # title
ftab
```


```{r}
to_tab = filter(patients, sex=='Female')
ptab = ftab = desc_table_cat(indata = to_tab, 
                             vars = c('pregnant'), 
                             group_var = 'any_prone', 
                             var_levels = c('No','Yes'),
                             include_total = TRUE,
                             heading = c('No prone','Some prone')) 
table_nums(name="preg_table","Pregnant (women only).", level=2) # title
ptab
```


```{r}
co_vars = c('comorbidity_hypertension','comorbidity_diabetes','comorbidity_chronic_cardiac_disease','comorbidity_chronic_kidney_disease','comorbidity_malignant_neoplasm','comorbidity_smoking','comorbidity_chronic_alcohol_abuse_eot','comorbidity_immuno_competent_eot','comorbidity_obesity')
ftab = desc_table_cat(indata = patients, 
                      vars = co_vars, 
                      group_var = 'any_prone', 
                      var_levels = c('No','Yes'),
                      include_total = TRUE,
                      heading = c('No prone','Some prone')) 
table_nums(name="comorb_table","Comorbidities.", level=2) # title
ftab
```


```{r}
# create time variables
patients = mutate(patients,
                  date_end = ifelse(is.na(date_hospital_discharge)==FALSE, date_hospital_discharge, date_death), # end date of death or discharge
                  date_end = as.Date(date_end, origin='1970-01-01'),
                  days_symptoms_hosp = as.numeric(date_admission - date_first_symptom) +0.5, 
                  days_hosp = as.numeric(date_end - date_admission) + 0.5,
                  prone_start_post_ecmo = ifelse(prone_start < date_ecmo & !is.na(prone_start), date_ecmo, prone_start), # ignore prone days prior to ECMO start
                  duration_prone = as.numeric(prone_end - prone_start_post_ecmo) + 0.5,
                  duration_prone = ifelse(prone_end <= date_ecmo, 0, duration_prone), # set to zero if prone ended on or before ECMO
                  duration_prone = ifelse(is.na(duration_prone), 0, duration_prone), # set missing to zero
                  duration_mv = date_mech_vent_discontinued - date_mechanical_ventilation +0.5)
#
time_vars = c('days_symptoms_hosp','days_hosp','duration_mv','dur_pre_ecmo_vent','duration_prone')
#
ftab = desc_table_cont(indata=patients, vars=time_vars, group_var='any_prone', var_levels = c('No','Yes'), heading=c('No prone','Some prone')) %>%
  fontsize(size=10, part='all') %>%
  autofit()
table_nums(name="time_table", "Time variables (in days)", level=2) # title
ftab
```

The duration of prone is only the days after starting ECMO.

```{r}
tab = group_by(patients, site_name) %>%
  summarise(count = n()) %>%
  ungroup() %>%
  mutate(xaxis = rank(count, ties='random'))
hplot = ggplot(data=tab, aes(x=xaxis, y=count))+
  geom_col(fill='cyan4', col=grey(0.7))+
  theme_bw()+
  scale_x_continuous(breaks=c(1,seq(10,60,10),n_sites))+
  ylab('Number of patients')+
  xlab('Site')
figure_nums(name="patients", "Patients per site", level=2) # title
hplot
# stats for text
site_med = quantile(tab$count, probs=0.5)
site_q1 = quantile(tab$count, probs=0.25)
site_q3 = quantile(tab$count, probs=0.75)
```

Sites are ordered by the number of patients.
The median number of patients per site is `r site_med` with an inter-quartile range of `r site_q1` to `r site_q3`.


```{r}
ftab = desc_table_cat(indata = patients, 
                      vars = 'outcome', 
                      group_var = 'any_prone', 
                      var_levels = c('No','Yes'),
                      include_total = TRUE,
                      heading = c('No prone','Some prone')) 
table_nums(name="outcome_table", "Final outcome", level=2) # title
ftab
```


### Patients over calendar time

```{r, fig.width=8, fig.height=6.5}
# calculate number in ECMO beds and number prone every day
time_dates = select(patients, pin, date_ecmo, date_ecmo_discontinued, prone_start, prone_end)
denom = numer = NULL
for (k in 1:nrow(time_dates)){
  if(!is.na(time_dates$date_ecmo[k])&!is.na(time_dates$date_ecmo_discontinued[k])){
    if(time_dates$date_ecmo_discontinued[k] >= time_dates$date_ecmo[k]){ # remove few errors
      denom = c(denom, seq(time_dates$date_ecmo[k], time_dates$date_ecmo_discontinued[k], 1))
    }
  }
  if(!is.na(time_dates$prone_start[k])){numer = c(numer, seq(time_dates$prone_start[k], time_dates$prone_end[k], 1))}
}
prone_count = data.frame(table(numer)); names(prone_count) = c('date','numer')
denom_count = data.frame(table(denom)); names(denom_count) = c('date','denom')
counts = full_join(prone_count, denom_count, by='date') %>%
  replace(., is.na(.), 0) %>% # replace NA with zero
  mutate(not = denom - numer, # numbers not on ECMO
         date = as.Date(as.numeric(as.character(date)), origin='1970-01-01'))
# data for plot 1
dplot1 = select(counts, date, not, numer) %>%
  tidyr::gather(`numer`, `not`, key='type', value = 'count', -`date`) %>%
  mutate(
    count = ifelse(is.na(count), 0, count))
# data for plot 2
dplot2 = mutate(counts, p = 100 * numer / denom) # percent
# plot 1
x_breaks = as.Date(paste('01-', 1:12, '-2020', sep=''), format='%d-%m-%Y')
plot1 = ggplot(data=dplot1, aes(x=date, y=count, fill=factor(type)))+
  geom_bar(stat='identity')+
  theme_bw()+
  ylab('Number of patients')+
  scale_x_continuous(breaks=x_breaks, labels=month.abb)+
  xlab('')+ # not needed for top plot
  scale_fill_manual(NULL, values=c(c("#FF8C00", "#00688B")), labels=c('ECMO and supine','ECMO and prone'))+
  theme(legend.position = c(0.8,0.75),
        plot.margin=unit(c(t = 1, r = 1, b = 0, l = 1), "pt"))
# plot 2
plot2 = ggplot(data=dplot2, aes(x=date, y=p))+
  geom_line()+
  geom_smooth(method='loess', span=0.85)+
  scale_x_continuous(breaks=x_breaks, labels=month.abb)+
  ylab('Percent of prone patients')+
  xlab('Date')+
  theme_bw()+
  theme(plot.margin=unit(c(t = 1, r = 1, b = 5, l = 1), "pt"))
figure_nums(name="prone_time", "Numbers of ECMO and prone patients over calendar time", level=1) # title
grid.arrange(plot1, plot2, nrow=2)
# export
jpeg(filename='figures/prone/prone_calendar.jpg', width=5, height=4, units='in', res=400, quality = 100)
grid.arrange(plot1, plot2, nrow=2)
invisible(dev.off())
```

The top plot shows the total number of patients on ECMO every day, and the number of ECMO patients in prone position.
The bottom plot shows the daily percent of patients on ECMO (black line) and a smoothed estimate of the percent (blue line). The proportion of patients in prone position has decreased over time. The increase in the smoothed estimate to October is based on very few patients. 

We do not have the exact calendar time because dates have been scrambled by around plus/minus 3 days to preserve patient anonymity. 

# Model diagram

The figure below shows the four states in the multi-state survival model. Patients start in the "ECMO & supine" state on the day they begin ECMO. Patients in the prone position on their first day of ECMO start in the "ECMO & Prone" state. Patients can move from Supine to Prone, and back from Prone to Supine.

```{r}
states = c('ECMO &\nsupine','ECMO &\nprone','Death','Discharge\nfrom hospital')
# Matrix of joins between states:
M = matrix(data=c(
  2,1,
  1,2,
  3,1,
  4,1,
  3,2,
  4,2
), byrow = TRUE, ncol=2)
figure_nums(name="model_diagram", "The four states in the model and the possible movements between states", level=1) # title
multistate_diagram(states=states, links = M, pos = c(1, 2, 1), arr.pos=0.65, box.size = 0.14) # see 99_functions.R
jpeg(filename='figures/prone/state_diagram.jpg', width=4, height=4, units='in', res=400, quality = 100)
multistate_diagram(states=states, links = M, pos = c(1, 2, 1), arr.pos=0.65, box.size = 0.14) # see 99_functions.R
invisible(dev.off())
```


### Table of transitions between states

This table shows the transitions (movements) between the states (`r figure_nums("model_diagram", display = "cite")`).

```{r}
tab = as.data.frame(with(my.data, table(from, to))) %>%
  mutate(from = case_when(
    from==1 ~ 'ECMO & Supine',
    from==2 ~ 'ECMO & Prone')
  ) %>%
  tidyr::spread(to, Freq)
# table header
table_header <- data.frame(
  col_keys = c('from', 'cens','1','2','3','4'),
  h1 = c('', rep("State moved to", 5)),
  h2 = c('State moved from', 'Censored','Supine','Prone','Discharge','Death'),
  stringsAsFactors = FALSE )
ftab = flextable(tab) %>%
  set_header_df(mapping = table_header, key = "col_keys" ) %>% # add header
  merge_h(i=1, part = "header") %>% # merge "To" column headers
  align(align='center', part='all') %>% # centre text
  theme_box() %>%
  autofit() 
table_nums(name="transitions", "Transition numbers between states") # title
ftab
```

# Models of prone

Here we examine whether there are important predictors of whether patients get prone positioning. This is important because prone positioning was not a randomly allocated treatment, hence patients who get prone may have important differences from patients who do not. This could then cause confounding when we examine the effects of prone on death and discharge. Understanding the differences between patients who did and did not receive prone positioning will help us understand the potential biases and will also be use to create weights for a propensity weighted analysis.

### Model diagram (for models of prone)

```{r}
states = c('ECMO &\nsupine','ECMO &\nprone','Death','Discharge\nfrom hospital')
# Matrix of joins between states:
M = matrix(data=c(
  2,1,
  3,1,
  4,1
), byrow = TRUE, ncol=2)
figure_nums(name="model_diagram2", "The three states in the model and the possible movements between states", level=1) # title
multistate_diagram(states=states, links = M, pos = c(1, 2, 1), arr.pos=0.65, box.size = 0.14) # see 99_functions.R
jpeg(filename='figures/prone/state_diagram_prone.jpg', width=4, height=4, units='in', res=400, quality = 100)
multistate_diagram(states=states, links = M, pos = c(1, 2, 1), arr.pos=0.65, box.size = 0.14) # see 99_functions.R
invisible(dev.off())
```


### Cumulative risk plot for prone positioning

The plot below shows the times that patients received prone positioning. For this analysis we examine the time to first prone position and do not consider whether a patient was later moved to supine.

```{r}
## reminder:
# State 1: ECMO & supine
# State 2: Prone
# State 3: Discharge
# State 4: Death

# once a patient hits prone, we are no longer interested for this analysis
for_prone = filter(my.data, 
                   from==1 | (from==2 & entry==0)) %>% # only transitions from not-prone or prone at time zero
  mutate(exit = ifelse(from==2, 0, exit), # reset time to zero for those in prone at entry
         to = ifelse(from==2, 2, to)) %>% # switch those at time zero to prone
  group_by(pin) %>%
  arrange(pin, entry) %>%
  slice(1) %>% # just first transition (not interested in later transitions)
  ungroup()

# cumulative incidence
cum_inc = with(for_prone, cuminc(ftime=exit, fstatus=to, cencode='cens'))
g = ggcompetingrisks(cum_inc,
                     xlab = 'Days since ECMO started',
                     ylab = 'Cumulative probability',
                     xlim = c(0,censor_day), # 
                     title = '',
                     #                 gsep = '_', # dummy to prevent groups being created from spaces
                     cumcensor = TRUE, # does not work
                     legend.title='',
                     risk.table = TRUE,
                     conf_int = TRUE)
# output modified plot
gmod = g + geom_line(size = 1.05) + # make lines thicker
  scale_x_continuous(breaks=c(0, 20, 40, 60, 80))+
  scale_y_continuous(breaks=seq(0, 0.4, 0.1))+
  scale_colour_manual(name="State moved to:", values=c(2,3,1), labels=c("Prone","Discharge",'Death')) +  # change labels
  theme(
    panel.grid.major = element_line(colour=grey(0.8)),
    strip.background = element_blank(), 
    strip.text.x = element_blank()) # remove facet_wrap labels
figure_nums(name="cumunc_prone", 'Cumulative probabilities over time of moving to prone, discharge and death states') # title
gmod
jpeg(filename='figures/prone/cumulative_prone.jpg', width=5, height=4, units='in', res=400, quality = 100)
print(gmod)
invisible(dev.off())
```

Many patients were already prone on the day ECMO started. Those patients who died or were discharged may have eventually experienced prone positioning, but experienced these other events before they could experience prone. We need to account for this censoring when examining the predictors of prone.

## Survival model for prone positioning

We use a Weibull survival model to examine the time to prone position. This models the time to prone position whilst accounting for censoring due to death and discharge, or administrative censoring if the patient is still on ECMO at the end of the study.

```{r, winbugs_model}
# bugs model
bfile = 'bugs_survival_prone.txt'
bugs = file(bfile, 'w')
cat('model{
    for (i in 1:N){
      days[i] ~ dweib(r, mu[i])I(cens[i],) # right-censoring
      log(mu[i]) <- int + beta[1]*sex[i] + beta[2]*age[i] + beta[3]*calendar_time[i] + beta[4]*bmi[i] + beta[5]*prone_before[i] + centre[site[i]]
      bmi[i] ~ dnorm(mu.bmi, tau.bmi) # missing BMI
      prone_before[i] ~ dbern(p) # missing prone before
    }
    # priors
    mu.bmi ~ dnorm(0, 0.1) # vaguely informative (scaled)
    p ~ dbeta(1, 1) # 
    tau.bmi ~ dgamma(1, 1)
    int ~ dnorm(0, 0.0001)
    for(k in 1:5){beta[k] ~ dnorm(0, 0.0001)}
    for(k in 1:M){ # random effect for site
      centre[k] ~ dnorm(0, tau.centre)
    }
    tau.centre ~ dgamma(1, 1)
    r ~ dexp(1)
}', file=bugs)
close(bugs)
```

```{r, include=FALSE}
# what events are censored
cens_events = c('3','4','cens') # discharge, death and admin censoring; moving to prone ('2') is the event
#
for_prone = mutate(for_prone,
                   site_num = as.numeric(as.factor(site_name)), # site number for winbugs
                   # add tiny time because time zero not allowed
                   days = ifelse(to %in% cens_events, NA, exit+0.001), # A right censored observation has a missing observed time ...
                   cens = ifelse(to %in% cens_events, exit+0.001, 0)) # ... and a minimum time equal to the censored time
# scalars:
N = nrow(for_prone)
M = max(for_prone$site_num)

# set up survival data for winbugs
bdata = list(N = N, 
             M = M,
             site = for_prone$site_num,
             days = for_prone$days,
             cens = for_prone$cens,
             prone_before = as.numeric(for_prone$ecmo_prone_before=='Yes'),
             sex = as.numeric(for_prone$sex=='Male'),
             age = (for_prone$age-50)/10, # scaled
             bmi = (for_prone$bmi-30)/5, # scaled
             calendar_time = (as.numeric(for_prone$date_ecmo)-18367)/31 ) # centre and scale to one month
```

```{r, include=FALSE}
## run the survival model in WinBUGS
n.chains= 2
MCMC = 20000 # increase to 10,000 for final run
thin = 5 # 

# initial values for single variable outcome (use mean for intercept)
inits = list(r=1, beta=rep(0,5), p=0.5, int=0, tau.centre=1, centre=rep(0, M), mu.bmi=0, tau.bmi=1) 
inits = rep(list(inits), n.chains) # repeat initial values per chain

## run WinBUGS
parameters = c('r','beta','int','centre','tau.centre','bmi')
# do not run model if there are already estimates (save time)
any_model = length(dir(path='results', pattern='prone')) > 0
# run model
if(any_model == FALSE){
  bugs_results_prone = bugs(data=bdata, inits=inits, parameters=parameters, model.file=bfile, DIC=TRUE, n.chains=n.chains, n.iter=MCMC*thin*2, n.thin=thin, bugs.seed=87947, debug=debug, bugs.directory=bugs_location)
  save(bugs_results_prone, file='results/survival_prone.RData')
}
if(any_model == TRUE){
  load('results/survival_prone.RData')
}
# 
```

The table below is the hazard ratios and 95% credible intervals for the survival model with prone. 
A hazard ratio of 1 indicates no change in risk, whereas a hazard ratio above 1 indicates increased risk, and a hazard ratio below 1 indicated decreased risk.

```{r}
## tabulate the results
# get beta estimates
beta = bugs_extract(inbugs=bugs_results_prone, var='beta', exp=TRUE, names=c('Male','Age (+10 years)','Calendar time (+30 days)','BMI (+5 kg/m2)',"Prone before ECMO"), digits=2, make.ci=TRUE) # from 99_functions.R
ftab = rename(beta, 'HR' = 'mean') %>%
  flextable() %>%
  theme_box() %>% # nice theme
  align(align='center', part='all') %>% # centre text
  autofit()
table_nums(name="hr_weibull_prone", "Hazard ratios and 95% credible intervals for prone position from the Weibull model", level=1) # title, increase level
ftab
```

Older patients have a reduced hazard of being put in a prone position (HR = `r beta$mean[2]`).
Patients with a higher BMI have a reduced hazard of being put in a prone position (HR = `r beta$mean[4]`).

There were `r sum(for_prone$to=='cens')` patients who were administratively censored, meaning they were not yet discharged or dead by the end of the study.

## Survival model for prone positioning without random site effect

Here we repeat the above model but without the random site effect.

```{r, winbugs_model}
# bugs model
bfile = 'bugs_survival_prone_no_site.txt'
bugs = file(bfile, 'w')
cat('model{
    for (i in 1:N){
      days[i] ~ dweib(r, mu[i])I(cens[i],) # right-censoring
      log(mu[i]) <- int + beta[1]*sex[i] + beta[2]*age[i] + beta[3]*calendar_time[i] + beta[4]*bmi[i] + beta[5]*prone_before[i]
      bmi[i] ~ dnorm(mu.bmi, tau.bmi) # missing BMI
      prone_before[i] ~ dbern(p) # missing prone before
    }
    # priors
    mu.bmi ~ dnorm(0, 0.1) # vaguely informative (scaled)
    p ~ dbeta(1, 1) # 
    tau.bmi ~ dgamma(1, 1)
    int ~ dnorm(0, 0.0001)
    for(k in 1:5){beta[k] ~ dnorm(0, 0.0001)}
    r ~ dexp(1)
}', file=bugs)
close(bugs)
```

```{r, include=FALSE}
# set up survival data for winbugs (without site)
bdata = list(N = N, 
             days = for_prone$days,
             cens = for_prone$cens,
             prone_before = as.numeric(for_prone$ecmo_prone_before=='Yes'),
             sex = as.numeric(for_prone$sex=='Male'),
             age = (for_prone$age-50)/10, # scaled
             bmi = (for_prone$bmi-30)/5, # scaled
             calendar_time = (as.numeric(for_prone$date_ecmo)-18367)/31 ) # centre and scale to one month
```

```{r, include=FALSE}
## run the survival model in WinBUGS
n.chains= 2
MCMC = 20000 # increase to 10,000 for final run
thin = 5 # 

# initial values for single variable outcome (use mean for intercept)
inits = list(r=1, beta=rep(0,5), p=0.5, int=0, mu.bmi=0, tau.bmi=1) 
inits = rep(list(inits), n.chains) # repeat initial values per chain

## run WinBUGS
parameters = c('r','beta','int','bmi')
# do not run model if there are already estimates (save time)
any_model = length(dir(path='results', pattern='prone_no_site')) > 0
# run model
if(any_model == FALSE){
  bugs_results_prone_no_site = bugs(data=bdata, inits=inits, parameters=parameters, model.file=bfile, DIC=TRUE, n.chains=n.chains, n.iter=MCMC*thin*2, n.thin=thin, bugs.seed=87947, debug=debug, bugs.directory=bugs_location)
  save(bugs_results_prone_no_site, file='results/survival_prone_no_site.RData')
}
if(any_model == TRUE){
  load('results/survival_prone_no_site.RData')
}
# 
```


```{r}
## tabulate the results
# get beta estimates
beta = bugs_extract(inbugs=bugs_results_prone_no_site, var='beta', exp=TRUE, names=c('Male','Age (+10 years)','Calendar time (+30 days)','BMI (+5 kg/m2)',"Prone before ECMO"), digits=2, make.ci=TRUE) # from 99_functions.R
ftab = rename(beta, 'HR' = 'mean') %>%
  flextable() %>%
  theme_box() %>% # nice theme
  align(align='center', part='all') %>% # centre text
  autofit()
table_nums(name="hr_weibull_prone", "Hazard ratios and 95% credible intervals for prone position from the Weibull model with no site effect")
ftab
```

Older patients have a reduced hazard of being put in a prone position (HR = `r beta$mean[2]`).
Patients with a higher BMI have a reduced hazard of being put in a prone position (HR = `r beta$mean[4]`).



### Check Weibull assumption for time to prone

```{r}
# see http://www.drizopoulos.com/courses/emc/ep03_%20survival%20analysis%20in%20r%20companion
# a) refit model using R (non-Bayesian)
for_residual = mutate(for_prone, 
                      ecmo_prone_before = ifelse(is.na(ecmo_prone_before), 2, ecmo_prone_before),
                      bmi = ifelse(is.na(bmi), 30, bmi)) # simple replace of missing with median
#                      )
fit_weib <- survreg(Surv(exit+0.001, event=to =='2') ~ bmi + sex + age + date_ecmo + ecmo_prone_before, data = for_residual) # event is prone
# b) make residuals
fitted_values <- fit_weib$linear.predictors
resids <- (log(fit_weib$y[, 1]) - fitted_values) / fit_weib$scale
# c) plot
resKM <- survfit(Surv(resids, event=to =='2') ~ 1, data = for_residual)
plot(resKM, mark.time = FALSE, xlab = "AFT Residuals", ylab = "Survival Probability")
xx <- seq(min(resids), max(resids), length.out = 35)
yy <- exp(- exp(xx))
lines(xx, yy, col = "red", lwd = 2)
legend("bottomleft", c("KM estimate", "95% CI KM estimate", 
    "Survival function of Weibull distribution"), 
    lty = c(1,2,1), col = c(1,1,2), bty = "n")
```

# Models of death and discharge

Here we examine the key outcomes of death and discharge. We examine whether prone positioning is associated with death or discharge whilst controlling for potential confounders.

```{r, death_probs}
# Analysis ------------------------------------------------------------------------

## Cox model stratified by transition
c_2 <- coxph(Surv(entry, exit, status==1) ~ strata(trans), data= my.data_ext, method = "breslow")
# msfit calculates baseline hazards
msf_2 <- msfit(c_2, trans = tra)
# probtrans calculates transition probabilities, prediction from day 0
pt_2 <- probtrans(msf_2, predt = 0)

# Retrieve predicted death risk at last day 
# Results in Table 3 (https://bmcmedresmethodol.biomedcentral.com/articles/10.1186/s12874-020-01082-z/tables/3)
pred_1 = (pt_2[[1]]) %>% # Patients starting ECMO
  mutate(from='ECMO & supine') %>%
  tail(1) # estimate on last day
pred_2 = (pt_2[[2]]) %>%  # Patients starting Prone
  mutate(from='ECMO & prone') %>%
  tail(1) # estimate on last day
preds = bind_rows(pred_1, pred_2) %>%
  mutate(pstate4 = roundz(pstate4*100, 1), # death probability, change to percent
         se4 = roundz(se4*100, 1),
         death = paste(pstate4, ' (', se4, ')', sep='')) %>%
  select(from, death) 
```

```{r}
# ELOS gives expected length of stay for up to censor days days in the states for patients in the states at day 0
# Results shown in Table 3
LOS_mat_2 <- ELOS(pt_2, censor_day)

### Bootstrapped confidence intervals for length of stay
# change class of data frame
class(my.data_ext) <- c("msdata", "data.frame")
# give transition matrix attribute
tmat <- attr(my.data_ext, "trans")
# bootstrap performed with msboot function in mstate, 1000 samples
set.seed(1906)
B = 200 # Should be 5000 for final run
my.data_ext$id = as.numeric(as.factor(my.data_ext$pin)) # needed for function
Example2_boot <- msboot(theta = LOS_boot, censor_day=censor_day, data=my.data_ext, id="id", B = B) # 

# Confidence Intervals
confidence_intervals = reshape2::melt(Example2_boot) %>%
  filter(!is.na(value)) %>%
  group_by(Var1) %>%
  summarise(lower = quantile(value, probs=0.025),
            upper = quantile(value, probs=0.975)) %>%
  ungroup()

# merge with estimates
both = bind_cols(as.vector(LOS_mat_2), confidence_intervals[1:16,]) %>% # assumes 4x4 matrix
  rename('mean' = '...1') %>%
  mutate(mean = roundz(mean, 1),
         lower = roundz(lower, 1),
         upper = roundz(upper, 1),
         cell = paste(mean, ' (', lower, ', ', upper, ')', sep=''))
```

```{r, include=FALSE}
### Expected lengths of stay in days for ending in each of the four states
# Initial Distribution of the patients in the 4 states at day 0
# get starting numbers
starts = filter(my.data, entry==0) %>% 
  group_by(from) %>% 
  summarise(n=n()) %>% 
  ungroup() %>% 
  mutate(p = prop.table(n))
init_dis_2 <- c(starts$p, 0, 0) # to update

# Multiply LOS_mat_1 wifth initial distribution to get weighted average of expected lengths of stay for entire cohort
# Results shown in Table 3
LOS_cohort_2 <- (init_dis_2 %*% LOS_mat_2) # sums for each of the four "to" states
los_cohort_tab = data.frame(t(roundz(LOS_cohort_2,1)))
colnames(los_cohort_tab) = c("ECMO", "Prone", "Discharge", "Death")

# Create weighted average of transition probabilities for progress of entire cohort
# Used to produce full cohort results in Table 3 
pt_fc_2 <- pt_2
fc_2 <- pt_fc_2[[1]] * starts$p[1] + pt_fc_2[[2]] * starts$p[2]
pt_fc_2[[1]] <- fc_2
```

```{r, table}
## make nice table
states = c("ECMO & supine", "ECMO & prone", "Discharge", "Death")
#
both_mat = matrix(both$cell, ncol=4)
colnames(both_mat) = states
# more formatting
LOS_mat_tab = as.data.frame(both_mat)
LOS_mat_tab$from = states
LOS_mat_tab = LOS_mat_tab[1:2,] # discard death and discharge from rows
# add death risk
#LOS_mat_tab = left_join(LOS_mat_tab, preds, by='from')
LOS_mat_tab  = select(LOS_mat_tab, from, everything()) # from from to start
#

## nice table
# table header
table_header <- data.frame(
  col_keys = c( 'from', states), #, 'death' ),
  h1 = c('', rep("State moved to", 4)),#, ""),
  h2 = c('State moved from', states),#, 'Percent dead'),
  stringsAsFactors = FALSE )
# make table
ftab = flextable(LOS_mat_tab) %>%
  set_header_df(mapping = table_header, key = "col_keys" ) %>% # add header
  merge_h(i=1, part = "header") %>% # merge "To" column headers
  theme_box() %>% # nice theme
  align(align='center', part='all') %>% # centre text
  autofit()
table_nums('elos', 'Expected length of stay in days', level=1) # title, increase level
ftab
```

The table shows expected lengths of stay between states, with the current state in the rows on the left and the new state in the columns along the top. The results are up to day `r censor_day`. These results do not adjust for potential confounders.

The days in parentheses are 95% confidence intervals based on `r B` bootstrap replications.

<!--- The right-most column shows the estimated death percentage and standard error in parentheses.--->
  
  The results use `r n_patients` patients.

```{r}
figure_nums(name="state_plot", "Plot of predicted probabilities for the four states") # increase level
plot_it = function(){
  par(mai=c(0.9,0.9,0.5,0.1), las=1) # b, l, t, r; leave room at top for legend
  colours = c("khaki1","indianred1","cornflowerblue","gray")
  plot(pt_fc_2, from = 1,  ord = c(4,2,1,3), type= "filled", cols = colours,
       lwd= 2, xlab = "Days since ECMO started", ylab = "Predicted Probabilities", 
       cex.lab = 1.25, legend = rep("", 4),
       main= NULL, axes=FALSE)
  # better x-axis ticks:
  axis(side=2, at=seq(0,1,0.2), labels=TRUE) # y-axis
  # 90 days
  axis(side=1, at=seq(0,90,10), labels=FALSE)
  axis(side=1, at=seq(0,90,5), labels=TRUE, tick=FALSE)
  # legend at top
  par(xpd=TRUE)
  labels = c('ECMO & supine','ECMO and prone','Discharge','Death')
  legend(0, 1.2, labels, col=colours, border=NA, fill=colours, bty='n', ncol=2)
}
plot_it()
jpeg(filename='figures/prone/patients_states_time.jpg', width=5, height=4, units='in', res=400, quality = 100)
plot_it()
invisible(dev.off())
```

There is relatively little change in patients' states after day 40.

## Cumulative risk plots

```{r, fig.width=8, fig.height=6.5}
# for nice labels
my.data = mutate(my.data, from.nice=case_when(
  from==1 ~ 'Supine',
  from==2 ~ 'Prone'
  ))
# cumulative incidence
cum_inc = with(my.data, cuminc(ftime=exit, fstatus=to, group=from.nice, cencode='cens'))
# remove lines not needed (prone to prone and supine to supine)
cum_inc[[2]] = NULL
cum_inc[[2]] = NULL
# plot
g = ggcompetingrisks(cum_inc,
                 xlab = 'Days since ECMO started',
                 ylab = 'Cumulative probability',
                 xlim = c(0, censor_day), # 
                 title = '',
#                 gsep = '_', # dummy to prevent groups being created from spaces
                 cumcensor = TRUE, # does not work
                 legend.title='',
                 risk.table = TRUE,
                 conf_int = TRUE)
# create plot with better facet labels
to_plot = mutate(g$data, 
                 facet = paste('State moved from: ', group, sep=''))
cum_plot = ggplot(data=to_plot, aes(x=time, y=est, col=factor(event)))+
  geom_line(size = 1.05) + # make lines thicker
  facet_wrap(~facet)+
  xlab('Days since ECMO started')+
  ylab('Cumulative probability')+
  theme_bw()+
  scale_x_continuous(breaks=c(0, 20, 40, 60, 80))+
  scale_y_continuous(breaks=seq(0, 0.9, 0.1))+
  scale_colour_manual(name="State moved to:", values=c(4,2,3,1), labels=c("Supine","Prone","Discharge",'Death')) +  # change labels
  theme(
    panel.grid.minor = element_blank(),
    panel.grid.major = element_line(colour=grey(0.8)),
    text = element_text(size=20), 
    legend.position = 'top',
    legend.title = element_text(size=16),
    legend.text = element_text(size=16))
figure_nums(name="cuminc_death", "Cumulative probabilities over time for the four states.")
cum_plot
jpeg('figures/prone/competing_risks_ecmo_prone.jpg', width=7, height=4.5, units='in', res=400, quality=100)
print(cum_plot)
invisible(dev.off())
```

The plots shows the cumulative probability over time of moving between states. The left panel shows the probabilities for patients in the "ECMO and prone" state, and the right panel shows the probabilities for patients in the "ECMO and supine" state. There is a steady accumulation of deaths in patients in the supine group (black line in right panel), although this slows after around 40 days. The movement of patients from the supine to the prone group greatly reduces after around 12 days (red line in right panel). Patients in the supine position are more likely to eventually experience discharge than patients in the prone position (green lines).

## Survival models

Here we model the survival time and outcome (discharge/death) with the aim of examining prone position whilst adjusting for confounders.

```{r, winbugs_model_death_discharge}
# bugs model
bfile = 'bugs_survival_prone_death.txt'
bugs = file(bfile, 'w')
cat('model{
  for (i in 1:N){
    days[i] ~ dweib(r, mu[i])I(cens[i],) # right-censoring
    log(mu[i]) <- int + beta[1]*sex[i] + beta[2]*age[i] + beta[3]*calendar_time[i] + beta[4]*bmi[i] + beta[5]*prone[i] + beta[6]*prone_before[i]+ centre[site[i]]
    bmi[i] ~ dnorm(mu.bmi, tau.bmi) # missing BMI
    prone_before[i] ~ dbern(p) # missing prone before
  }
  ## priors
  # missing data:
  mu.bmi ~ dnorm(0, 0.1) # vaguely informative (scaled)
  tau.bmi ~ dgamma(1, 1)
  p ~ dbeta(1, 1) # 
  # regression model:
  int ~ dnorm(0, 0.0001)
  for(k in 1:6){beta[k] ~ dnorm(0, 0.0001)}
  for(k in 1:M){ # random effect for site
    centre[k] ~ dnorm(0, tau.centre)
  }
  tau.centre ~ dgamma(1, 1)
  r ~ dgamma(10, 10)
}', file=bugs)
close(bugs)
```

#### Time to death and discharge

```{r, include=FALSE}
## prepare the data for 'winbugs' or 'R', need to use clock reset method
prepare_survival = function(indata, cens_events, clock_reset= TRUE, version='winbugs'){
bugs_data = mutate(indata,
    id_num = as.numeric(as.factor(pin)), # make ID number for patients
    site_num = as.numeric(as.factor(site_name)), # make ID number for sites
    # clock reset
    exit = ifelse(entry > 0, exit - entry, exit), # subtract starting days if delayed start
    # add tiny time because time zero not allowed
    days = ifelse(to %in% cens_events, NA, exit+0.001), # A right censored observation has a missing observed time ...
    cens = ifelse(to %in% cens_events, exit+0.001, 0) # ... and a minimum time equal to the censored time
                   )

if(version=='R'){
  return(bugs_data)
}
if(version=='winbugs'){
N = nrow(bugs_data)
M = max(bugs_data$site_num)

# set up survival data for winbugs
bdata = list(N = N, 
             M = M,
             prone = as.numeric(bugs_data$from==2),
             prone_before = as.numeric(bugs_data$ecmo_prone_before=="Yes"),
             site = bugs_data$site_num,
             days = bugs_data$days,
             cens = bugs_data$cens,
             sex = as.numeric(bugs_data$sex=='Male'),
             age = (bugs_data$age-50)/10, # scaled
             bmi = (bugs_data$bmi-30)/5, # scaled
             calendar_time = (as.numeric(bugs_data$date_ecmo)-18367)/31 ) # centre and scale to one month
  return(bdata)
} # end of if
} # end of function
for_death = prepare_survival(indata=my.data, cens_events=c('1','2','3','cens')) # supine, prone, discharge or admin censored; event is death ('4')
for_discharge = prepare_survival(indata=my.data, cens_events=c('1','2','4','cens')) # supine, prone, death or admin censored; event is discharge ('3')
```

```{r, include=FALSE}
## run the survival models in WinBUGS
# could remove individual effect, it is not helping much
n.chains= 2
MCMC = 10000 # increase to 10,000 for final run
thin = 5 # increase to 5 for final run

# initial values for single variable outcome (use mean for intercept)
inits = list(r=1, beta=rep(0,6), int=0, tau.centre=1, centre=rep(0, for_death$M), mu.bmi=0, tau.bmi=1, p=0.74) 
inits = rep(list(inits), n.chains) # repeat initial values per chain
parameters = c('r','beta','int','centre','tau.centre','bmi') # ,'median','diff'
  
##
# a) death
# do not run model if there are already estimates (save time)
any_model = length(dir(path='results', pattern='bugs_death')) > 0
# run model
if(any_model == FALSE){
  bugs_results_death = bugs(data=for_death, inits=inits, parameters=parameters, model.file=bfile, DIC=FALSE, n.chains=n.chains, n.iter=MCMC*thin*2, n.thin=thin, bugs.seed=87947, debug=debug, bugs.directory=bugs_location)
  save(bugs_results_death, file='results/survival_bugs_death.RData')
}
if(any_model == TRUE){
  load('results/survival_bugs_death.RData')
}
# b) discharge
any_model = length(dir(path='results', pattern='bugs_discharge\\.RData')) > 0
# run model
if(any_model == FALSE){
  bugs_results_discharge = bugs(data=for_discharge, inits=inits, parameters=parameters, model.file=bfile, DIC=FALSE, n.chains=n.chains, n.iter=MCMC*thin*2, n.thin=thin, bugs.seed=87947, debug=debug, bugs.directory=bugs_location)
  save(bugs_results_discharge, file='results/survival_bugs_discharge.RData')
}
if(any_model == TRUE){
  load('results/survival_bugs_discharge.RData')
}
# 
```


The table below is the hazard ratios (HRs) and 95% credible intervals for the competing risks of death and discharge. 
A hazard ratio of 1 indicates no change in risk, whereas a hazard ratio above 1 indicates increased risk, and a hazard ratio below 1 indicated decreased risk.

```{r}
## tabulate the results
# get beta estimates
beta_death = bugs_extract(inbugs=bugs_results_death, var='beta', exp=TRUE, names=c('Male','Age (+10 years)','Calendar time (+30 days)','BMI (+5 kg/m2)','Prone position','Prone before ECMO'), digits=2, make.ci=TRUE) # from 99_functions.R
beta_discharge = bugs_extract(inbugs=bugs_results_discharge, var='beta', exp=TRUE, names=NULL, digits=2, make.ci=TRUE) # from 99_functions.R
for_table = bind_cols(beta_death, beta_discharge) %>%
  select(-row, -Var) 
names(for_table) = c('var','mean1','ci1','mean2','ci2')
#
header = read.table(header=TRUE, stringsAsFactors = FALSE, sep="!", text='
col_keys!colB!colA
var! !Variable
mean1!Death!HR
ci1!Death!95% CI
mean2!Discharge!HR
ci2!Discharge!95% CI
')
#
ftab = flextable(for_table) %>%
  set_header_df(mapping = header, key = "col_keys" ) %>%
  merge_h(part = "header", i = 1) %>%
  theme_box() %>% # nice theme 
  align(align='center', part='all') %>% # centre text
  autofit()
table_nums(name="hr_weibull", "Hazard ratios and 95% credible intervals for death and discharge from the Weibull model")
ftab
## numbers used in text:
# bayesian p-value for prone
pval_prone = sum(bugs_results_death$sims.list$beta[,5]<0) / (length(bugs_results_death$sims.list$beta[,5]) + 1)
# get number administratively censored
censored = group_by(my.data, pin) %>%
  arrange(pin, entry) %>%
  slice(n()) %>% # last per patient
  filter(to == 'cens') # admin censored
```

Prone position somewhat reduces the hazard of death (HR = `r for_table$mean1[5]`). 

Older age increases the hazard of death (HR = `r for_table$mean1[2]`). Older age decreases the hazard of discharge (HR = `r for_table$mean2[2]`), which increases the length of time on ECMO.
Calendar time has a strong effect on discharge (HR = `r for_table$mean2[3]`), with more days on ECMO (reduced discharge hazard) over calendar time. 

There were `r nrow(censored)` patients who were administratively censored, meaning they were not yet discharged or dead by the end of the study.

### Check Weibull assumption for death and discharge

#### a) Death

```{r}
# see http://www.drizopoulos.com/courses/emc/ep03_%20survival%20analysis%20in%20r%20companion
# a) refit model using R (non-Bayesian)
for_residual = mutate(my.data, 
                      exit = ifelse(entry > 0, exit - entry, exit), # subtract starting days if delayed start
                      ecmo_prone_before = ifelse(is.na(ecmo_prone_before), 2, ecmo_prone_before), # mode
                      bmi = ifelse(is.na(bmi), 30, bmi)) # simple replace of missing with median
#                      )
fit_weib <- survreg(Surv(exit+0.001, event=to =='4') ~ bmi + sex + age + date_ecmo + I(from==2) + ecmo_prone_before, data = for_residual) # event is death
# b) make residuals
fitted_values <- fit_weib$linear.predictors
resids <- (log(fit_weib$y[, 1]) - fitted_values) / fit_weib$scale
# c) plot
resKM <- survfit(Surv(resids, event=to =='4') ~ 1, data = for_residual)
plot(resKM, mark.time = FALSE, xlab = "AFT Residuals", ylab = "Survival Probability")
xx <- seq(min(resids), max(resids), length.out = 35)
yy <- exp(- exp(xx))
lines(xx, yy, col = "red", lwd = 2)
legend("bottomleft", c("KM estimate", "95% CI KM estimate", 
    "Survival function of Weibull"), 
    lty = c(1,2,1), col = c(1,1,2), bty = "n")
# residual plot
# qqnorm(resids); qqline(resids)
```

#### b) Discharge

```{r}
# see http://www.drizopoulos.com/courses/emc/ep03_%20survival%20analysis%20in%20r%20companion
# a) refit model using R (non-Bayesian)
fit_weib <- survreg(Surv(exit+0.001, event=to =='3') ~ bmi + sex + age + date_ecmo + I(from==2) + ecmo_prone_before, data = for_residual) # event is death
# b) make residuals
fitted_values <- fit_weib$linear.predictors
resids <- (log(fit_weib$y[, 1]) - fitted_values) / fit_weib$scale
# c) plot
resKM <- survfit(Surv(resids, event=to =='3') ~ 1, data = for_residual)
plot(resKM, mark.time = FALSE, xlab = "AFT Residuals", ylab = "Survival Probability")
xx <- seq(min(resids), max(resids), length.out = 35)
yy <- exp(- exp(xx))
lines(xx, yy, col = "red", lwd = 2)
legend("bottomleft", c("KM estimate", "95% CI KM estimate", 
    "Survival function of Weibull"), 
    lty = c(1,2,1), col = c(1,1,2), bty = "n")
# residual plot
# qqnorm(resids); qqline(resids)
```

## Cumulative survival models

The survival models in this section focus on the patients' final (or cumulative) outcome of death and discharge, whereas the previous models examined the instantaneous hazard of experiencing death or discharge. 

The table below shows the cumulative hazard ratios and 95% confidence intervals for death and discharge.

```{r}
my.data_crr = mutate(my.data, 
                     # clock reset
                     exit = ifelse(entry > 0, exit - entry, exit), # subtract starting days if delayed start
                     # prepare variables
                     prone = as.numeric(from==2),
                     prone_before = as.numeric(ecmo_prone_before=='Yes'),
                     age = (age-50)/10, # scaled
                     bmi = (bmi-30)/5, # scaled
                     calendar_time = (as.numeric(date_ecmo)-18367)/31,
                     sex = as.numeric(sex=='Male'))
# small amounts of missing
my.data_crr = mutate(my.data_crr,
                     bmi = ifelse(is.na(bmi)==TRUE, 0, bmi), # use mean
                     prone_before = ifelse(is.na(prone_before)==TRUE, 1, prone_before)) # use mode
#
cov1 = as.matrix(select(my.data_crr, sex, age, calendar_time, bmi, prone, prone_before))
# death model
cum_rr_death = with(my.data_crr, crr(ftime=exit, fstatus=to==4, cov1=cov1, cencode='cens', maxiter = 100))
# discharge model
cum_rr_discharge = with(my.data_crr, crr(ftime=exit, fstatus=to==3, cov1=cov1, cencode='cens', maxiter = 100))
s_death = summary(cum_rr_death) # cannot use broom
s_discharge = summary(cum_rr_discharge) # cannot use broom
table_death = as.data.frame(s_death$coef); names(table_death) = c('coef','exp_coef','SE','z','pval')
table_discharge = as.data.frame(s_discharge$coef); names(table_discharge) = c('coef','exp_coef','SE','z','pval')
tab = bind_rows(table_death, table_discharge, .id='type') # combine tables
table = mutate(tab,
               z = qnorm(0.975),
               lower = coef - (z*SE), # make CIs
               upper = coef + (z*SE),
               HR = roundz(exp(coef), 2),
               lower = roundz(exp(lower), 2),
               upper = roundz(exp(upper), 2),
               CI = paste(lower, ' to ', upper, sep='')) %>%
  select(type, HR, CI) 
ftable = bind_cols(filter(table, type==1) %>% select(-type),
                   filter(table, type==2) %>% select(-type))
names(ftable) = c('HR1','CI1','HR2','CI2')
ftable = mutate(ftable, Variable=c('Male','Age (+10 years)','Calendar time (+30 days)','BMI (+5 kg/m2)',"Prone position","Prone before ECMO")) %>%
  select('Variable', everything()) %>% # move Variable first
  flextable() %>%
  add_header('Variable'='', 'HR1' = 'Death', 'CI1' = 'Death', 'HR2' = 'Discharge', 'CI2' = 'Discharge') %>%
  merge_h(i=1, part = "header") %>% # merge "To" column headers
  theme_box() %>% # nice theme
  align(align='center', part='all') %>% # centre text
  autofit() 
table_nums(name="crr_hazards", "Hazard ratios and 95% confidence intervals using a cumulative probability regression model")
ftable
```

Prone position decreases the probability of death and discharge.

##### Plot of model predictions for cumulative regression

```{r}
## plot predictions
# make predictions
pred.cov1 = as.matrix(data.frame(sex=1, age=0, calendar_time=0, bmi=0, prone=c(0,1)))
pred_death = as.matrix(predict(cum_rr_death, cov1 = pred.cov1))
pred_discharge = as.matrix(predict(cum_rr_discharge, cov1 = pred.cov1))
class(pred_death) = 'matrix'
class(pred_discharge) = 'matrix'
colnames(pred_death) = c('days','0','1')
colnames(pred_discharge) = c('days','0','1')
pred = bind_rows(as.data.frame(pred_death), as.data.frame(pred_discharge), .id='type') %>%
  tidyr::gather(`0` , `1`, key='prone', value='crr', -type) %>%
  mutate(type_nice = factor(type, levels=1:2, labels=c('Death','Discharge')))
# plot
pred_plot = ggplot(data=pred, aes(x=days, y=crr, col=factor(prone)))+
  geom_line(size=1.05)+
  theme_bw()+
  scale_color_manual('Prone', labels=c('No','Yes'), values=c('cyan4','darkorange'))+
  xlab('Days since ECMO started')+
  ylab('Cumulative probability')+
  theme(legend.position = c(0.15, 0.85))+
  facet_wrap(~type_nice)
figure_nums(name="crr_plot", "Plot of predicted cumulative probabilities of death and discharge over time for cumulative regression model for patients in prone and supine position.") # title
pred_plot
jpeg(filename='figures/prone/predicted_cumulative.jpg', width=5, height=4, units='in', res=400, quality = 100)
print(pred_plot)
invisible(dev.off())
```

The plots above show the cumulative predicted probabilities of death and discharge for a patient of average age and BMI. By the end of the follow-up time (day 80) there are very large gaps in the predicted probability of both death and discharge for patients in supine and prone position.  

# Statistical methods section for paper

We used multi-state survival analysis to examine this data because there are key changes that happen over time during a patient's ICU stay that must be accounted for. A simple cross-sectional analysis, e.g., examining the association between a patient experiencing any prone position (yes/no) and death (yes/no) using a chi-squared test or similar, over-simplifies the data and can introduce serious biases, such as the time-dependent bias [@Beyersmann08]. A simple approach also ignores the issue of competing risks, which can easily lead to flawed results and conclusions [@Wolkewitz2014]. In this case death and discharge are competing events, because once a patient has experienced one, they cannot experience the other. Because of this issue we used cumulative incidence plots to graphically show events over time rather than the more familiar Kaplan-Meier plots, because Kaplan-Meier plots generally over-estimate cumulative risks when there are competing events [@Wolkewitz2014].

We used a multi-state model with four states (see `r figure_nums("model_diagram", display = "cite")`). 
Patients started in the prone or supine states, and could move between these states during their ECMO treatment. 
We plotted the modelled number of patients in each state using the methods described in @Hazard2020. 
We used a parametric Weibull survival model to examine the hazards of moving to the death or discharge state, and whether this depended on prone positioning whilst adjusting for potential confounders. We also adjusted for each site using a random intercept, which allows sites to vary in their baseline risks. 
We scaled continuous variables so that the hazard ratios were more clinically relevant, e.g., age per ten year increase.

A key concern was confounders of prone position, which would be variables that predict both: 1) whether a patient is moved to a prone position, 2) the outcomes of death or discharge. To investigate this we used a survival model for time to prone position and included characteristics of the patients that we expected to influence whether they moved into a prone position during their stay. 

The model estimates were made using a Bayesian framework and so results are presented using a 95% credible interval instead of the more familiar 95% confidence interval. A 95% credible interval has a 95% probability of containing the estimate of interest, whereas a 95% confidence interval has a far less intuitive. interpretation [@Lunn12]. We avoid presenting p-values as there are often misinterpreted [@Goodman2008]. Instead we focus on the size of the estimated effects and their 95% credible intervals, which are more clinically relevant.

To verify our parametric survival models we also fitted non-parametric Cox models to check that they gave similar hazard ratios and confidence intervals. These models also used the multi-state approach. We used a leave-one-country out sensitivity analysis to check that the results were not dependent on one country.

We used summary statistics to describe the sample of patients, using percentages for categorical variables, and medians and inter-quartile ranges for continuous variables.

We used `r sessionInfo()$R.version$version.string` [@R] and WinBUGS version 1.4.2 [@Lunn12] for all analyses. The cumulative incidence plots and models were fitted using the 'cmprsk' package [@cmprsk] and the Cox models using the 'survival' package [@survival].
All the R and WinBUGS code used in the analysis is available here: https://github.com/agbarnett/covid_cccc. 

<!--- To check: Patients were administratively censored if they were still in ICU when the data were extracted (31 October 2020). --->

<!--- A good recent introduction to survival analysis is @Andersen2020. --->


## Alertness over time

```{r}
# turn my.data into long data (daily)
long = NULL
my.data.cens = mutate(my.data,
                      to = ifelse(to=='cens', '0', to)) %>%
  filter(exit - entry > 0.5) # remove few small jumps
for (k in 1:nrow(my.data.cens)){
  this_row = my.data.cens[k,]
  f = data.frame(pin=this_row$pin, days=seq(this_row$entry, this_row$exit-1, 1), state=as.numeric(this_row$from))
  f_last = data.frame(pin=this_row$pin, days=this_row$exit, state=as.numeric(this_row$to))
  long = bind_rows(long, f, f_last)
}
# quick check
check = group_by(long, days, state) %>%
  tally()
cplot = ggplot(data=check, aes(x=days, y=n, fill=factor(state)))+
  geom_bar(stat='identity')
#cplot

# get daily data
small = select(patients, pin, date_ecmo)
to_plot = full_join(daily, small, by='pin') %>%
  mutate(days = as.numeric(date_daily - date_ecmo)) %>%
  filter(!is.na(days),
         days >= 0,
         days <= censor_day) %>%
  select(pin, days, date_daily, glasgow_coma_score, avpu)

# merge with state information
to_plot = left_join(to_plot, long, by=c('pin','days')) %>%
  filter(state %in% c(1,2)) # just supine and prone

# daily mean by state
to_plot_mean = group_by(to_plot, days, state) %>%
  summarise(#meana = mean(avpu, na.rm=TRUE), # categorical
            meang = mean(glasgow_coma_score, na.rm=TRUE)) %>%
  filter(!is.na(meang))
# plot
gplot = ggplot(data=to_plot_mean, aes(x=days, y=meang, col=factor(state)))+
  geom_point()+
  scale_color_manual(NULL, labels=c('Supine','Prone'), values=1:2)+
  geom_smooth(span=0.9)+
  xlab('Days since ECMO start')+
  ylab('Average Glasgow coma score')+
  theme_bw()+
  theme(legend.position = c(0.15,0.85))
figure_nums(name="glasgow_plot", "Plot of daily average Glasgow coma scores by position.") # title
gplot
```


The plot shows a steady rise in score in the supine position. The patterns in the prone group are divergent, but are reliant on a small number of patients.

### Addition details for a methods appendix

_The following additional details could be added to the appendix of the paper._

There was a small amount of missing data for BMI (n = `r sum(is.na(my.data$bmi))`) and prone position prior to ECMO (n = `r sum(is.na(my.data$ecmo_prone_before))`), and these missing values were imputed using multiple imputation in the Bayesian model [@Lunn12]. We used a flat Beta(1,1) prior for the binary variable of prone position prior to ECMO, and a normal distribution prior for body mass index with a mean of 30 kg/m2 and standard deviation of `r round(5*sqrt(1/0.5))` kg/m2. 

The Bayesian estimates were made using `r n.chains` chains with a burn-in and sample of size `r MCMC` thinned by `r thin`. The convergence and mixing of the chains was checked visually.

# Appendices

These appendices contain additional results and checks of the models. These could be appendices to the paper.

## Plot of individual patient journeys

This plot tries to show individual patient journeys from admission to discharge or death.
The panels are split by the patient's final state (death, discharge or censored) and whether the started in a prone or supine position. 

```{r, fig.width=8, fig.height=7}
# State 0: Censored
# State 1: ECMO & supine
# State 2: Prone
# State 3: Discharge
# State 4: Death
## loop to expand days
expand = NULL
my.data = mutate(my.data, 
                 to = ifelse(to=='cens', 0, to), # changed censored back to 0
                 from = as.numeric(from),
                 to = as.numeric(to))
for (i in 1:nrow(my.data)){
  if(nrow(my.data[i,]) > 0){
    f = with(my.data[i,], data.frame(id=pin, state=from, days=entry:(exit-1)))
    flast = with(my.data[i,], data.frame(id=pin, state=to, days=exit))
  }
  if(nrow(my.data[i,]) == 0){
    f = NULL
    flast = with(my.data[i,], data.frame(id=pin, state=to, days=exit))
  }
  expand = bind_rows(expand, f, flast)
}
# order plot by exit time
exit_time = group_by(expand, id) %>%
  summarise(maxd = max(days)) %>%
  ungroup() %>%
  mutate(n = n(),
         maxd = maxd + runif(n=n), # add random noise to prevent ties
         xaxis = rank(maxd))
expand = left_join(expand, exit_time)
## split by events
# a) prone on entry
events = filter(expand, state==2 & days==0) %>% 
  select(id) %>%
  mutate(facet=2) # prone
# a) prone on entry
ultimate = group_by(expand, id) %>%
  arrange(-days) %>%
  slice(1) %>%
  select(id, state) %>%
  rename('final' = 'state')
# merge
expand_plus = left_join(expand, events, by='id') %>%
  left_join(ultimate, by='id') %>%
  mutate(facet = ifelse(is.na(facet)==TRUE, 1, facet)) %>% # switch to supine
  group_by(facet, final) %>%
  mutate(xaxis_plus = rank(xaxis), # redo x-axis to stop gaps
         xaxis_plus = as.numeric(as.factor(xaxis_plus)), # make into integers
         both = paste(facet, '.', final, sep=''),
         both_nice = case_when(
           both == '1.0' ~ 'Started in supine and censored',
           both == '2.0' ~ 'Started in prone and censored',
           both == '1.3' ~ 'Started in supine and discharged',
           both == '2.3' ~ 'Started in prone and discharged',
           both == '1.4' ~ 'Started in supine and died',
           both == '2.4' ~ 'Started in prone and died'
         )
  )
# just plot prone and supine, events done using facets
expand_plus = filter(expand_plus, state %in% c(1,2))
#
gplot = ggplot(data=expand_plus, aes(x=xaxis_plus, y=days, group=factor(id), col=factor(state)))+
  geom_line()+
  scale_x_discrete(breaks=NULL)+
  ylab('Days since ECMO')+
  xlab('')+
  scale_colour_manual('State', 
                      values=c('dark orange','blue'), 
                      labels=c('Supine','Prone'))+
  coord_flip()+
  theme_bw()+
  facet_wrap( ~ both_nice, ncol = 1, scales='free_y')
figure_nums(name="trajectories", "Plots of individual patients over time", level=2) # title
gplot
```

## Daily results for patients with a position change

Here we look at daily variables for patients who experience a change from supine to prone. 
Using within patient change helps control for confounders because patients act as their own control.
We examine the two weeks before and after this change. Not all patients have data that covers two weeks before and after the change. 

```{r}
# keep patients with both supine and prone
to_keep = filter(expand, state%in% c(1,2)) %>% # only prone and supine
  group_by(id) %>%
  summarise(diff = max(state) - min(state)) %>%
  filter(diff != 0) %>% # some change
  ungroup()
# add date information
small_patients = select(patients, pin, ecmo_start, prone_start, prone_end)
to_keep = left_join(to_keep, small_patients, by=c('id'='pin'))
n_diff = nrow(to_keep)  
# merge with daily data
daily_plot = left_join(to_keep, daily, by=c('id'='pin')) %>%
  mutate(days = date_daily - ecmo_start,
         diff = date_daily - prone_start,
         prone_effect = ifelse(diff<0, 0, diff)) %>% # change point at day zero
  filter(abs(diff) <= 14) %>% # keep to 2 weeks either side
  select(id, date_daily, ecmo_start, prone_start, diff, prone_effect, eotd_peep) %>%
  filter(!is.na(eotd_peep)) # remove results with missing outcome
# model within person change
model = lmer(eotd_peep ~ diff + prone_effect + (1|id), data=daily_plot) # just random intercept
newdata = data.frame(id=daily_plot$id[1], diff=seq(-14,14,1)) %>%
  mutate(prone_effect = ifelse(diff<0, 0, diff))
newdata$pred = predict(model, newdata=newdata, re.form=~0)
#
tplot = ggplot(data=daily_plot, aes(x=diff, y=eotd_peep, group=id))+
  geom_line(col=grey(0.5))+
  geom_line(data=newdata, aes(x=diff, y=pred), col='dark red')+ # add fitted regression line
  theme_bw()+
  xlab('Days since prone positioning start')
figure_nums(name="prone_change", "Individual patients results before and after the change to prone position (grey lines) and the estimated average change from a regression model (red line).") # title
tplot
# estimates and confidence intervals
s = as.data.frame(summary(model)$coefficients); s$var = rownames(s)
b = as.data.frame(confint(model)); b$var = rownames(b)
ests = left_join(s, b, by='var') %>%
  filter(var == 'prone_effect') %>%
  rename('lower' ='2.5 %',
         'upper' ='97.5 %')
```

There are `r n_diff` patients who move from prone to supine or vice versa, and there are `r nrow(daily_plot)` daily observations for these patients.

The plot shows the individual results for patients (grey lines) before and after their move to prone. The red line is the estimated mean change from a regression model that has a change point in the slope over time on the day that patients switched to prone. 

For PEEP there was little change after moving to prone position, with a change in the slope per day of `r roundz(ests$Estimate,2)` with a 95% confidence interval from `r roundz(ests$lower,2)` to `r roundz(ests$upper,2)`.

## Site effects

Random effects for sites from the survival models.

##### Site effects (prone position)

```{r, fig.width=5, fig.height=6}
# get site estimates
load('results/survival_prone.RData') # get estimates 
centres = bugs_extract(inbugs=bugs_results_prone, var='^centre', exp=TRUE, arrange=TRUE) %>% # from 99_functions.R
  mutate(over = as.numeric(lower>1) + 1) # for colouring estimates
# annotation
labels = data.frame(row=0, mean=1, lower=0, upper=0, over=1, label=c('Increased hazard','Decreased hazard'))
# plot
cplot = ggplot(data=centres, aes(x=row, y=mean, ymin=lower, ymax=upper, col=factor(over)))+
  geom_hline(lty=2, yintercept=1)+
  scale_color_manual(NULL, values=c('cyan4','dark red'))+
  geom_point()+
  scale_y_log10()+
  scale_x_continuous(breaks=NULL, labels = NULL, limits=c(0,M), expand=c(0.01,0.01))+
  geom_errorbar(width=0)+
  geom_text(data=labels, size=3, aes(x=row, y=mean, label=label), col='black', hjust=c(-0.1, 1.1))+
  xlab('')+
  ylab('Hazard ratio')+
  coord_flip()+
  theme_bw()+
  theme(legend.position = 'none',
        panel.grid.minor.y = element_blank(), 
        panel.grid.major.y = element_blank())
figure_nums(name="site_prone", "Site effects from the Weibull survival model for prone position", level=1) # title
cplot
jpeg('figures/prone/left_out_prone.jpg', width=4.5, height=5.5, units='in', res=500, quality=100)
print(cplot)
invisible(dev.off())
```

The plot shows the mean hazard ratio per site (circle) and 95% credible interval (horizontal line) for the `r n_sites` sites. The dotted vertical line is at a hazard ratio of 1 and indicated no change in hazard. Sites below the line have a decreased hazard of prone position, whilst sites above the line have an increased hazard. Sites with a lower limit above the line have been coloured red. Sites have been ordered using their mean hazard ratio. 
The wide credible intervals indicate large uncertainty in most sites, and this is because of the small number of patients in most sites. The hazard ratio is on a log scale (base 10).

There are `r sum(centres$over==2)` sites that have a relatively large hazard ratio for putting patients in prone position. This could be because prone positioning is an accepted practice at those sites.

##### Site effects (prone positioning): details on the influential sites

```{r}
# get centre numbers that look high
high = filter(centres, over==2) %>%
  mutate(
    text = row.names(.),
    num = as.numeric(str_remove_all(text, '[^0-9]'))) %>% # get centre number
  pull(num)
# add back to original data
for_prone_original = mutate(for_prone,
                   high = site_num %in% high)
# now get statistics per site
stats = group_by(for_prone_original, site_num, high) %>%
  summarise(n = n(),
            exit = mean(exit, na.rm=TRUE), # average time
            censored = sum(as.numeric(to=='cens')), # number censored
            prone = sum(as.numeric(to=='2'))) # movements to prone
ftab = filter(stats, high==TRUE) %>%
  select(-high) %>%
  flextable() %>%
  theme_box() %>%
  autofit()
ftab
```

The table above shows the summary statistics for the three sites that were relative outliers in terms of prone positioning.


##### Site effects (death and discharge)

```{r, fig.width=5, fig.height=6}
# get centre estimates
centres_death = bugs_extract(inbugs=bugs_results_death, var='^centre', exp=TRUE) %>% # from 99_functions.R
  mutate(type = 1)
# annotation
centres_discharge = bugs_extract(inbugs=bugs_results_discharge, var='^centre', exp=TRUE) %>% # from 99_functions.R
  mutate(type = 2)
# annotation
labels = data.frame(row=0, mean=1, lower=0, upper=0, label=c('Increased hazard','Decreased hazard'))
# combine estimates
centres = bind_rows(centres_death, centres_discharge) %>%
  group_by(row) %>%
  mutate(mean_hr = mean(mean)) %>% # calculate average HR
  ungroup() %>%
  mutate(order = as.numeric(as.factor(mean_hr)),
         order = ifelse(type==2, order+0.2, order))  # jitter
# plot
cplot = ggplot(data=centres, aes(x=order, y=mean, ymin=lower, ymax=upper, col=factor(type)))+
  geom_hline(lty=2, yintercept=1)+
  geom_point()+
  scale_y_log10()+
  scale_x_continuous(breaks=NULL, labels = NULL, limits=c(0,M+0.2), expand=c(0.01,0.01))+
  scale_color_manual(NULL, values=c('indianred','skyblue'), labels=c('Death','Discharge'))+
  geom_errorbar(width=0)+
  geom_text(data=labels, size=3, aes(x=row, y=mean, label=label), col='black', hjust=c(-0.1, 1.1))+
  xlab('')+
  ylab('Hazard ratio')+
  coord_flip()+
  theme_bw()+
  theme(panel.grid.minor.y = element_blank(), panel.grid.major.y = element_blank())
figure_nums(name="site_death", "Site effects from the Weibull survival model for death and discharge")
cplot
```

The hazard ratio axis is on log scale (base 10). Each row is a site and there are two estimates per site: one for death and one for discharge.

There was more variability between sites in the hazard of discharge than the hazard of death. This can be seen because the site-specific estimates for death (red dots) are generally closer to 1 than the estimates for discharge (blue dots).

## Estimates of the Weibull shape parameter

Here we show the estimates of the Weibull shape parameter from the parametric survival models. The table shows the mean and 95% credible interval.

```{r}
# get shape estimates
r_prone = bugs_extract(inbugs=bugs_results_prone, var=c('^r|tau.centre')) %>% # from 99_functions.R; added centre because 1 parameter does not work!
  filter(Var=='r')
r_death = bugs_extract(inbugs=bugs_results_death, var=c('^r|tau.centre')) %>% # from 99_functions.R; added centre because 1 parameter does not work!
  filter(Var=='r')
r_discharge = bugs_extract(inbugs=bugs_results_discharge, var=c('^r|tau.centre')) %>% # from 99_functions.R; added centre because 1 parameter does not work!
  filter(Var=='r')
r = bind_rows(r_prone, r_death, r_discharge, .id = 'model') %>%
  mutate(model = factor(model, levels=1:3, labels=c('Prone','Death','Discharge')),
         mean = roundz(mean, 1),
         CI = paste(roundz(lower, 1), ' to ', roundz(upper, 1), sep='')) %>%
  select(-Var, -lower, -upper) 
#
ftab = flextable(r) %>%
  theme_box() %>% # nice theme
  align(align='center', part='all') %>% # centre text
  autofit()
table_nums(name="table_r", "Means and 95% credible intervals for the Weibull shape parameters", level=1) # title, increase level
ftab
```

A shape parameter under 1 means the event rate decreases over time, whereas a shape parameter over 1 means the event rate increases with time.
The shape parameter is well below 1 for the prone model, as most patients experience prone early in their stay.
The shape parameter is well above 1 for the discharge model, as few patients are discharged in the first few days after starting ECMO.

## Baseline hazards

Plots of the estimated baseline hazard function from the Weibull survival models.
Say why this is useful. Cite Altman.

```{r}
# times to predict over:
x = c(0.5, 1:(censor_day/2)) # only need half of censor time
# winbugs estimates
r = bugs_results_prone$sims.list$r # from MCMC
mu = exp(bugs_results_prone$sims.list$int) # log-link
mat = matrix(data=0, nrow=length(r), ncol=length(x))
# alternative notation (from Wikipedia)
for (i in 1:length(r)){ # loop through MCMC
  mat[i,] = mu[i]*r[i]*(x^(r[i]-1))
}
f = data.frame(x=x, hazard = colMeans(mat), lower=apply(mat, 2, quantile, probs=0.025), upper=apply(mat, 2, quantile, probs=0.975))
hazplot = ggplot(f, aes(x=x, y=hazard, ymin=lower, ymax=upper))+
  geom_ribbon(alpha=0.2)+
  geom_line()+
  xlab('Days')+
  ylab('Hazard function')+
  theme_bw()
figure_nums(name="baseline_a", "Baseline hazard for survival model of prone position")
hazplot
```

The black line is the mean hazard function and the grey area is the 95% credible interval.
The hazard of prone position is highest in the first few days and then sharply declines. This matches the data, as most patients are prone on day zero, or get moved to prone soon after being on ECMO.


```{r}
# times to predict over:
x = c(0.5, 1:(censor_day/2)) # only need half of censor time
## a) discharge
# winbugs estimates
r = bugs_results_discharge$sims.list$r # from MCMC
mu = exp(bugs_results_discharge$sims.list$int) # log-link
mat = matrix(data=0, nrow=length(r), ncol=length(x))
# alternative notation (from Wikipedia)
for (i in 1:length(r)){ # loop through MCMC
  mat[i,] = mu[i]*r[i]*(x^(r[i]-1)) # hazard function
}
f1 = data.frame(type=2, x=x, hazard = colMeans(mat), lower=apply(mat, 2, quantile, probs=0.025), upper=apply(mat, 2, quantile, probs=0.975))
## b) death
# winbugs estimates
r = bugs_results_death$sims.list$r # from MCMC
mu = exp(bugs_results_death$sims.list$int) # log-link
mat = matrix(data=0, nrow=length(r), ncol=length(x))
# alternative notation (from Wikipedia)
for (i in 1:length(r)){ # loop through MCMC
  mat[i,] = mu[i]*r[i]*(x^(r[i]-1)) # hazard function
}
f2 = data.frame(type=1, x=x, hazard = colMeans(mat), lower=apply(mat, 2, quantile, probs=0.025), upper=apply(mat, 2, quantile, probs=0.975))
f = bind_rows(f1, f2) %>%
  mutate(type.nice = factor(type, levels=1:2, labels=c('Death','Discharge')))
hazplot = ggplot(f, aes(x=x, y=hazard, ymin=lower, ymax=upper))+
  geom_ribbon(alpha = 0.2)+
  geom_line()+
  scale_color_manual(NULL, values=c('indianred','skyblue'), labels=c('Death','Discharge'))+
  xlab('Days')+
  ylab('Hazard function')+
  theme_bw()+
  facet_wrap(~type.nice, scales='free_y')
figure_nums(name="baseline_b", "Baseline hazard for survival model of death and discharge")
hazplot
```

The hazard of discharge increases greatly over time. The hazard of death also increases over time, but with a far shallower slope.

### Alternative Cox survival models

Here we use non-parametric Cox models to confirm the results from the parametric Weibull survival models. These models use the multi-state approach, so patients could move from prone to supine (or vice versa) over time.
These models account for correlated results from the same site. We use a simple mean imputation for patients missing BMI.

The table shows the mean hazard ratios (HRs) and 95% confidence intervals for death and discharge.

```{r}
# use function to run Cox models (also used below in leave-one-out analysis)
run_cox_models = function(in_discharge, in_death, digits=2, cumulative_prone=FALSE){
  # convert lists to data frames
  if(class(in_discharge) == 'list'){in_discharge = as.data.frame(in_discharge)}
  if(class(in_death) == 'list'){in_death = as.data.frame(in_death)}
  # a) discharge
  in_discharge = mutate(in_discharge,
                        bmi = ifelse(is.na(bmi)==TRUE, 0, bmi), # simple imputation of mean
                        prone_before = ifelse(is.na(prone_before)==TRUE, 1, prone_before)) # simple imputation of mode
  # weibull - not used
  #m_disch = survreg(Surv(days, event=in_discharge$cens==0, type='right') ~ in_discharge$sex + in_discharge$age +in_discharge$calendar_time + bmi + #in_discharge$prone + cluster(in_discharge$site_num))
  # cox
  if(cumulative_prone == FALSE ){ # censored = 0 (FALSE) is the event
    in_discharge = mutate(in_discharge,
                          days = ifelse(is.na(days)==TRUE, cens, days)) # get days from censored or uncensored
    c_disch = coxph(Surv(days, event=cens==0, type='right') ~ sex + age + calendar_time + bmi + prone + prone_before + cluster(site), data=in_discharge)
  }
  if(cumulative_prone == TRUE){
    c_disch = coxph(Surv(time=entry, time2=exit, event= to == 3) ~ sex + age + calendar_time + bmi + prone + prone_cum + cluster(site_num), data=in_discharge)
  }
  ests_disch = tidy(c_disch, conf.int = TRUE, exponentiate = TRUE) %>%
    mutate(
      estimate = roundz(estimate, digits),
      conf.low = roundz(conf.low, digits),
      conf.high = roundz(conf.high, digits),
      CI = paste(conf.low, ' to ', conf.high, sep='')) %>%
    select(-term, -std.error, -statistic, -p.value) %>%
    rename('HR2' = 'estimate',
           'CI2' = 'CI',
           'conf.low2' = 'conf.low',
           'conf.high2' = 'conf.high')
  # b) death
  in_death = mutate(in_death,
                    bmi = ifelse(is.na(bmi)==TRUE, 0, bmi), # simple imputation of mean
                    prone_before = ifelse(is.na(prone_before)==TRUE, 1, prone_before)) # simple imputation of mode
  # weibull - not used
  #  m_death = survreg(Surv(days, event=in_death$cens==0, type='right') ~ bmi + in_death$age + in_death$sex + in_death$calendar_time + in_death$prone + cluster(in_death$site_num))
  # cox
  if(cumulative_prone == FALSE ){ # censored = 0 (FALSE) is the event
    in_death = mutate(in_death,
                      days = ifelse(is.na(days)==TRUE, cens, days)) # get days from censored or uncensored
    c_death = coxph(Surv(days, event=cens == 0, type='right') ~ bmi + age + sex + calendar_time + prone + prone_before + cluster(site), data=in_death)
  }
  if(cumulative_prone == TRUE ){
    c_death = coxph(Surv(time=entry, time2=exit, event=to == 4) ~ bmi + age + sex + calendar_time + prone + prone_cum + cluster(site_num), data=in_death)
  }
  ests_death = tidy(c_death, conf.int = TRUE, exponentiate = TRUE) %>%
    mutate(
      estimate = roundz(estimate, digits),
      conf.low = roundz(conf.low, digits),
      conf.high = roundz(conf.high, digits),
      CI = paste(conf.low, ' to ', conf.high, sep = '')) %>%
    select(-std.error, -statistic, -p.value) %>%
    rename('HR1' = 'estimate',
           'CI1' = 'CI',
           'conf.low1' = 'conf.low',
           'conf.high1' = 'conf.high')
  for_tab = bind_cols(ests_death, ests_disch) %>%
    mutate(term = str_remove(string=term, pattern='in_death\\$'))
  # return
  to.return = list()
  to.return$table = for_tab
  to.return$discharge_model = c_disch
  to.return$death_model = c_death
  to.return$death_data = in_death # add data
  to.return$discharge_data = in_discharge # 
  return(to.return) 
}
cox_models = run_cox_models(in_discharge = for_discharge, in_death = for_death) # run above function
for_tab = cox_models$table %>% 
  select(term, HR1, CI1, HR2, CI2) 
## nice table
header = read.table(header=TRUE, stringsAsFactors = FALSE, sep="!", text='
col_keys!colB!colA
term! !Variable
HR1!Death!HR
CI1!Death!95% CI
HR2!Discharge!HR
CI2!Discharge!95% CI
')
ftab = flextable(for_tab) %>%
  set_header_df(mapping = header, key = "col_keys") %>%
  merge_h(part = "header", i = 1) %>%
  theme_box() %>%# nice theme
  align(align='center', part='all') %>%# centre text
  autofit()
table_nums(name="hr_cox", "Hazard ratios and 95% credible intervals for death and discharge from the Cox model")
ftab 
```

Prone position reduces the hazard of death and discharge. 

#### Censoring times

```{r}
# run simple model of censoring times, recommended by https://arxiv.org/pdf/2005.13271.pdf
times_censor = survival::survfit(Surv(time=entry, time2=exit, event= to==0) ~ 1 + cluster(pin), data=my.data) # event is censored
# report median censoring time
g = ggsurvplot(times_censor, data = my.data,
               legend.labs=NULL,
               xlim=c(0,90),
               break.time.by=30,
                  conf.int =FALSE,
                  censor =FALSE) 
g
figure_nums(name="censor_times", "Kaplan-Meier plot of the censoring times.")
#
cens_stats = surv_median(times_censor)
```

The Kaplan-Meier plot above shows the censoring times. Patients could be censored prior to 90 days based on their last date of observation. Patients were censored at 90 days if they were still in hospital. The median day of censoring is `r cens_stats$median`.

#### Checking for influential patients

In this section we check whether there are influential patients in the Cox survival analysis. An influential patient is one who has a large effect of the estimates and hence any inference.

The plot below shows the 'dfbetas' which are the estimated changes in the regression coefficients after deleting each observation patient in turn. The estimates are on the log-scale and zero means no change. We have separate plots for the discharge and death models.

```{r}
figure_nums(name="inf_discharge", "Influential observations for the Cox survival model of discharge", level=2) # title
# test of influential observations
influence_discharge = ggcoxdiagnostics(fit=cox_models$discharge_model, sline=FALSE, type = 'dfbeta', linear.predictions = FALSE, ggtheme = theme_bw())
influence_discharge
```

There is one influential patient for prone for the model examining discharge. This is the one patient who was discharged in prone position. Without that patient the hazard of discharge is much greater in 

```{r}
# look at influential observation(s)
index = filter(influence_discharge$data, covariate=='prone', abs(res) >= 0.6) %>% pull(xval)
to_table = cox_models$discharge_data[index,] %>%
  select(site, sex, age, bmi, prone, days, cens) %>%
  mutate(
    cens = ifelse(cens==0, 'No', 'Yes'), # nicer labels for table
    prone = ifelse(prone==0, 'No', 'Yes'),
    sex = ifelse(sex==1, 'Male', 'Female'),
    days = round(days),
    age = (age*10)+50, # back-transform
    bmi = (bmi*5)+30,
    bmi =round(bmi*10)/10) %>%
  rename('censored' = 'cens')
ftab = flextable(to_table) %>%
  autofit() %>%
  theme_box()
ftab
```

```{r}
figure_nums(name="inf_death", "Influential observations for the Cox survival model of death", level=2) # title
# test of influential observation(s)
influence_death = ggcoxdiagnostics(fit=cox_models$death_model, sline=FALSE, type = 'dfbeta', linear.predictions = FALSE, ggtheme = theme_bw())
influence_death
```

There are some influential patients for prone for the model examining deaths, although the size is relatively small at around 0.1, meaning the estimated reduced risk of death would be larger without these patients. For comparison the estimate for prone with all the data is `r round(cox_models$death_model$coefficients[5]*100)/100`.

```{r}
# look at influential observation(s)
index = filter(influence_death$data, covariate=='prone', abs(res) >= 0.1) %>% pull(xval)
to_table = cox_models$death_data[index,] %>%
  select(site, sex, age, bmi, prone, days, cens) %>%
  mutate(
    cens = ifelse(cens==0, 'No', 'Yes'), # nicer labels for table
    prone = ifelse(prone==0, 'No', 'Yes'),
    sex = ifelse(sex==1, 'Male', 'Female'),
    days = round(days),
    age = (age*10)+50, # back-transform
    bmi = (bmi*5)+30,
    bmi =round(bmi*10)/10) %>%
  rename('censored' = 'cens')
ftab = flextable(to_table) %>%
  autofit() %>%
  theme_box()
ftab
```

These are all patients in the prone position who died in relatively short times ("days" column).

#### Check of proportional hazards assumption for Cox models

```{r}
prop_disch = cox.zph(run_cox_models(in_discharge = for_discharge, in_death = for_death)$discharge_model)
prop_death = cox.zph(run_cox_models(in_discharge = for_death, in_death = for_death)$death_model)
for_table = bind_rows(as.data.frame(prop_disch$table), as.data.frame(prop_death$table), .id='outcome') %>%
  mutate(Variable = row.names(.),
         Variable = str_remove_all(string=Variable, pattern='in_discharge\\$|\\.|[0-9]'), # tidy names
         chisq = round(chisq*100, 0)/100,
         p = format.pval(p, eps=0.001, digits=3),
         outcome = factor(as.numeric(outcome), levels=1:2, labels=c("Discharge",'Death'))) %>%
  select(outcome, Variable, everything())
ftab = flextable(for_table) %>%
  merge_v(j=1) %>% # 
  theme_box() %>%
  autofit()
table_nums(name="ph_check", "Check of proportional hazards assumption for Cox models")
ftab
```

The table above shows no clear concern about the proportional hazards assumption using a Cox model. 
The only variable that may be worth investigating is the effect of age on discharge.

### Leave one country out sensitivity analysis

Here we leave one country out in turn and check that the estimates for prone remain similar. The reasoning is to check that the findings for prone are not reliant on one country. We use the Cox survival models.

```{r, fig.width=7}
# get country list
tab = table(my.data$country)
countries = names(tab[tab>1]) # just for countries with more than one patient
# loop through countries
all_ests = NULL
for (this_country in countries){
  data.to.use = filter(my.data, country != this_country)
  for_death = prepare_survival(indata=data.to.use, cens_events=c('1','2','3','cens')) # supine, prone, discharge or admin censored; event is death ('4')
  for_discharge = prepare_survival(indata=data.to.use, cens_events=c('1','2','4','cens')) # supine, prone, death or admin censored; event is discharge ('3')
  ests = run_cox_models(in_discharge = for_discharge, in_death = for_death, digits=10)$table %>% # run above function to run Cox models
    filter(str_detect(term, 'prone$')) %>% # not prone before
    select('HR1','conf.low1','conf.high1','HR2','conf.low2','conf.high2') %>%
    mutate_all(~as.numeric(.x)) %>% # change rounded numbers back to numbers
    mutate(country = this_country)
  all_ests = bind_rows(all_ests, ests)
}
# plot
leave_plot1 = ggplot(data=all_ests, aes(x=country, y=HR1, ymin=conf.low1, ymax=conf.high1))+
  ggtitle('Death')+
  geom_point()+
  geom_errorbar(width=0)+
  theme_bw()+
  xlab('')+
  ylab('Hazard ratio')+
  coord_flip()
leave_plot2 = ggplot(data=all_ests, aes(x=country, y=HR2, ymin=conf.low2, ymax=conf.high2))+
  ggtitle('Discharge')+
  geom_point(col='blue')+
  geom_errorbar(width=0, col='blue')+
  theme_bw()+
  xlab('')+
  ylab('Hazard ratio')+
  coord_flip()
figure_nums(name="leave_country", "Hazard ratios and 95% confidence intervals for death and discharge after leaving out each country") # title
grid.arrange(leave_plot1, leave_plot2, ncol=2)
```

The plot shows the mean hazard ratio (dot) and 95% confidence interval.
The estimates change when the USA is left out. This is because the one very influential patient is from the USA.

## Convergence of Bayesian models

Here we visually check the convergence of the estimates for the Bayesian survival models. We plot the intercept (int), regression parameters (beta 1 to 5), and Weibull shape parameter (r).

```{r}
# plot r, beta and intercept
to_plot = c('^r$','^beta','^int')
index = str_detect(pattern=paste(to_plot, collapse='|'), string=dimnames(bugs_results_death$sims.array)[[3]])
chain1 = reshape2::melt(bugs_results_death$sims.array[,1,index])
chain2 = reshape2::melt(bugs_results_death$sims.array[,2,index])
chains = bind_rows(chain1, chain2, .id='chain')
chain_plot = ggplot(data=chains, aes(x=Var1, y=value, col=factor(chain)))+
  geom_line()+
  scale_color_manual('Chain', values=c(c("#FFB90F", "#00B2EE")))+
  xlab('')+
  ylab('')+
  theme_bw()+
  facet_wrap(~Var2, scales='free_y')+
  theme(legend.position = c(0.8,0.1))
figure_nums(name="chains_death", "Markov Chain Monte Carlo estimates for the Weibull survival model of death", level=2) # title
chain_plot
```

```{r}
# plot r, beta and intercept
index = str_detect(pattern=paste(to_plot, collapse='|'), string=dimnames(bugs_results_discharge$sims.array)[[3]])
chain1 = reshape2::melt(bugs_results_discharge$sims.array[,1,index])
chain2 = reshape2::melt(bugs_results_discharge$sims.array[,2,index])
chains = bind_rows(chain1, chain2, .id='chain')
chain_plot = ggplot(data=chains, aes(x=Var1, y=value, col=factor(chain)))+
  geom_line()+
  scale_color_manual('Chain', values=c(c("#FFB90F", "#00B2EE")))+
  xlab('')+
  ylab('')+
  theme_bw()+
  facet_wrap(~Var2, scales='free_y')+
  theme(legend.position = c(0.8,0.1))
figure_nums(name="chains_discharge", "Markov Chain Monte Carlo estimates for the Weibull survival model of discharge", level=2) # title
chain_plot
```

#### R version and packages used

This information can be useful for reproducibility purposes.

```{r}
sessionInfo()
```

#### References
