==> 1_obese.Rmd <==
---
title: "COVID and obesity in ECMO patients"
author: "Adrian Barnett"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  word_document
link-citations: yes
bibliography: bibs.bib
---
  
```{r, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, warning=FALSE, message=FALSE, error=FALSE, comment='', dpi=400)
options(width=1000) # Wide pages
options(scipen=999) # avoid scientific presentation

## libraries and functions
library(mice) # for making imputed data
library(mitools) # for summarising imputations
library(dplyr)
options(dplyr.summarise.inform = FALSE)
library(tidyr)
#library(mstate)
library(stringr)
library(flextable)
library(broom)
library(mstate)
library(cmprsk)
source('crr-addson.R') # additional functions for crr, see https://www.nature.com/articles/bmt2009359
library(survminer) # for nice plots of competing risks
source('99_functions.R')
source("Martin/ext_mstate.R")
# for captions:
library(captioner)
table_nums <- captioner(prefix = "Table", levels=2, type = c("n", "c")) # number and character
figure_nums <- captioner(prefix = "Figure", levels=2, type = c("n", "c"))
library(diagram) # for state diagrams

# maximum day for plots and estimates
censor_day = 90

## Questions:
# "The key question in my mind was does morbid obesity impact survival in patients on ecmo with covid 19."
# How do obese people do on VV ECMO?

## prepare CCC set (with id, from, to, entry, exit; from and to are state numbers)
load('data/COVID_Data_obese.RData') # 0_read_data.R

## create BMI category using threshold from Jeff
# have to do before exclusions as it is needed for MV patients too
patients = mutate(patients, 
        # if no final outcome then unknown:
        outcome = ifelse(outcome=='', 'Unknown', outcome), 
        # if no death date then switch to unknown:
        outcome = ifelse(outcome=='Death' & is.na(date_death), 'Unknown', outcome),
        # prone
        any_prone = as.numeric(is.na(prone_start)==FALSE), # have a prone start date
        any_prone = case_when(
          any_prone == 0 ~ 'No',
          any_prone == 1 ~ 'Yes'
        ),
        # pregnant
        pregnant = case_when(
          pregnant == 0 ~ 'No',
          pregnant == 1 ~ 'Yes',
          pregnant == 2 ~ 'Unknown/missing',
          is.na(pregnant) == TRUE ~ 'Unknown/missing'
        ),
        # create apache category
        apache_cat = case_when(
          apache_ii <= median(patients$apache_ii, na.rm=TRUE) ~ "Low",
          apache_ii > median(patients$apache_ii, na.rm=TRUE) ~ "High",
          is.na(apache_ii) ~ "Missing"
        ),
        bmi_cat1 = case_when(
          bmi <= 35 ~ "Low (<= 35)",
          bmi > 35 ~ "High (> 35)",
          is.na(bmi) ~ "Missing"
        ), # and alternative cut-off
        bmi_cat = case_when(
          bmi <= 35 ~ "Low (<= 35)",
          bmi > 35 ~ "High (> 35)",
          is.na(bmi) ~ "Missing"
        ),
        )

## impute BMI using MICE
source('1_obese_impute.R') 

## Mechanically Ventilated patients (used later)
mv_patients = filter(patients, 
                  age >= 18, # adults only
                  sex %in% c('Female', 'Male'),
                  #bmi_cat != 'Missing', # no missing BMI - keep for imputation
                  !is.na(date_icu), # must be in ICU
                  !is.na(date_mechanical_ventilation)) %>% # must have started MV
  mutate(ecmo_group = as.numeric(ecmo_type=='Venous-Venous'),
         ecmo_group = factor(ecmo_group, levels=0:1, labels=c('Non-ECMO','ECMO')))
n_mv_patients = nrow(mv_patients)

## apply inclusions for the obese patients
patients = filter(patients, 
                  age >= 18, # adults only
                  #bmi_cat != 'Missing', # no missing BMI
                  ecmo_type == 'Venous-Venous', # just VV ECMO
      sex %in% c('Female', 'Male'))
n_total = nrow(patients)
patients = filter(patients,
      !is.na(date_icu), # must be in ICU
      !is.na(date_ecmo)) # must have an ECMO start date
n_patients = nrow(patients)

## Now add imputed BMI to patient data
patients_imputed = mv_patients_imputed = list()
for (k in 1:N.impute){
  complete = complete(imputed_data, k) %>%
    select('pin','bmi')
  patients_imputed[[k]] = select(patients, -bmi) %>% # remove BMI
    left_join(complete, by='pin') # add imputed BMI
  mv_patients_imputed[[k]] = select(mv_patients, -bmi) %>% # remove BMI
    left_join(complete, by='pin') # add imputed BMI
}
```

This report uses the data available on `r data_date$month` `r data_date$day`.

```{r, include=FALSE}
# make the transition data plus imputed versions
source('1_obese_transitions.R')
```

We examined adult patients (aged 18+) who experienced Venous-Venous ECMO during their stay. For some analysis we also included all adult patients from the date they began mechanical ventilation.

For the analysis of ECMO only patients, we excluded patients with insufficient information on the dates of ECMO, discharge and death. The total number of patients excluded for this reason was `r length(excluded)`.

For the analysis of ECMO and MV patients, we excluded patients with insufficient information on the dates of MV, discharge and death. The total number of patients excluded for this reason was `r length(mv_excluded)`. We excluded 31 patients with missing BMI.

The total number of patients for the ECMO only patients is `r n_patients`. The data come from `r length(unique(patients$site_name))` sites.

The total number of patients for the ECMO and MV patients is `r n_mv_patients`. The data come from `r length(unique(mv_patients$site_name))` sites.

```{r, include=FALSE}
first_date = format(min(patients$date_admission), '%d-%b-%Y')
last_date = format(max(patients$date_admission), '%d-%b-%Y')
```

The first patient was admitted on `r first_date` and the last patient on `r last_date`.

# Statistical methods section

We used survival analysis to examine both the time on ECMO or mechanical ventilation, and the patients' final outcome (death or discharge) whilst accounting for censoring (patients transferred to another hospital or still in hospital at the end of data collection). As death and discharge are competing risks, we cannot use the standard approach of a Kaplan-Meier plot, but instead plot cumulative probability curves over time.
A guide to competing risks models for clinicians is available [@Scrucca2010].

Our research questions examined the impact of body mass index on patient outcomes. We examined whether BMI was associated with the risk of discharge or death, and this analysis was done in two groups: i) ECMO patients only, ii) ECMO and MV patients. We also examined if BMI was associated with being put on ECMO.

In summary plots we split BMI into low (below 35&nbsp;kg/m^2^) and high (above 35&nbsp;kg/m^2^) as this allowed us to easily compare groups. However, in the survival models we kept BMI as continuous as this increased our statistical power to find an association between BMI and death or discharge [@Greenland1995]. We scaled BMI to a per 5&nbsp;kg/m^2^ increase in order to show the risks for a meaningful change in BMI. Similarly we scaled age to a 10&nbsp;year increase. 

```{r, include=FALSE}
thresholds = seq(25, 40, 1) # BMI thresholds to test ... (used in text below but mainly later)
```

We tested if the risk of BMI had a threshold. We fitted multiple separate survival models thresholds from `r min(thresholds)` to `r max(thresholds)`&nbsp;kg/m^2^ and selected the best model using the Bayesian information criteria (BIC). The BIC examines both model fit and complexity and aims to select a parsimonious model. 

We randomly imputed the small number of missing BMI (number missing = `r sum(is.na(trans.data$bmi))` ECMO patients only, `r sum(is.na(trans.data.mv$bmi))` ECMO and MV patients).
We used Multivariate Imputation by Chained Equations (MICE) to create five imputed data sets [@Buuren2011]. BMI was imputed using the variables: country, ethnicity, height, weight, age, sex, APACHE II and SOFA.
Separate surivival models were run for each of the five imputed data sets and then combined. The results therefore hopefully account for the small amount of missing BMI data.

All analyses were made using R version 3.0.4 [@R]. The survival models were fitted using the "cmprsk" and "survival" packages [@cmprsk;@survival]. The results of the survival models are presented as hazard ratios (HRs) with 95% confidence intervals (CIs).

### Missing data

For some descriptive statistics we split BMI into high and low using a threshold of 35&nbsp;kg/m^2^. Because of the missing BMI data we keep a third category of missing. So the categorical BMI variable has three categories of low, high and missing. 

For SOFA and APACHE II score we split these variables into the three categories of:

* Low, below median
* High, above median
* Missing

The table below is the cross-tabulation of categorical BMI and APACHE II.

```{r}
tab = group_by(patients, bmi_cat, apache_cat) %>%
  summarise(count = n()) %>% 
  ungroup() %>%
  tidyr::spread(apache_cat, count)  %>%
  rename('BMI category' = 'bmi_cat')
# nice table
typology = data.frame(col_keys=c('bmi_cat','High','Low','Missing'),
                      colA = c('','APACHE','APACHE','APACHE'),
                      colB = c('BMI','High','Low','Missing'))
ftab = flextable(tab) %>%
   set_header_df(mapping = typology, key = "col_keys" ) %>%
   merge_h(part = "header", i=1) %>%
   theme_box() %>%
   align(align='center', part='all') %>%
  autofit()
table_nums(name="cross_tab", "Cross-tabulation of the categorical BMI and APACHE variables") # title
ftab
```

#### Summary diagram

The diagram below summarises the three research questions:

* Q1 = Question 1 uses the transitions from ECMO to death and discharge. Does the risk of death and discharge depend on BMI for ECMO patients?

* Q2 = Question 2 uses the transitions from MV to ECMO, death and discharge. Are there differences in the patient characteristics for patients who get put on ECMO compared with those who are discharged or die before being put on ECMO.

* Q3 = Question 3 uses the transitions from ECMO to death and discharge, and from MV to death and discharge. Does the association between BMI and death or discharge differ for patients on MV and ECMO?


```{r, fig.width=6.5, fig.height=4.5}
states = c('MV start','ECMO start','Death','Discharge\nfrom hospital')
# Matrix of joins between states:
M = matrix(data=c(
  2,1,
  3,1,
  4,1,
  3,2,
  4,2
), byrow = TRUE, ncol=2)
labels = c('Q2','Q2+Q3','Q2+Q3','Q1+Q3','Q1+Q3')
figure_nums(name="model_diagram_summary", "The four patient states and the three research questions") # title, increase title number
multistate_diagram(states=states, labels=labels, links = M, pos = c(1, 2, 1), arr.pos=0.65, box.size = 0.14) # see 99_functions.R
```


# Summary tables and plots (ECMO patients only)

#### a) Continuous variables (ECMO patients only)

```{r}
#
cont_vars = c('age','bmi','sofa','apache_ii')

# to here, add totals
selection = select(patients, pin, bmi_cat, all_of(cont_vars)) %>%
  tidyr::gather(key='Variable', value='Value', -pin, -bmi_cat)
tab = group_by(selection, Variable, bmi_cat) %>%
  summarise(Non_missing = sum(!is.na(Value)),
            Missing = sum(is.na(Value)),
            Median = roundz(median(Value, na.rm=TRUE),0),
            Q1 = roundz(quantile(Value, probs=0.25, na.rm=TRUE),0),
            Q3 = roundz(quantile(Value, probs=0.75, na.rm=TRUE),0),
            IQR = paste(Q1 , ' to ', Q3, sep='')) %>%
  select(Variable, bmi_cat, Non_missing, Missing, Median, IQR)  %>%
  rename('BMI category' = 'bmi_cat')
# total
totals = group_by(selection, Variable) %>%
  summarise(Non_missing = sum(!is.na(Value)),
            Missing = sum(is.na(Value)),
            Median = roundz(median(Value, na.rm=TRUE),0),
            Q1 = roundz(quantile(Value, probs=0.25, na.rm=TRUE),0),
            Q3 = roundz(quantile(Value, probs=0.75, na.rm=TRUE),0),
            IQR = paste(Q1 , ' to ', Q3, sep='')) %>%
  mutate(bmi_cat  = 'Total') %>%
  select(Variable, bmi_cat, Non_missing, Missing, Median, IQR)  %>%
  rename('BMI category' = 'bmi_cat')
# combine
tab = bind_rows(tab, totals) %>%
  arrange(Variable, 'BMI category')
#  pivot_wider(names_from = bmi_cat, values_from = c(Non_missing, Missing, Median, IQR)) 
#, names_glue = "{bmi_cat}_{.value}", names_sort=FALSE
ftab = flextable(tab) %>%
  theme_box() %>%
  merge_v(j=1) %>%
  autofit()
table_nums(name="summary_cont", "Summary statistics for continuous variables") # title
ftab
```

The above results for BMI as without imputing missing BMI.

IQR = inter-quartile range.

```{r, fig.width=6}
# boxplots
to_plot = select(patients, pin, bmi_cat, cont_vars) %>%
  pivot_longer(!c(pin, bmi_cat), names_to = "variable", values_to = "value") %>%
  filter(!is.na(value),
         bmi_cat != 'Missing') # do not include missing in plot
bplot = ggplot(data=to_plot, aes(x=value, col=bmi_cat))+
  geom_boxplot(fill='grey95')+
  xlab('')+
  scale_colour_manual('BMI category', values=c('goldenrod2','darkseagreen3'))+
  scale_y_continuous(breaks=NULL)+
  facet_wrap(~variable, scales='free')+
  theme_bw()+
  theme(panel.grid.minor = element_blank())
figure_nums(name="boxplots", "Boxplots of continuous variables by Body Mass Index category") # title
bplot
```

#### b) Categorical variables (ECMO patients only)

```{r}
# dropped ,'comorbidity_obesity'
cat_vars = c('sex','country','ethnicity','bmi_cat','apache_cat','any_prone')
#
tab = select(patients, pin, bmi_cat, all_of(cat_vars)) %>%
  tidyr::gather(key='Variable', value='Value', -pin, -bmi_cat) %>%
  group_by(Variable, bmi_cat, Value) %>%
  tally() %>%
  group_by(Variable, bmi_cat) %>%
  mutate(percent = roundz(prop.table(n)*100,0)) %>% # calculate percent
  arrange(Variable, bmi_cat, -n) %>% # arrange from high to low within variables
  mutate(cell = paste(n, ' (', percent, ')', sep='')) %>%
  select(-n, -percent) %>%
  pivot_wider(values_from = cell, names_from=bmi_cat, values_fill='0')
ftab = flextable(tab) %>%
  merge_v(j = c("Variable") ) %>% # merge rows
  theme_box() %>%
  autofit()
table_nums(name="summary_cat", "Summary statistics for categorical variables") # title
ftab
```

#### c) Pregancy (ECMO female patients under 50 only)

```{r}
#
tab = filter(patients, sex == 'Female', age < 50) %>%
  select(pin, bmi_cat, pregnant) %>%
  tidyr::gather(key='Variable', value='Value', -pin, -bmi_cat) %>%
  group_by(Variable, bmi_cat, Value) %>%
  tally() %>%
  group_by(Variable, bmi_cat) %>%
  mutate(percent = roundz(prop.table(n)*100,0)) %>% # calculate percent
  arrange(Variable, bmi_cat, -n) %>% # arrange from high to low within variables
  mutate(cell = paste(n, ' (', percent, ')', sep='')) %>%
  select(-n, -percent) %>%
  pivot_wider(values_from = cell, names_from=bmi_cat, values_fill='0')
ftab = flextable(tab) %>%
  merge_v(j = c("Variable") ) %>% # merge rows
  theme_box() %>%
  autofit()
table_nums(name="pregnant_cat", "Summary table for pregnancy") # title
ftab
```

#### d) Co-morbidities by low / high BMI (ECMO patients only)

```{r}
to_table = select(patients, pin, bmi_cat, starts_with('comorb')) %>%
  select(-comorbidity_other_risk_factors, -comorbidity_hba1c_units, -comorbidity_diabetes_mellitus_hba1c, -comorbidity_immuno, -comorbidity_aids_hiv_cd4_count) %>% # remove characters or those with just missing
  pivot_longer(!c(pin,bmi_cat), names_to='Comorbidity') %>%
  mutate(Value = case_when(
    value == 0 ~ 'No',
    value == 1 ~ 'Yes',
    is.na(value) == TRUE ~ 'Missing'
  )) %>%
  group_by(Comorbidity, bmi_cat, Value) %>%
  tally() %>%
  group_by(Comorbidity, bmi_cat) %>%
  mutate(percent = round(prop.table(n)*100)) %>%
  ungroup() %>%
  mutate(Comorbidity = str_remove_all(Comorbidity, pattern='^comorbidity_'),
         Comorbidity = str_replace_all(Comorbidity, pattern='_', replacement=' '),
         cell = paste(n, ' (', percent, ')', sep='')) %>%
  select(-n, -percent) %>%
  pivot_wider(values_from = cell, names_from=bmi_cat, values_fill='0')
ftab = flextable(to_table) %>%
  merge_v(j = c("Comorbidity") ) %>% # merge rows
  theme_box() %>%
  autofit()
table_nums(name="comorb", "Counts and percentages of comorbidities by BMI category") # title
ftab
```

#### e) Ventilation settings (ECMO patients only)

On the day of ECMO or the day after for those with missing data on the day of ECMO.

```{r}
#FiO2
#PEEP, cm H2O
#Static compliance, mL/cm H2O
#PaO2/FiO2
#PCO2
# make reduced daily data
daily_vars = c('pa_o2_fi_o2','pa_o2','pa_co2','p_h','eotd_peep','compliance_respiratory_system')
this_daily = select(daily, pin, date_daily, all_of(daily_vars))
# show settings on date of ECMO initiation or day after
date_e = select(patients, pin, bmi_cat, date_ecmo) # day of ecmo
date_e1 = mutate(date_e, date_ecmo = date_ecmo+1) # day after
all_dates = bind_rows(date_e, date_e1, .id='day') # bind two dates
# merge dates with daily
full_data = left_join(all_dates, this_daily, by=c('pin'='pin', 'date_ecmo'='date_daily')) %>% 
  pivot_longer(-c(pin, date_ecmo, bmi_cat, day), values_to='Value', names_to='Variable') %>%
  filter(!is.na(Value)) %>%
  group_by(pin, Variable) %>%
  arrange(pin, Variable, day) %>% # sort by date
  slice(1) # take first non-missing per patient and variable
# now do summary stats
tab = filter(full_data, Variable !='p_h') %>%
  group_by(Variable, bmi_cat) %>%
  summarise(N = sum(!is.na(Value)),
            Median = roundz(median(Value, na.rm=TRUE),0),
            Q1 = roundz(quantile(Value, probs=0.25, na.rm=TRUE),0),
            Q3 = roundz(quantile(Value, probs=0.75, na.rm=TRUE),0),
            IQR = paste(Q1 , ' to ', Q3, sep='')) %>%
  select(Variable, bmi_cat, N, Median, IQR)  %>%
  rename('BMI category' = 'bmi_cat')
# add totals
totals = filter(full_data, Variable !='p_h') %>%
  group_by(Variable) %>%
  summarise(N = sum(!is.na(Value)),
            Median = roundz(median(Value, na.rm=TRUE),0),
            Q1 = roundz(quantile(Value, probs=0.25, na.rm=TRUE),0),
            Q3 = roundz(quantile(Value, probs=0.75, na.rm=TRUE),0),
            IQR = paste(Q1 , ' to ', Q3, sep='')) %>%
  mutate(bmi_cat = 'Total') %>%
  select(Variable, bmi_cat, N, Median, IQR)  %>%
  rename('BMI category' = 'bmi_cat')
tab = bind_rows(tab, totals) %>%
  arrange(Variable, 'BMI category')
ftab = flextable(tab) %>%
  theme_box() %>%
  merge_v(j=1) %>%
  autofit()
table_nums(name="settings", "Ventilation settings") # title
ftab

## just ph with different dp
tab = filter(full_data, Variable =='p_h') %>%
  group_by(Variable, bmi_cat) %>%
  summarise(N = sum(!is.na(Value)),
            Median = roundz(median(Value, na.rm=TRUE),2),
            Q1 = roundz(quantile(Value, probs=0.25, na.rm=TRUE),2),
            Q3 = roundz(quantile(Value, probs=0.75, na.rm=TRUE),2),
            IQR = paste(Q1 , ' to ', Q3, sep='')) %>%
  select(Variable, bmi_cat, N, Median, IQR)  %>%
  rename('BMI category' = 'bmi_cat')
# add totals
totals = filter(full_data, Variable =='p_h') %>%
  group_by(Variable) %>%
  summarise(N = sum(!is.na(Value)),
            Median = roundz(median(Value, na.rm=TRUE),0),
            Q1 = roundz(quantile(Value, probs=0.25, na.rm=TRUE),0),
            Q3 = roundz(quantile(Value, probs=0.75, na.rm=TRUE),0),
            IQR = paste(Q1 , ' to ', Q3, sep='')) %>%
  mutate(bmi_cat = 'Total') %>%
  select(Variable, bmi_cat, N, Median, IQR)  %>%
  rename('BMI category' = 'bmi_cat')
tab = bind_rows(tab, totals) %>%
  arrange(Variable, 'BMI category')
ftab = flextable(tab) %>%
  theme_box() %>%
  merge_v(j=1) %>%
  autofit()
table_nums(name="settings2", "Ventilation settings (ph)") # title
ftab

```


#### f) Time variables by low / high BMI (ECMO patients only)

```{r, time_table}
# create time variables
patients = mutate(patients,
                  # calculate end dates
      date_end = ifelse(is.na(date_hospital_discharge)==FALSE, date_hospital_discharge, date_death), # end date of death or discharge
      date_end = as.Date(date_end, origin='1970-01-01'),
      #
      date_end_icu = ifelse(is.na(date_discharge)==FALSE, date_discharge, date_death), # end date of death or discharge
      date_end_icu = as.Date(date_end_icu, origin='1970-01-01'),
      #
      days_hosp = as.numeric(date_end - date_admission) + 0.5,
      days_icu = as.numeric(date_end_icu - date_icu) + 0.5,
      days_symptoms_hosp = as.numeric(date_admission - date_first_symptom) +0.5, 
      #
      duration_ecmo = as.numeric(date_ecmo_discontinued - date_ecmo) + 0.5,
      duration_mv = as.numeric(date_mech_vent_discontinued - date_mechanical_ventilation +0.5),
# new (Jan 2021)
      days_admission_mv = as.numeric(date_mechanical_ventilation - date_admission) + 0.5,
days_admission_ecmo = as.numeric(date_ecmo - date_admission) + 0.5)

#
time_vars = c('days_symptoms_hosp','days_hosp','days_icu','duration_ecmo','duration_mv','days_admission_mv','days_admission_ecmo')
#
selected = select(patients, pin, bmi_cat, all_of(time_vars)) %>%
   pivot_longer(!c(pin, bmi_cat), names_to='Variable', values_to='Value')
#
stats = group_by(selected, Variable, bmi_cat) %>%
  summarise(Non_missing = sum(!is.na(Value)),
            Missing = sum(is.na(Value)),
            Median = roundz(median(Value, na.rm=TRUE),0),
            Q1 = roundz(quantile(Value, probs=0.25, na.rm=TRUE),0),
            Q3 = roundz(quantile(Value, probs=0.75, na.rm=TRUE),0),
            IQR = paste(Q1 , ' to ', Q3, sep='')) %>%
  select(Variable, bmi_cat, Non_missing, Missing, Median, IQR) %>%
  mutate(Variable = case_when(
    Variable=='duration_ecmo'~ 'Duration of ECMO',
    Variable=='duration_mv'~ 'Duration of MV', 
    Variable=='days_hosp'~ 'Days in hospital', 
    Variable=='days_icu'~ 'Days in ICU',
    Variable=='days_admission_mv'~ 'Days from hospital admission to MV', 
    Variable=='days_admission_ecmo'~ 'Days from hospital admission to ECMO', 
    Variable=='days_symptoms_hosp'~ 'Days between symptoms and hospital admission'
#    TRUE ~ as.character(x)
    )) %>%
  rename('BMI category' = 'bmi_cat')
# add totals
totals = group_by(selected, Variable) %>%
  summarise(Non_missing = sum(!is.na(Value)),
            Missing = sum(is.na(Value)),
            Median = roundz(median(Value, na.rm=TRUE),0),
            Q1 = roundz(quantile(Value, probs=0.25, na.rm=TRUE),0),
            Q3 = roundz(quantile(Value, probs=0.75, na.rm=TRUE),0),
            IQR = paste(Q1 , ' to ', Q3, sep='')) %>%
  mutate(
    bmi_cat = 'Total',
    Variable = case_when(
    Variable=='duration_ecmo'~ 'Duration of ECMO',
    Variable=='duration_mv'~ 'Duration of MV', 
    Variable=='days_hosp'~ 'Days in hospital', 
    Variable=='days_icu'~ 'Days in ICU',
    Variable=='days_admission_mv'~ 'Days from hospital admission to MV', 
    Variable=='days_admission_ecmo'~ 'Days from hospital admission to ECMO', 
    Variable=='days_symptoms_hosp'~ 'Days between symptoms and hospital admission'
#    TRUE ~ as.character(x)
    )) %>%
  select(Variable, bmi_cat, Non_missing, Missing, Median, IQR) %>%
  rename('BMI category' = 'bmi_cat')
tab = bind_rows(stats, totals) %>%
  arrange(Variable, 'BMI category')
#
ftab = flextable(tab) %>%
  theme_box() %>%
  fontsize(size=10, part='all') %>%
  merge_v(j=1) %>%
  autofit()
table_nums(name="summary_time", "Summary statistics for time variables (in days)") # title
ftab
```


#### g) ECMO times for zero vs non-zero days (ECMO patients only)

```{r, ecmo_days_plot}
dplot = ggplot(data=patients, aes(x=days_admission_ecmo))+
  geom_bar(fill='darkseagreen2', col='grey', width=1)+
  xlab('Days from hospital admission to ECMO')+
  ylab('Frequency')+
  scale_x_continuous(limits=c(-7,21))+ # remove outliers
  theme_bw()
dplot
```


# Survival models (ECMO patients only)

## Model diagram

The figure below shows the three states in the multi-state survival model. We followed patients from the date then began on ECMO. Death and discharge are competing risks. Patients were censored if they did not have a date of death or discharge. We censored them using their last date of daily data or the last day data were collected (`r paste(data_date$day, data_date$month, sep=' ')`).

```{r, fig.width=4, fig.height=3}
states = c('ECMO','Death','Discharge\nfrom hospital')
# Matrix of joins between states:
M = matrix(data=c(
  2,1,
  3,1
), byrow = TRUE, ncol=2)
figure_nums(name="model_diagram", "The three states in the model and the possible movements between states", level=1) # title, increase title number
multistate_diagram(states=states, links = M, pos = c(1, 2), arr.pos=0.65, box.size = 0.14) # see 99_functions.R
jpeg(filename='figures/obese/state_diagram.jpg', width=3, height=3, units='in', res=400, quality = 100)
multistate_diagram(states=states, links = M, pos = c(1, 2), arr.pos=0.65, box.size = 0.14) # see 99_functions.R
invisible(dev.off())
```

### Table of transitions between states (ECMO patients only)

This table below shows the transitions (movements) between the states.

```{r}
tab = as.data.frame(with(trans.data, table(from, to))) %>%
  mutate(from = case_when(
    from==1 ~ 'ECMO')
  ) %>%
  tidyr::spread(to, Freq)
# table header
table_header <- data.frame(
  col_keys = c('from','0', '2','3'),
  h1 = c('', rep("State moved to", 3)),
  h2 = c('State moved from', 'Censored','Discharge','Death'),
  stringsAsFactors = FALSE )
ftab = flextable(tab) %>%
  set_header_df(mapping = table_header, key = "col_keys" ) %>% # add header
  merge_h(i=1, part = "header") %>% # merge "To" column headers
  align(align='center', part='all') %>% # centre text
  theme_box() %>%
  autofit() 
table_nums(name="transitions", "Transition numbers between states", level=1) # title, increase level number
ftab
```

## Cumulative probability curves (ECMO patients only)

```{r, fig.width=7}
# cumulative incidence
trans.data = mutate(trans.data,
                 bmi_cat_num = as.numeric(as.factor(bmi_cat))) # need simply category for cuminc
cum_inc = with(trans.data, cuminc(ftime=exit, fstatus=to, group=bmi_cat_num, cencode='0'))
g = ggcompetingrisks(cum_inc)
# make new plot
bmi_labels = c("High (> 35)","Low (<= 35)" ,"Missing")
for_plot = mutate(g$data, 
                  event_factor = factor(event, levels=2:3, labels=c('Discharge','Death')),
                 bmi_cat_factor = factor(group, levels=1:3, labels=bmi_labels)) %>%
  filter(bmi_cat_factor != "Missing") # remove missing
cplot = ggplot(data=for_plot, aes(x=time, y=est, col=factor(event)))+
  geom_step(size=1.05)+
  xlab('Days since ECMO started')+
  ylab('Cumulative probability')+
  scale_x_continuous(breaks=c(0, 20, 40, 60, 80))+
  scale_colour_manual(name="Outcome:", values=c('dodgerblue','indianred1'), labels=c("Discharge",'Death')) +  
  facet_wrap(~bmi_cat_factor)+
  theme_bw()+
  theme(panel.grid.minor = element_blank())
figure_nums(name="cumunc", 'Cumulative probabilities over time of discharge and death. The panels show the categories of BMI.') # title
cplot
jpeg(filename='figures/obese/cumulative_risks.jpg', width=6, height=4, units='in', res=400, quality = 100)
print(cplot)
invisible(dev.off())
```

#### Alternative plot of cumulative risks (ECMO patients only)

This plot uses the same curves as above but presented in an alternative grouping.

```{r, fig.width=7}
aplot = ggplot(data=for_plot, aes(x=time, y=est, col=factor(bmi_cat_factor)))+
  geom_step(size=1.05)+
  xlab('Days since ECMO started')+
  ylab('Cumulative probability')+
  scale_x_continuous(breaks=c(0, 20, 40, 60, 80))+
  scale_colour_manual(name="BMI:", values=c(c("#8470FF", "#4EEE94", "#CCCCCC")), labels=bmi_labels) +  
  facet_wrap(~event_factor)+
  theme_bw()+
  theme(panel.grid.minor = element_blank())
figure_nums(name="cumunc_alternative", 'Cumulative probabilities over time of discharge and death. The panels are death and discharge.') # title
aplot
jpeg(filename='figures/obese/cumulative_risks_alternative.jpg', width=6, height=4, units='in', res=400, quality = 100)
print(aplot)
invisible(dev.off())
```

The cumulative probability curves for death (right panel) are similar for the low and high BMI groups. The high BMI group has a higher probability of discharge (left panel), meaning these patients have generally shorter stays in hospital and appear more likely to be discharged alive.

The curves for patients with a missing BMI are not shown.

## Cumulative survival models (ECMO patients only)

```{r include=FALSE}
# get all data ready for model (now that summary stats are done)
for (k in 1:N.impute){
  #
  trans.data.imputed[[k]] = mutate(trans.data.imputed[[k]], 
                  age = (age-50)/10, # scaled
                  bmi = (bmi-30)/5, # scaled
                  calendar_time = (as.numeric(date_ecmo)-18389)/31, # median = 2020-05-07
                  sex = as.numeric(sex=='Male'))
  #
  trans.data.mv.imputed[[k]] = mutate(trans.data.mv.imputed[[k]], 
                  age = (age-50)/10, # scaled
                  bmi = (bmi-30)/5, # scaled 
                  calendar_time = (as.numeric(date_icu)-18389)/31, # median = 2020-05-07
                  sex = as.numeric(sex=='Male'))
}
```

The survival models in this section focus on the patients' final (or cumulative) outcome of death and discharge. 

The table below shows the cumulative hazard ratios and 95% confidence intervals for the competing outcomes death and discharge. The variables in the model were age, sex, calendar time, APACHE II and BMI. 

Here we examine BMI as a continuous risk, so each increase in BMI is important. The plots above categorise BMI into low and high, but keeping BMI as a continuous variable increases our statistical power to find a difference in risk.

```{r, include=FALSE}
## run models with imputed data
BIC_null = death_coef = disch_coef = death_vcov = disch_vcov = NULL
for (k in 1:N.impute){ # loop through imputed data
  # design matrix (because of categorical variables):
  M = with(trans.data.imputed[[k]], model.matrix(exit ~ age + calendar_time + sex+ bmi+ apache_cat)) # complete data from MICE imputation
  cov1 = as.matrix(M[,-1]) # exclude intercept
  # death model
  cum_rr_death = with(trans.data.imputed[[k]], crr(ftime=exit, fstatus=to==3, cov1=cov1, cencode='0', maxiter = 100))
  # discharge model
  cum_rr_discharge = with(trans.data.imputed[[k]], crr(ftime=exit, fstatus=to==2, cov1=cov1, cencode='0', maxiter = 100))
  # get BIC (used below)
  bic = modsel.crr(cum_rr_death)
  fr_death = data.frame(model='Death', BIC=bic$BIC[2], threshold=0, k=k)
  bic = modsel.crr(cum_rr_discharge)
  fr_discharge = data.frame(model='Discharge', BIC=bic$BIC[2], threshold=0, k=k)
  BIC_null = bind_rows(BIC_null, fr_death, fr_discharge)
  # add estimates
  death_coef[[k]] = cum_rr_death$coef
  death_vcov[[k]] = cum_rr_death$var
  disch_coef[[k]] = cum_rr_discharge$coef
  disch_vcov[[k]] = cum_rr_discharge$var
}
# get combined estimates across all imputations
death_crr = summary(MIcombine(death_coef, death_vcov)) %>% mutate(Outcome = 'Death' )
discharge_crr = summary(MIcombine(disch_coef, disch_vcov)) %>% mutate(Outcome = 'Discharge' )
to_table = bind_rows(death_crr, discharge_crr) %>% 
  mutate(HR = roundz(exp(results), 2),
         lower = roundz(exp(`(lower`), 2),
         upper = roundz(exp(`upper)`), 2),
         CI = paste(lower, ' to ', upper, sep='')) %>%
  select(Outcome, HR, CI) 
```

```{r}
ftable = bind_cols(filter(to_table, Outcome=='Death') %>% select(-Outcome),
                   filter(to_table, Outcome=='Discharge') %>% select(-Outcome))
names(ftable) = c('HR1','CI1','HR2','CI2')
ftable = mutate(ftable, 
                Variable = names(cum_rr_discharge$coef),
                Variable = case_when( # nice rename
                  Variable == 'age'~ 'Age (+10 years)',
                  Variable == 'sex'~ 'Sex = Male',
                  Variable == 'calendar_time'~ 'Calendar Time (+1 month)',
                  Variable == 'apache_catMissing'~ 'APACHE = missing',
                  Variable == 'apache_catLow'~ 'APACHE = low',
                  Variable == 'bmi'~ 'BMI (+5 kg/m2)',
                  Variable == 'bmi_catMissing'~ 'BMI = missing',
                  Variable == 'bmi_catLow (<= 35)'~ 'BMI = low (<=35)',
                  TRUE ~ as.character(Variable)
                )) %>%
  select('Variable', everything()) %>% # move Variable first
  flextable() %>%
  add_header('Variable'='', 'HR1' = 'Death', 'CI1' = 'Death', 'HR2' = 'Discharge', 'CI2' = 'Discharge') %>%
  merge_h(i=1, part = "header") %>% # merge "To" column headers
  theme_box() %>% # nice theme
  align(align='center', part='all') %>% # centre text
  autofit() 
table_nums(name="crr_hazards", "Hazard ratios and 95% confidence intervals using a cumulative probability regression model. For patients on ECMO.")
ftable
```

Older age is associated with an increase in the risk of death and decrease in the risk of discharge.

A higher BMI is associated with an increased risk of death, but no clear association with discharge.

### Threshold in BMI (ECMO patients only)

Here we test if there is a threshold in the association between BMI and death/discharge.

```{r threshold, fig.width=7}
# function to make threshold
thresh_it = function(x, threshold){
  y = as.numeric(x > threshold) * (x-threshold)
  return(y)
}
## run models with imputed data
BIC = NULL
thresholds_s = (thresholds - 30)/5 # ... now standardise to match data
for (thresh_num in 1:length(thresholds)){
for (k in 1:N.impute){ # loop through imputed data
# design matrix (because of categorical variables):
  complete = trans.data.imputed[[k]] %>%
    mutate(bmi_thresh = thresh_it(bmi, thresholds_s[thresh_num])) # add threshold
  M = with(complete, model.matrix(exit ~ age + calendar_time + sex+ bmi+ bmi_thresh + apache_cat)) # complete data from MICE imputation
  cov1 = as.matrix(M[,-1]) # exclude intercept
  # death model
  cum_rr_death = with(complete, crr(ftime=exit, fstatus=to==3, cov1=cov1, cencode='0', maxiter = 100))
# discharge model
  cum_rr_discharge = with(complete, crr(ftime=exit, fstatus=to==2, cov1=cov1, cencode='0', maxiter = 100))
  # get BIC
  bic = modsel.crr(cum_rr_death)
  f_death = data.frame(model='Death', BIC=bic$BIC[2], threshold=thresholds[thresh_num], k=k)
  bic = modsel.crr(cum_rr_discharge)
  f_discharge = data.frame(model='Discharge', BIC=bic$BIC[2], threshold=thresholds[thresh_num], k=k)
  BIC = bind_rows(BIC, f_death, f_discharge)
}
}

## now plot BIC
# first add model without threshold
to_plot = bind_rows(BIC_null, BIC, .id='any') %>%
  mutate(threshold = ifelse(threshold==0, min(thresholds)-1, threshold)) %>%
  group_by(model, any, threshold) %>%
  summarise(mean = mean(BIC), n=n()) %>%
  ungroup()
#
bic_plot = ggplot(data=to_plot, aes(x=threshold, y=mean, col=factor(any), group=factor(any)))+
  geom_point(size=3)+
  geom_line()+
  scale_x_continuous(breaks=c(25,30,35,40), labels=c('25','30','35','40'))+ # blank label for no threshold
  scale_color_manual('Threshold', values=c('goldenrod2','darkseagreen3'), labels=c('No','Yes'))+
  xlab('BMI threshold (kg/m^2^)')+
  ylab("Bayesian Information Criterion (BIC)")+
  facet_wrap(~model, scales='free_y')+
  theme_bw()+
  theme(panel.grid.minor = element_blank())
bic_plot
```

The best model (smallest BIC) is a model without a threshold. That is true for death and discharge.

# ECMO and MV patients

Here we examine the outcomes of patients on mechanical ventilation. The aim is to provide a comparison to the ECMO patients in terms of the key question of the effect of BMI. The patients must have been started on mechanical ventilation.

The total number of patients is `r format(nrow(mv_patients),big.mark=',')`. The data come from `r length(unique(mv_patients$site_name))` sites.

## Demographics for ECMO and MV patients

#### a) Continuous variables (ECMO and MV patients)

```{r}
#
tab = select(mv_patients, pin, ecmo_group, bmi_cat, all_of(cont_vars)) %>%
  tidyr::gather(key='Variable', value='Value', -pin, -bmi_cat, -ecmo_group) %>%
  group_by(Variable, ecmo_group, bmi_cat) %>%
  summarise(Non_missing = sum(!is.na(Value)),
            Missing = sum(is.na(Value)),
            Median = roundz(median(Value, na.rm=TRUE),0),
            Q1 = roundz(quantile(Value, probs=0.25, na.rm=TRUE),0),
            Q3 = roundz(quantile(Value, probs=0.75, na.rm=TRUE),0),
            IQR = paste(Q1 , ' to ', Q3, sep='')) %>%
  select(Variable, ecmo_group, bmi_cat, Non_missing, Missing, Median, IQR)  %>%
  rename('BMI category' = 'bmi_cat')
ftab = flextable(tab) %>%
  theme_box() %>%
  merge_v(j=1:2) %>%
  autofit()
table_nums(name="summary_cont_MV", "Summary statistics for continuous variables", level=1) # title
ftab
```

#### b) Categorical variables (ECMO and MV patients)

```{r}
# dropped ,'comorbidity_obesity'
cat_vars = c('sex','country','ethnicity','bmi_cat','apache_cat')
#
tab = select(mv_patients, pin, ecmo_group, bmi_cat, all_of(cat_vars)) %>%
  tidyr::gather(key='Variable', value='Value', -pin, -bmi_cat, -ecmo_group) %>%
  group_by(Variable, ecmo_group, bmi_cat, Value) %>%
  tally() %>%
  group_by(Variable, ecmo_group, bmi_cat) %>%
  mutate(percent = roundz(prop.table(n)*100,0)) %>% # calculate percent
  arrange(Variable, ecmo_group, bmi_cat, -n) %>% # arrange from high to low within variables
  mutate(cell = paste(n, ' (', percent, ')', sep='')) %>%
  select(-n, -percent) %>%
  pivot_wider(values_from = cell, names_from=bmi_cat, values_fill='0')
ftab = flextable(tab) %>%
  merge_v(j = c("Variable",'ecmo_group') ) %>% # merge rows
  theme_box() %>%
  autofit()
table_nums(name="summary_cat_mv", "Summary statistics for categorical variables") # title
ftab
```

#### c) Co-morbidities by low / high BMI (ECMO and MV patients)

```{r}
to_table = select(mv_patients, pin, ecmo_group, bmi_cat, starts_with('comorb')) %>%
  select(-comorbidity_other_risk_factors, -comorbidity_hba1c_units, -comorbidity_diabetes_mellitus_hba1c, -comorbidity_immuno, -comorbidity_aids_hiv_cd4_count) %>% # remove characters or those with just missing
  pivot_longer(!c(pin,bmi_cat,ecmo_group), names_to='Comorbidity') %>%
  mutate(Value = case_when(
    value == 0 ~ 'No',
    value == 1 ~ 'Yes',
    is.na(value) == TRUE ~ 'Missing'
  )) %>%
  group_by(Comorbidity, ecmo_group, bmi_cat, Value) %>%
  tally() %>%
  group_by(Comorbidity, ecmo_group, bmi_cat) %>%
  mutate(percent = round(prop.table(n)*100)) %>%
  ungroup() %>%
  mutate(Comorbidity = str_remove_all(Comorbidity, pattern='^comorbidity_'),
         Comorbidity = str_replace_all(Comorbidity, pattern='_', replacement=' '),
         cell = paste(n, ' (', percent, ')', sep='')) %>%
  select(-n, -percent) %>%
  pivot_wider(values_from = cell, names_from=bmi_cat, values_fill='0')
ftab = flextable(to_table) %>%
  merge_v(j = c("Comorbidity",'ecmo_group') ) %>% # merge rows
  theme_box() %>%
  fontsize(size=10, part='all') %>%
  autofit()
table_nums(name="comorb_mv", "Counts and percentages of comorbidities by BMI category") # title
ftab
```

#### d) Time variables by low / high BMI (ECMO and MV patients)

```{r, time_table_mv}
# create time variables
mv_patients = mutate(mv_patients,
                  # calculate end dates
      date_end = ifelse(is.na(date_hospital_discharge)==FALSE, date_hospital_discharge, date_death), # end date of death or discharge
      date_end = as.Date(date_end, origin='1970-01-01'),
      #
      date_end_icu = ifelse(is.na(date_discharge)==FALSE, date_discharge, date_death), # end date of death or discharge
      date_end_icu = as.Date(date_end_icu, origin='1970-01-01'),
      #
      days_hosp = as.numeric(date_end - date_admission) + 0.5,
      days_icu = as.numeric(date_end_icu - date_icu) + 0.5,
      days_symptoms_hosp = as.numeric(date_admission - date_first_symptom) +0.5, 
      #
      duration_ecmo = as.numeric(date_ecmo_discontinued - date_ecmo) + 0.5,
      duration_mv = as.numeric(date_mech_vent_discontinued - date_mechanical_ventilation +0.5),
# new (Jan 2021)
      days_admission_mv = as.numeric(date_mechanical_ventilation - date_admission) + 0.5,
days_admission_ecmo = as.numeric(date_ecmo - date_admission) + 0.5)

#
time_vars = c('days_symptoms_hosp','days_hosp','days_icu','duration_ecmo','duration_mv','days_admission_mv','days_admission_ecmo')
#
stats = select(mv_patients, pin, bmi_cat, ecmo_group, all_of(time_vars)) %>%
   pivot_longer(!c(pin, bmi_cat, ecmo_group), names_to='Variable', values_to='Value') %>%
  group_by(Variable, bmi_cat, ecmo_group) %>%
  summarise(Non_missing = sum(!is.na(Value)),
            Missing = sum(is.na(Value)),
            Median = roundz(median(Value, na.rm=TRUE),0),
            Q1 = roundz(quantile(Value, probs=0.25, na.rm=TRUE),0),
            Q3 = roundz(quantile(Value, probs=0.75, na.rm=TRUE),0),
            IQR = paste(Q1 , ' to ', Q3, sep='')) %>%
  select(Variable, ecmo_group, bmi_cat, Non_missing, Missing, Median, IQR) %>%
  mutate(Variable = case_when(
    Variable=='duration_ecmo'~ 'Duration of ECMO',
    Variable=='duration_mv'~ 'Duration of MV', 
    Variable=='days_hosp'~ 'Days in hospital', 
    Variable=='days_icu'~ 'Days in ICU',
    Variable=='days_admission_mv'~ 'Days from hospital admission to MV', 
    Variable=='days_admission_ecmo'~ 'Days from hospital admission to ECMO', 
    Variable=='days_symptoms_hosp'~ 'Days between symptoms and hospital admission'
#    TRUE ~ as.character(x)
    )) %>%
  rename('BMI category' = 'bmi_cat')
ftab = flextable(stats) %>%
  theme_box() %>%
  fontsize(size=8, part='all') %>% # smaller font
  merge_v(j=1:2) %>%
  autofit()
table_nums(name="summary_time_mv", "Summary statistics for time variables (in days)") # title
ftab
```



## Model diagram (time to ECMO)

Here we examine the time to ECMO adjusting for the competing risks of death and discharge. This answers the question about whether there are patient characteristics associated with being put on ECMO.

The figure below shows the four states in the multi-state survival model. In this model we  examine the time to the first event after starting mechanical ventilation, the three events are starting ECMO, death or discharge. We do not examine patients' times after starting ECMO because that is considered above.

```{r, fig.width=4, fig.height=3}
states = c('MV start','ECMO start','Death','Discharge\nfrom hospital')
# Matrix of joins between states:
M = matrix(data=c(
  2,1,
  3,1,
  4,1
), byrow = TRUE, ncol=2)
figure_nums(name="model_diagram_mv", "The four states in the model and the possible movements between states", level=1) # title, increase title number
multistate_diagram(states=states, links = M, pos = c(1, 3), arr.pos=0.65, box.size = 0.14) # see 99_functions.R
```

### Table of transitions between states

This table below shows the transitions (movements) between the states.

```{r}
three.state = filter(trans.data.mv) %>%
  filter(from == '1') # just MV patients
tab = as.data.frame(with(three.state, table(from, to))) %>%
  mutate(from = case_when(
    from=='1' ~ 'MV start'),
    to = case_when(
    to=='2' ~ 'ECMO',
    to=='4' ~ 'Death',
    to=='3' ~ 'Discharge',
    to=='cens' ~ 'Censored')
  ) %>%
  tidyr::spread(to, Freq)
# table header
table_header <- data.frame(
  col_keys = c('from','Censored', 'ECMO','Discharge','Death'),
  h1 = c('', rep("State moved to", 4)),
  h2 = c('State moved from', 'Censored','ECMO','Discharge','Death'),
  stringsAsFactors = FALSE )
ftab = flextable(tab) %>%
  set_header_df(mapping = table_header, key = "col_keys" ) %>% # add header
  merge_h(i=1, part = "header") %>% # merge "To" column headers
  align(align='center', part='all') %>% # centre text
  theme_box() %>%
  autofit() 
table_nums(name="transitions_mv", "Transition numbers between states for the model examining time to ECMO")
ftab
```

### Cumulative risks for MV patients

```{r, fig.width=7}
# cumulative incidence
three.state = mutate(three.state,
          bmi_cat_num = as.numeric(as.factor(bmi_cat))) # need simply category for cuminc
cum_inc = with(three.state, cuminc(ftime=exit, fstatus=to, group=bmi_cat_num, cencode='cens'))
g = ggcompetingrisks(cum_inc)
# make new plot
bmi_labels = c("High (> 35)","Low (<= 35)" ,"Missing")
for_plot = mutate(g$data, 
                  bmi_cat_factor = factor(group, levels=1:3, labels=bmi_labels))  %>%
  filter(bmi_cat_factor !='Missing')
cplot = ggplot(data=for_plot, aes(x=time, y=est, col=factor(event)))+
  geom_step(size=1.05)+
  xlab('Days since mechanical ventilation started')+
  ylab('Cumulative probability')+
  scale_x_continuous(breaks=c(0, 20, 40, 60, 80))+
  scale_colour_manual(name="Outcome:", values=c('purple','dodgerblue','indianred1'), labels=c("ECMO","Discharge",'Death')) +  
  facet_wrap(~bmi_cat_factor)+
  theme_bw()+
  theme(panel.grid.minor = element_blank())
figure_nums(name="cumunc_mv", 'Cumulative probabilities over time of starting ECMO, discharge and death. The panels show the three categories of BMI.') # title
cplot
jpeg(filename='figures/obese/cumulative_risks_mv.jpg', width=6, height=4, units='in', res=400, quality = 100)
print(cplot)
invisible(dev.off())
```

The cumulative risk plots show that the patients with a lower BMI had a:

* lower probability of being put on ECMO (purple lines)
* higher probability of discharge (blue lines) 
* lower probability of death (red lines)

### Cumulative risk models post MV start (ECMO and MV patients)

```{r, include=FALSE}
## run models with imputed data
death_coef = disch_coef = ecmo_coef = ecmo_vcov = death_vcov = disch_vcov = NULL
for (k in 1:N.impute){ # loop through imputed data
  three_state = filter(trans.data.mv.imputed[[k]], from=='date_mechanical_ventilation') # just from MV
# design matrix (because of categorical variables):
  M = with(three_state, model.matrix(exit ~ age + calendar_time + sex+ bmi+ apache_cat)) # complete data from MICE imputation
  cov1 = as.matrix(M[,-1]) # exclude intercept
  # ecmo model
  cum_rr_ecmo = with(three_state, crr(ftime=exit, fstatus=to=='date_int_state_start', cov1=cov1, cencode='censored', maxiter = 100))
  # death model
  cum_rr_death = with(three_state, crr(ftime=exit, fstatus=to=='death', cov1=cov1, cencode='censored', maxiter = 100))
# discharge model
  cum_rr_discharge = with(three_state, crr(ftime=exit, fstatus=to=='discharge', cov1=cov1, cencode='censored', maxiter = 100))
  # add estimates
  ecmo_coef[[k]] = cum_rr_ecmo$coef
  ecmo_vcov[[k]] = cum_rr_ecmo$var
  death_coef[[k]] = cum_rr_death$coef
  death_vcov[[k]] = cum_rr_death$var
  disch_coef[[k]] = cum_rr_discharge$coef
  disch_vcov[[k]] = cum_rr_discharge$var
}
# get combined estimates across all imputations
ecmo_crr = summary(MIcombine(ecmo_coef, ecmo_vcov)) %>% mutate(Outcome = 'ECMO')
death_crr = summary(MIcombine(death_coef, death_vcov)) %>% mutate(Outcome = 'Death')
discharge_crr = summary(MIcombine(disch_coef, disch_vcov)) %>% mutate(Outcome = 'Discharge')
to_table = bind_rows(death_crr, discharge_crr, ecmo_crr) %>% 
  mutate(HR = roundz(exp(results), 2),
         lower = roundz(exp(`(lower`), 2),
         upper = roundz(exp(`upper)`), 2),
         CI = paste(lower, ' to ', upper, sep='')) %>%
  select(Outcome, HR, CI) 
```

```{r}
ftable = bind_cols(filter(to_table, Outcome=='ECMO') %>% select(-Outcome),
                   filter(to_table, Outcome=='Death') %>% select(-Outcome),
                   filter(to_table, Outcome=='Discharge') %>% select(-Outcome))
names(ftable) = c('HR1','CI1','HR2','CI2','HR3','CI3')
ftable = mutate(ftable, 
                Variable = names(cum_rr_discharge$coef),
                Variable = case_when( # nice rename
                  Variable == 'age'~ 'Age (+10 years)',
                  Variable == 'sex'~ 'Sex = Male',
                  Variable == 'calendar_time'~ 'Calendar time (+1 month)',
                  Variable == 'apache_catMissing'~ 'APACHE = missing',
                  Variable == 'apache_catLow'~ 'APACHE = low',
                  Variable == 'bmi'~ 'BMI (+5 kg/m2)',
                  Variable == 'bmi_catMissing'~ 'BMI = missing',
                  Variable == 'bmi_catLow (<= 35)'~ 'BMI = low (<=35)',
                  TRUE ~ as.character(Variable)
                )) %>%
  select('Variable', everything()) %>% # move Variable first
  flextable() %>%
  add_header('Variable'='', 'HR1' = 'ECMO', 'CI1' = 'ECMO', 'HR2' = 'Death', 'CI2' = 'Death', 'HR3' = 'Discharge', 'CI3' = 'Discharge') %>%
  merge_h(i=1, part = "header") %>% # merge "To" column headers
  theme_box() %>% # nice theme
  align(align='center', part='all') %>% # centre text
  fontsize(size=10, part='all') %>% 
  autofit() 
table_nums(name="crr_hazards_mv", "Hazard ratios and 95% confidence intervals examining time to ECMO")
ftable
```

Older patients were much less likely to be put on ECMO (hazard ratio below 1). Men were more likely to be put on ECMO (hazard ratio above 1). Higher BMI was somewhat associated with an increased risk of being put on ECMO. 

# Are patients on ECMO at higher risk of death and discharge compared with patients on MV

### Model diagram 

```{r, fig.width=4, fig.height=3}
states = c('MV start','ECMO start','Death','Discharge\nfrom hospital')
# Matrix of joins between states:
M = matrix(data=c(
  2,1,
  3,1,
  4,1,
  4,2,
  3,2
), byrow = TRUE, ncol=2)
colours = c('yellow','skyblue','indianred2','grey')
figure_nums(name="model_diagram_full", "The four states in the model and the possible movements between states", level=1) # title, increase title number
multistate_diagram(box.colours= colours, states=states, links = M, pos = c(1, 2, 1), arr.pos=0.65, box.size = 0.14) # see 99_functions.R
```


```{r, include=FALSE}
### to plot state numbers over time
# get starting numbers
starts = filter(trans.data.mv, entry==0) %>%  # can use any of the imputations
  group_by(from) %>% 
  tally() %>% 
  ungroup() %>% 
  mutate(p = prop.table(n))
init_dis_2 <- c(starts$p, 0, 0, 0) # four states

## Cox model stratified by transition
trans.data.full_ext = ext_mstate(trans.data.mv, traM=tra) #  
c_2 <- coxph(Surv(entry, exit, status==1) ~ strata(trans), data= trans.data.full_ext, method = "breslow")
# msfit calculates baseline hazards
msf_2 <- msfit(c_2, trans = tra)
# probtrans calculates transition probabilities, prediction from day 0
pt_fc_2 <- probtrans(msf_2, predt = 0)

# Create weighted average of transition probabilities for progress of entire cohort
# Used to produce full cohort results in Table 3 
fc_2 <- pt_fc_2[[1]] * starts$p[1]
pt_fc_2[[1]] <- fc_2
```


```{r state_plot, fig.width=6.5}
figure_nums(name="state_plot", "Plot of predicted probabilities over time for the four states") # increase level
plot_it = function(){
  par(mai=c(0.9,0.9,0.5,0.1), las=1) # b, l, t, r; leave room at top for legend
  plot(pt_fc_2, from = 1,  ord = c(4,2,1,3), type= "filled", cols = colours,
       lwd= 2, xlab = "Days since ECMO started", ylab = "Predicted Probabilities", 
       cex.lab = 1.25, legend = rep("", 4),
       main= NULL, axes=FALSE)
  # better x-axis ticks:
  axis(side=2, at=seq(0,1,0.2), labels=TRUE) # y-axis
  # 90 days
  axis(side=1, at=seq(0,90,10), labels=FALSE)
  axis(side=1, at=seq(0,90,5), labels=TRUE, tick=FALSE)
  # legend at top
  par(xpd=TRUE)
  labels = c('MV','ECMO','Death','Discharge')
  legend(0, 1.2, labels, col=colours, border=NA, fill=colours, bty='n', ncol=4)
}
plot_it()
jpeg(filename='figures/obese/patients_states_time_four.jpg', width=5, height=4, units='in', res=400, quality = 100)
plot_it()
invisible(dev.off())
```

## Models comparing the effect of BMI in MV and ECMO

Here we examine whether the effect of BMI on death differs for patients in MV or on ECMO. We repeat the model to examine the effect of BMI on discharge.

```{r, include=FALSE}
source('1_obese_cox_full.R')
```

```{r}
ftable = bind_cols(filter(to_table, Outcome=='Death') %>% select(-Outcome),
                   filter(to_table, Outcome=='Discharge') %>% select(-Outcome)) 
names(ftable) = c('HR1','CI1','HR2','CI2')
ftable = mutate(ftable, 
                Variable = names(cox_death$coefficients),
                Variable = case_when( # nice rename
                  Variable == 'age'~ 'Age (+10 years)',
                  Variable == 'sex'~ 'Sex = Male',
                  Variable == 'calendar_time'~ 'Calendar time (+1 month)',
                  Variable == 'apache_catMissing'~ 'APACHE = missing',
                  Variable == 'apache_catLow'~ 'APACHE = low',
                  Variable == 'bmi'~ 'BMI (+5 kg/m2)',
                  Variable == 'bmi:fromdate_mechanical_ventilation' ~ 'BMI interaction with MV',
                  Variable == 'bmi_catMissing'~ 'BMI = missing',
                  Variable == 'bmi_catLow (<= 35)'~ 'BMI = low (<=35)',
                  TRUE ~ as.character(Variable)
                )) %>%
  filter(Variable != 'fromdate_mechanical_ventilation') %>%
  select('Variable', everything()) %>% # move Variable first
  flextable() %>%
  add_header('Variable'='', 'HR1' = 'Death', 'CI1' = 'Death', 'HR2' = 'Discharge', 'CI2' = 'Discharge') %>%
  merge_h(i=1, part = "header") %>% # merge "To" column headers
  theme_box() %>% # nice theme
  align(align='center', part='all') %>% # centre text
  autofit() 
table_nums(name="crr_hazards_bmi", "Hazard ratios and 95% confidence intervals using a cumulative probability regression model.", level=1) # increase table number
ftable
```

The key result in the table above is the "BMI interaction with MV" which is the change in the effect of BMI for patients on MV compared with patients on ECMO. The hazard ratios for both death and discharge are below one, meaning that the association between BMI and discharge/death is reduced in patients on MV compared with patients on ECMO. 

#### Probability of death by BMI

```{r}
figure_nums(name="bmi_interaction_plot", "Plot of predicted cumulative probabilities of death and discharge over time for a patient with a BMI of 30 or 35.") # increase level
print(plot_bmi_death) # from 1_obese_cox_full.R
```

The plot shows the estimated probability of death over time for patients with a BMI of 30 and 35. The higher BMI patients have a slightly reduced probability of death when on MV, and slightly increased risk of death when on ECMO.

# Appendix

Here we check key model assumptions for the survival models. The results shown below would likely not go in the main section of any paper, but could be added to an online appendix.

```{r, include=FALSE}
## refit zero knot model ##

# just for one imputation of BMI
# design matrix (because of categorical variables):
M = with(trans.data.imputed[[1]], model.matrix(exit ~ age + calendar_time + sex+ bmi+ apache_cat)) # complete data from MICE imputation
cov1 = as.matrix(M[,-1]) # exclude intercept
# death model
cum_rr_death = with(trans.data.imputed[[1]], crr(ftime=exit, fstatus=to==3, cov1=cov1, cencode='0', maxiter = 100))
# discharge model
cum_rr_discharge = with(trans.data.imputed[[1]], crr(ftime=exit, fstatus=to==2, cov1=cov1, cencode='0', maxiter = 100))
```


## Survival times of censored observations (ECMO patients only)

```{r, fig.width=7}
cen = survfit(Surv(time=exit, event=to=='0') ~ 1, data=trans.data)
g = ggsurvplot(cen, 
           break.time.by=10, 
           xlim=c(0,90))
# alternative plot
cen_plot = ggplot(g$data.survplot, aes(x=time, y=1-surv, ymin = 1-lower, ymax=1-upper))+
  geom_ribbon(alpha=0.5, fill='pink')+
  geom_step(size=1.05, col='dark red')+
  scale_y_continuous(limits=c(0,1))+
  xlab('Days since ECMO start')+
  ylab('Proportion censored')+
  scale_x_continuous(breaks=seq(0,90,10))+
  theme_bw()+
  theme(panel.grid.minor= element_blank())
figure_nums(name="censored", "Kaplan-Meier plot of the proportion of censored observations over time") # 
cen_plot
```

The greatest number of censored patients is on day `r censor_day`, the last day of follow-up for this study when patients not yet discharged or dead are censored.


## Schoenfeld residuals (ECMO patients only)

```{r, fig.width=6}
## plot residuals against failure time
# data frame of times
times = data.frame(times = cum_rr_death$uftime) %>%
  mutate(index=  1:n())
# data frame of names
names = data.frame(name = paste('X', 1:length(cum_rr_death$coef), sep=''),
                   variable = names(cum_rr_death$coef))
# set-up results in long format
to_plot = pivot_longer(data.frame(cum_rr_death$res), cols=everything()) %>%
  group_by(name) %>%
  mutate(index = 1:n()) %>%
  ungroup() %>%
  left_join(times, by='index') %>%
  left_join(names, by='name')
res_plot = ggplot(data=to_plot, aes(x=times, y=value))+
  geom_hline(yintercept=0, lty=2)+
  geom_point()+
  geom_smooth(col='dark red', method = 'loess', span=0.9, se = FALSE)+
  xlab('Days since ECMO')+
  ylab('Residual')+
  facet_wrap(~variable, scales='free_y')+
  theme_bw()+
  theme(panel.grid.minor = element_blank())
figure_nums(name="Schoenfeld", "Schoenfeld residuals for the six variables used in the survival model of time to death.") # 
res_plot
```

The residuals are for the survival models that examine the time to death and discharge for ECMO patients. Ideally the mean (red line) should be consistent over time. There is some evidence of a time-dependent effect for age, meaning the effect of age changes during a patient's time on ECMO. The mean line for BMI is relatively flat, hence the effect of BMI appears consistent during the time on ECMO. 

## Sensitivity analysis excluding large BMIs (ECMO patients only)

```{r include=FALSE}
bmi_extreme = 70
bmi_extreme_stan = (bmi_extreme - 30)/5 # standardise
n_exclude_bmi = sum(trans.data$bmi >= bmi_extreme, na.rm=TRUE)
```

The distribution of BMI is right-skewed with a few very large BMIs that could potentially be influential. To examine this we ran a sensitivity analysis that excludes the `r n_exclude_bmi` patients with a BMI over 70.

```{r, include=FALSE}
## run models with imputed data
death_coef = disch_coef = death_vcov = disch_vcov = NULL
for (k in 1:N.impute){ # loop through imputed data
  this_data = filter(trans.data.imputed[[k]], bmi < bmi_extreme_stan) # exclude large BMIs (use standardised scale here)
# design matrix (because of categorical variables):
  M = with(this_data, model.matrix(exit ~ age + calendar_time + sex+ bmi+ apache_cat)) # complete data from MICE imputation
  cov1 = as.matrix(M[,-1]) # exclude intercept
  # death model
  cum_rr_death = with(this_data, crr(ftime=exit, fstatus=to==3, cov1=cov1, cencode='0', maxiter = 100))
# discharge model
  cum_rr_discharge = with(this_data, crr(ftime=exit, fstatus=to==2, cov1=cov1, cencode='0', maxiter = 100))
  # add estimates
  death_coef[[k]] = cum_rr_death$coef
  death_vcov[[k]] = cum_rr_death$var
  disch_coef[[k]] = cum_rr_discharge$coef
  disch_vcov[[k]] = cum_rr_discharge$var
}
# get combined estimates across all imputations
death_crr = summary(MIcombine(death_coef, death_vcov)) %>% mutate(Outcome = 'Death' )
discharge_crr = summary(MIcombine(disch_coef, disch_vcov)) %>% mutate(Outcome = 'Discharge' )
to_table = bind_rows(death_crr, discharge_crr) %>% 
  mutate(HR = roundz(exp(results), 2),
         lower = roundz(exp(`(lower`), 2),
         upper = roundz(exp(`upper)`), 2),
         CI = paste(lower, ' to ', upper, sep='')) %>%
  select(Outcome, HR, CI) 
```

```{r}
ftable = bind_cols(filter(to_table, Outcome=='Death') %>% select(-Outcome),
                   filter(to_table, Outcome=='Discharge') %>% select(-Outcome))
names(ftable) = c('HR1','CI1','HR2','CI2')
ftable = mutate(ftable, 
                Variable = names(cum_rr_discharge$coef),
                Variable = case_when( # nice rename
                  Variable == 'age'~ 'Age (+10 years)',
                  Variable == 'sex'~ 'Sex = Male',
                  Variable == 'calendar_time'~ 'Calendar time (+1 month)',
                  Variable == 'apache_catMissing'~ 'APACHE = missing',
                  Variable == 'apache_catLow'~ 'APACHE = low',
                  Variable == 'bmi'~ 'BMI (+5 kg/m2)',
                  Variable == 'bmi_catMissing'~ 'BMI = missing',
                  Variable == 'bmi_catLow (<= 35)'~ 'BMI = low (<=35)',
                  TRUE ~ as.character(Variable)
                )) %>%
  select('Variable', everything()) %>% # move Variable first
  flextable() %>%
  add_header('Variable'='', 'HR1' = 'Death', 'CI1' = 'Death', 'HR2' = 'Discharge', 'CI2' = 'Discharge') %>%
  merge_h(i=1, part = "header") %>% # merge "To" column headers
  theme_box() %>% # nice theme
  align(align='center', part='all') %>% # centre text
  autofit() 
table_nums(name="crr_hazards_exclude", "Hazard ratios and 95% confidence intervals using a cumulative probability regression model. For patients on ECMO. Excluding patients with a BMI over 70.")
ftable
```

A higher BMI is still associated with an increased risk of death.

## Sensitivity analysis of influential sites (ECMO patients only)

Here we check if the results are strongly influenced by one site using a leave-one-out sensitivity analysis. Each site is left out in turn and the difference in the estimated risk of BMI is plotted. The plot shows the mean log hazard ratio (dot) and 95% confidence interval (vertical line).

```{r, include=FALSE}
results_exist = length(dir(path='results', pattern='obesity_leave_out_site')) > 0
if(results_exist==TRUE){
  load('results/obesity_leave_out_site.RData')
}
if(results_exist==FALSE){
  results_left_out = NULL
  sites = unique(trans.data$site_name)
  for (this_site in sites){# loop through sites 
    model.death = model.discharge = list() # for storing imputed models
    for (k in 1:N.impute){
      this_data = complete(imputed_data, k) %>%
        filter(site_name != this_site) # exclude one site
      # design matrix (because of categorical variables):
      M = with(this_data, model.matrix(exit ~ age + calendar_time + sex+ bmi+ apache_cat)) # complete data from MICE imputation
      cov1 = as.matrix(M[,-1]) # exclude intercept
      # death model
      cum_rr_death = with(this_data, crr(ftime=exit, fstatus=to==3, cov1=cov1, cencode='0', maxiter = 100))
      # discharge model
      cum_rr_discharge = with(this_data, crr(ftime=exit, fstatus=to==2, cov1=cov1, cencode='0', maxiter = 100))
      # add estimates
      death_coef[[k]] = cum_rr_death$coef
      death_vcov[[k]] = cum_rr_death$var
      disch_coef[[k]] = cum_rr_discharge$coef
      disch_vcov[[k]] = cum_rr_discharge$var
    }
    # combine over imputations
  death_crr = summary(MIcombine(death_coef, death_vcov)) %>% 
  mutate(Outcome = 'Death',
         Variable = names(cum_rr_death$coef))
discharge_crr = summary(MIcombine(disch_coef, disch_vcov)) %>% 
  mutate(Outcome = 'Discharge',
         Variable = names(cum_rr_discharge$coef))
ests = bind_rows(death_crr, discharge_crr) %>% 
  filter(Variable == 'bmi') %>% # Just BMI
  mutate(site = this_site,
           HR = exp(results),
         lower = exp(`(lower`),
         upper = exp(`upper)`)) %>%
  select(Outcome, site, HR, lower, upper)   
  results_left_out = bind_rows(results_left_out, ests)
} #end of site loop
save(results_left_out, file='results/obesity_leave_out_site.RData')
} # end of results exist if
```

```{r, fig.width=7}
results_left_out = arrange(results_left_out, Outcome, HR) %>% # high to low
  group_by(Outcome) %>%
  mutate(xaxis = 1:n()) %>% # site number
  ungroup()
lplot = ggplot(data=results_left_out, aes(x=xaxis, y=HR, ymin=lower, ymax=upper))+
  geom_point()+
  geom_errorbar(width=0)+
  facet_wrap(~Outcome, scales='free_y')+
  scale_x_continuous(breaks = NULL)+
  xlab('Site left out')+
  ylab('Log hazard ratio')+
  theme_bw()+
  theme(panel.grid.minor = element_blank())
figure_nums(name="leave_out", "Difference in the estimated risk for BMI after leaving out each site.") # 
lplot
```

Results are ordered by the mean. The differences are small, hence the results are not dependent on one site.

## Martingale residuals 

<!--- Positive values mean that the patient died sooner than
expected (according to the model); negative values mean that
the patient lived longer than expected (or were censored) --->


These residuals are for the survival models that use all four states and compare discharge/death for patients in MV against discharge/death for patients on ECMO. Martingale residuals are usually asymmetric, but are useful for spotting outliers. 

##### a) age

```{r, fig.width=6}
source('1_obese_residuals.R')
figure_nums(name="mart1", "Martingale residuals from the survival model against age") # 
age_res
```

##### b) BMI

```{r, fig.width=6}
figure_nums(name="mart2", "Martingale residuals from the survival model against BMI") # 
bmi_res 
```

There are no concerning outliers in the residuals. 

# References
